Improved documentation for sigmoid_cross_entropy_with_logits.
The definition of logistic loss used by sigmoid_cross_entropy_with_logits is a
bit unusual. It took me some time to convince myself that it was equivalent to
a more common definition of logistic loss. I'd like to document the equivalence
because I think it will be helpful for other TensorFlow newbies.
Change: 112075662