Avoid allocation of std::function<> object per node executed, as well as a hash-table lookup per allocated output.
Instead, we now pre-compute the AllocatorAttributes for every output
tensor in the graph into an array (indexed by a base number per node +
output index), and changed OpKernelContext::Params to provide
a pointer to the base of the array for the node, rather than providing
a std::function<>.

Updated test code to avoid so much code duplication when setting up
the OpKernelContext::Params object in various places.

Used gtl::InlinedVector<...> instead of std::vector<...> in a few
places in tensorflow/core/kernels/reduction_ops_common.h

Didn't make a measurable change in overall performance but allocations and
time spent in the std::function destructor code was significantly reduced.
Change: 112103260