Removes static shape checks from certain image ops.
The following ops are affected:
* tf.image.flip_left_right()
* tf.image.flip_up_down()
* tf.image.per_image_whitening()
* tf.image.random_flip_left_right()
* tf.image.random_flip_up_down()
* tf.image.transpose_image()

This change allows these ops to be used in a setting where each image
may have a different shape (such as a JPEG-decoding pipeline). When
static shape information is available, the same checks are performed.

To improve the produced shape information, this change also makes the
shape function for `tf.reverse()` more precise when dealing with an
input with partially-known shape.
Change: 112277661