Note that master is installed with plain git clone
Since the docs are versioned, and the latest stable version is the default visible docset on the website, we should point out that the installation instructions actually install master, and also give instructions on how to install a specific branch. This change doesn't mention which branch it is on because we don't want to have to change this text in each release.

Closes #891.