Support more than two session factories.
This CL changes how the session factory is chosen. Previous an empty
target would always use DIRECT_SESSION, and a non-empty target would
always REMOTE_SESSION. In preparation for multiple distributed session
implementations (issue #23), we now delegate to the SessionFactory to
see whether it accepts a given SessionOptions. Existing programs
should continue to work unmodified.

NOTE: This CL assumes that the domains of the registered session factories
do not overlap. We may need to revisit this in future.
Change: 115008046