Adds implementations of PosixEnv::SchedClosure[After]().
The current implementations are very simplistic: they spawn a thread
in response to each call. This is needed to deal with the fact that
some users of SchedClosure rely on the ability to spawn blocking
closures, and we lack an unbounded threadpool.

TODO(mrry): Replace the currently-blocking users of this API with
asynchronous implementations.
Change: 115239594