Implements "bool strictness" for `tf.Tensor`.
Using a `tf.Tensor` as a Python boolean value is ambiguous: in library
code it typically means "is this tensor not None?", whereas in user
code it can mean "does this tensor evaluate to `True`" (which does
not work and leads to buggy code).

This change adds a warning when a tensor is used as a bool, and
replaces all known instances in TensorFlow with the appropriate `is
None` or `is not None` checks.

This is the first step towards addressing Issue #1454.
Change: 116984574