Fix python 3 vs 2 issue regarding bytes != str
Was able to reproduce the issue in python3. The problem comes down to 'str' and 'bytes' being equivalent types in python2, but in python3 they are not, since python3 encodes all strings as unicode. Because of this, they need to explicitly be converted to bytes using an encoding, which in this case is utf-8 since protobuf uses utf-8 for encoding strings.
Change: 117244230