Disable //third_party/tensorflow/core/distributed_runtime:master_test.
This test is flaky. We need a more robust solution for starting
multiple processes and connecting them; `PickUnusedPortOrDie()` is not
robust to multiple concurrent tests on the same host. A possible
approach would be to add a portserver, or add a handshake between the
parent test process and the server processes.
Change: 117293607