Add a half type to TensorFlow core, based on Eigen::half.
Note that this is only the type, not support for it in any ops,
so it is not useful for anything yet. In particular,
neither TF_CALL_REAL_NUMBER_TYPES nor TF_CALL_GPU_NUMBER_TYPES
list Eigen::half, so even though a lot of ops will end up
declaring support for the new type, calling them will fail at
runtime.
Change: 117825461