Add TensorBoard auto-reload behavior.
Auto-reloading defaults on and is controlled by a setting.

To have a place to put that setting, also add a settings panel.

The settings panel uses paper-dialog and is opened by clicking a settings icon button in the global top right.

Minor changes to how the toolbar is laid out by CSS to ensure that the settings all show up in a line and don't bunch up together.

Also, the "TensorBoard" message on the left no longer takes a fixed pixel width (everything is flex), so the toolbar will use space more appropriately when the window size is small.
Change: 119997005