TensorFlow: Simple placer accepts device mismatches for reference edges.
In the case where a user specifies an invalid assignment (assign on
a different device than the variable, for example), we now have
the SimplePlacer ignore the invalid assignment and only do what is possible.

The major reason for this change is that users and library code are forced to
override device functions for assignment.  For example:

with tf.device(some_function):
  var = tf.Variable(...)   # some_function places this on CPU
  input = tf.constant(...)
  # Necessary because some_function may otherwise place assign
  # on a different device.
  with ops.colocate_with(var):
    tf.assign(var, input)

To get rid of these colocation declarations, the placer needs to handle
and ignore situations where tf.assign() is placed on the "wrong" device.
Unlike other kindsof assignment, where a user may be surprised to know
that an op was placed on an unexpected device, reference edges force
that two ops are on the same device, so it is not as helpful to return
an error (the user should not have expected the assign to be on
another device anyway).
Change: 120760300