TensorBoard: Replace crosshairs with tooltips.
The crosshairs were implemented in a way that made them difficult to maintain: the chart component needed to talk to a runs component.
Here, the crosshairs are replaced with an overlay tooltip that shows the run, step, time, and value.
The tooltip never shows interpolated values, so the user can now distinguish real data from interpolations.
It also simplifies the architecture.
Change: 120949357