Make managed_session(close_summary_writer=False) actually work:  - Pass the close_summary_writer argument to the stop() method instead of    the close() method.  - Fix the unittest.
Additionaly, change the Supervisor to only write the graph once in a given
event file.  This makes the event files much smaller when reusing a
SummaryWriter across runs, as you often do for evals.

Add tests for that behavior in the Supervisor.managed_session tests.

The bug went unnoticed because of several factors:
 - We do ignore exceptions from close() so the bogus keyword arg was ignored.
 - The corresponding unittest was starting the supervisor summary_thread
   accidently, which caused it to enqueue the summary twice: the test saw the
   summary twice as expected, but from the wrong source.
Change: 122758728