Add create_scope_now option to make_template.
This is a non-breaking API change.

This gives the option of having a Template capture a scope at construction
time, rather than the time of first __call__. This can prove useful in a
situation where the Template is created but than passed into other code to be
called much later - it may be hard to be sure of the eventual scope name without
understanding all the lower level parts of the model, and potentially
configuration changes could lead to a different first __call__ location,
meaning a differently named variable, and potential problems saving and
loading.

By contrast, if create_scope_now_ is set to True, then wherever the template is
constructed defines the scope name. Lower level code which actually calls it is
free to switch around the order of __call__ without changing the variable name.
Change: 122874739