Use external repository for protobuf dependency. (#1289)
Currently, TensorFlow uses a Git submodule for its dependency on
protobuf. This was due to the hack used in protobuf for building Python
support with Bazel, which was required since Bazel's Python rules did
not support adding directories to `PYTHONPATH`. Now that the new
`imports` attribute has been added to the Python rules in Bazel 0.2 and
the hack for Python support in protobuf has been removed, this change
removes the `google/protobuf` Git submodule and adds an external
repository for including protobuf.

This patch also adds gmock.BUILD to the tensorflow repo.

Fixes #1069
Fixes #2021