TF debugger core callback and DebugGateway
This is the first of a series of CLs aimed at implementing the TF debugger (tfdb).

This C++ CL adds the node-outputs callback to ProcessOutputs() in ExecutorImpl and provides access to it via the DebugGateway class. This makes it possible to observe intermediate node outputs during a DirectSession.Run() call.
Change: 125882979