Fix gather gradient for empty slices
We can't use -1's in the reshape if the size is 0, since the -1 would be
ambiguous.  Also simplify the code now that array_ops.shape does its own
checking for fully defined shapes.

A previous version used too much colocate_with and broke a distributed
model.  This version uses colocate_with only for params following the
previous version of the code.
Change: 130720593