Move host_info.h internal, move public functions into cpu_info.h and make public.
Hostname() is not public, but kLittleEndian arguably is (some code
might need to know what the endianness is, like decode_raw_op,
so it probably should be public).  Similarly, NumSchedulableCPUs
is probably something a lot of code wants to know.  It is also
already used widely and seems hard to remove in short notice,
which suggests something about its utility.
Change: 131203680