Add SpaceToBatchND and BatchToSpaceND ops for > 2 spatial dimensions.
These new operations are more general versions of the existing
SpaceToBatch and BatchToSpace operations that support an arbitrary
number of spatial dimensions and a non-uniform block shape.

For backwards compatibility, the existing SpaceToBatch and BatchToSpace
operations are retained, but internally forward to the new, generic
implementation.

For convenience, the new operations support an arbitrary number of
trailing non-spatial dimensions, rather than requiring a single batch
dimension.

As an optimization, leading and trailing spatial dimensions with no
padding and a block size of 1 are collapsed such that they incur no
additional overhead.

The maximum number of (non-collapsed) spatial dimensions must be
specified at compile time in spacetobatch_functor.h, and is currently
set to 4.
Change: 133836914