Add generic Python interface for convolution and pooling
The existing Python interfaces in ops/nn.py for convolution and pooling
are both specific to the number of convolution dimensions and also, in
the case of pooling, specific to the type of pooling.  This leads to
needless variation and duplication.

This change adds a new `convolution` Python wrapper that provides a
uniform interface for 1-D, 2-D, or 3-D convolutions, and internally
dispatches to the correct implementation.  If 4-D or higher-dimensional
convolutions are supported in the future, an additional case can easily
be added.  It also supports atrous convolutions through the
dilation_rate parameter, which allows a separate dilation rate to be
specified for each spatial dimension.  This function subsumes the
functionality of the existing atrous_conv2d function, which is
nonetheless retained for backwards compatibility.

This change also adds a new `pool` Python wrapper that provides a
uniform interface for 1-D, 2-D, or 3-D max/avg pooling.  It also
supports atrous pooling through a dilation_rate parameter.

Both `convolution` and `pool` are implemented using the new function
`with_space_to_batch` to support atrous convolution/pooling.  In the
case that dilation_rate is uniformly 1 (non-atrous case), there is no
overhead: just the raw convolution/pooling operation is added the graph.
Change: 134624471