Get rid of the javascript error when switching to the TSNE tab.
This specifically adds a bit "deferUpdate" to setProjection which doesn't
update the scatter plot until notifyProjectionsUpdated is called later. This
is used for TSNE because we don't have projections until TSNE is run, which is expensive
and takes some time. Also adds a bit to not animate when switching to TSNE when recreating
the scene.
Change: 135226248