go: Generate wrapper functions for ops
This change fills in the body of the "genop" command-line
tool to generate Go wrapper functions for each TensorFlow op.
The generated API is as follows:

- One function generated per TensorFlow op
- Arguments to the function are: Scope, all inputs, all required attributes and a variadic list of optional attributes
- Outputs of the function are all outputs and an error
- The optional attributes of each op get their own type (<OpName>Attr)
- And there are a set of factory functions for that type, one for each optional attribute

See genop_test.go for a sample of how OpDefs turn into generated source code.

Operations with either:
(a) Reference typed input or outputs
(b) "func" valued attributes
Are skipped for now. Have to work out the appropriate representation in Go for them.
But a vast majority of operations do not use these, so we skip them.

The "description" and "summary" fields in the OpDef are used to generate documentation comments. However, the OpDef style is to use markdown and so some of that markdown does not end well in Go documentation. Living with that for now.

Another step towards #10
Change: 136412140