Fix shape function for the Variable op.
Well, almost. The shape function cannot distinguish
between an unknown ranked variable and a scalar variable.
This is for historical reasons, as the Variable operator
(like the Placeholder operator) has been in existence
since a time when an empty TensorShapeProto used to
imply an unknown rank.

There is a plan to "fix" this by introducing and transitioning
to PlaceholderV2 and VariableV2 operations (see https://github.com/tensorflow/tensorflow/commit/5e176998d92a64d78df57e9fb78582e5e7e4ebb6)

But until that happens, make the shape function for the
Variable op work for all but scalars.

An explanation of the changes to python:

- state_ops.py: This is done to keep the in-memory representation of the graph
  in python (which resulted in an unknown shape when set_shape=False)
  is consistent with what any C++ code thinks about the shape of the Variable op
- variable_scope_test.py: When an initializer is provided to tf.get_variable,
  it ultimately calls state_ops.variable with set_shape=False. Because of the
  change to state_ops.py, the shape attribute wouldn't end up with [73].
  My understanding is that the use of the shape attribute wasn't the intention
  of the test, so switching to another attribute maintains the intention.

Fixes #5106
Change: 137179498