Fix ThreadSanitizer warnings on std::vector<bool>.
Each worker thread in the thread pool writes to its designated entry in
the std::vector. The entry must be of a size that the CPU can write to
without stepping on the adjacent entries. Moreover, std::vector<bool> is
implemented using bitwise operations. The assignment operator sets or
clears a bit in a word, which requires both a read and a write and is
therefore not atomic.
Change: 137749423