C API: Do not take ownership of the TF_Tensors
Prior to this change, TF_*Run, TF_SetAttrTensor and TF_SetAttrTensorList
took ownership of the TF_Tensor*s of the feeds. This can make performance
client languages bothersome when the same Tensor is repeatedly fed into
multiple session executions as the memory for the feed tensor would need
to be re-allocated and filled in every time.

With this change, these functions no longer take ownership of the
TF_Tensor*.

The changes to the Go API implementation reflect the claimed benefits.
(Another step towards #10)
Change: 139169388