Use 'priority' field in device registration to allow external devices to register custom priority orderings.  By default, external devices are prioritized below CPU and GPU (maintaining existing behavior as of recently), but can choose to register a higher number if desired.
The default priority is set to 50, and our existing CPU and GPU
devices have registrations that have higher values.  A custom device
registration can choose an even higher value than any of ours, if they
want to be preferred by the simple_placer.  Note that in the future,
better placement algorithms may decide to just ignore this field, so
it's not a particularly good ordering to rely on.

Changed the test to not rely on device registrations of existing devices,
registering dummy factories instead.

This maintains the existing within-DeviceType priority ordering, and also maintains the previously existing DeviceSet ordering, with one exception that SYCL is preferred below GPU but still above CPU.  There were no guarantees about SYCL vs. GPU ordering before (it was pointer ordering before), so this is more deterministic and not harmful.
Change: 143723649