Change default initializer to glorot_uniform.
It is the same as uniform_unit_scaling for most inputs, inlcuding 2D ones,
but has a slight difference in larger-dimensional cases. This makes
default initialization the same as in Keras and some other frameworks.
Also correct edge cases in the initializer and the few smoke tests that rely on it.
Change: 144033334