Added a Tensorboard plugin that will serve data from a TensorFlow debugger
The TensorFlow debugger is currently internal-only, but hopefully not for long.
Currently, the plugin does nothing - its one handler just responds with same list
of op names provided in POST data. Eventually, it will respond with health pills.

The goal of this change is not to add functionality, but to lay out the organization
of where the logic will be located.
Change: 146499722