Remove the libtensorflow_c.so target.
This was identical to libtensorflow.so for all but
a few days in the life of libtensorflow_c.so's existence.
Remove libtensorflow_c.so to remove any confusion about
which target to use for the TensorFlow C library.

(https://github.com/tensorflow/tensorflow/commit/783c52e
and #3830
were the changes that resulted in two targets doing the same thing)

I have made attempts to remove use of libtensorflow_c.so in
other projects such as:
https://github.com/LaurentMazare/tensorflow-ocaml/pull/19
https://github.com/tensorflow/rust/pull/56
https://github.com/malmaud/TensorFlow.jl/pull/157
Change: 148665838