Refactor python BUILD targets.
The deps of the "framework" target did not reflect the actual imports
used in those files. Adding the missing deps causes circular
dependencies, so this patch refactors "framework" into per-file
targets. It still exposes the "framework" and
"framework_for_generated_wrappers" to avoid breaking the many targets
that have these as dependencies.
Change: 148834459