Check for whether to show health pill stats based on presence of -, 0, and + values.
We used to rely on the mean being infinite, but oddly, Tensorboard gives us those errant values as strings:

[1, 256, 256, 0, 0, 0, 0, 0, "Infinity", "-Infinity", "NaN", "NaN"]

This breaks logic that checks for !== Infinity because a string is not Infinity. Something in our stack might have difficulty reading Infinity and NaN from JSON.

For now, I think we might want to resort to checking for valid values. This resolves a console error.
Change: 148920617