Allow callers to control the names of the slot variables.
Other optimizers which have only a single slot variable allow control through
the 'name' constructor parameter, but the FtrlOptimizer has two variables.
Because they both are created with the same 'name' parameter, one of them has
name as a suffix, and the other has name + "_1" as a suffix.  This change
allows them to be specified in a more controllable way.
Change: 152802478