Python partial run binding should make call when feeds but no fetches are provided.
Although no fetches and some feeds being specified on a run call should be a no-op, it is perfectly valid for no fetches and some feeds to be provided in a partial_run call. This is because a subsequent partial_run call with fetches may need that earlier feed. (see test for example). This change fixes the python binding to allow for that.

PiperOrigin-RevId: 156396669