[XLA] Add a way to classify constraints as mandatory or non-mandatory.
The idea is that the backend might want to override operand constraints
after propagation has already assigned them. The example in the test
case uses custom constraints to enforce that all operands of an
instruction are the same. We can't assign that early because we don't
know the layout a priori. During propagation an instruction gets
assigned different layouts on output and operands, but when the
output constraint is propagated into the operands we can override
them now, making the constraint possible.

PiperOrigin-RevId: 156413491