Halve the number of CPU tests run in 9 files.
This pattern tests all tests twice (once for use_gpu=True, once for
use_gpu=False).

Switch to using a single test pattern that uses use_gpu=True,
which tests on GPU if a GPU is available, and tests on CPU otherwise.

the %test%_gpu target tests on GPU, the %test% target tests on CPU, so
both are covered.

The tricky situation is when you want to test locally on GPU, you have
to remember to run the _gpu target, but reducing test time is more
important than individual developers making mistakes at this point.

PiperOrigin-RevId: 157141876