Create a set of sample data for the scalars plugin
We aim to provide for each plugin a set of sample data satisfying the
following properties:

  - The data is of a realistic form.
  - The data is sufficient to exercise all capabilities of the plugin.
  - The data is not completely boring.

Such data then serves many purposes: it is documentation for this
particular plugin; it is documentation for TensorFlow and TensorBoard in
general; it can be used for manual testing of TensorBoard; and it can be
used in the plugin's unit tests. Ideally, we'll also be able to create a
script to aggregate the test data from all first-party plugins,
providing a single data set that can be used to comprehensively test
TensorBoard.

To start us off, this change adds such data for the scalars plugin. It's
realistic in that it has a few thousand steps and a few dozen runs. It
enables testing smoothing (due to the added noise), tag groups (due to
the various summary names), and run filters (due to the combinatorial
run parameters). And it's implementing a tiny physics simulation, so
it's not _entirely_ boring!

PiperOrigin-RevId: 157274669