[XLA] Fix a subtle issue in copy_insertion due the interaction between copy overriding logic and RecordIndicesToColocatingBuffers:
- When building instructions ShapeTree to be copy overriden, it is possible
that we create a single kCopy for two identical instructions. An example can
be:

    %tuple.19 = tuple(%constant.4, %constant.1793, %constant.1793)

where it is used in a while.init operand, and constant.1793 is read-only within
the loop and also used by another while loop. The copy overriding pass will then
create the following (logical, not finalized) tuple:

    %tuple.19 = tuple(%constant.4, %copy.5, %copy.5)

- In the subsequent pass RecordAmbiguousOrNonDistinctIndices, to add copies to
ensure point_to set is distinct, the duplicate %copy.5 are ignored because they
are not yet finalized, and these indices (1 and 2 in the example) are still
marked as to-be copied.

Therefore distinctiveness is lost.

This fix applies to the override building stage, to explicitly avoid creating
shared copies for non-distinct buffers.

PiperOrigin-RevId: 157872231