Remove tf-sidebar-helper
It wasn't actually helpful.

As a side-effect, this change makes the text plugin behave like the
other plugins in that its "runs"/"tags" route will list all runs, not
just those that include text summaries. (Without this, the existing
infrastructure breaks when it tries to look up non-text runs in the
run-to-tag mapping returned by /data/plugin/text/runs. We'll soon be
collapsing/optionalifying this infrastructure, but we want the text
plugin to be consistent, anyway, so now is a good time to do so.)

PiperOrigin-RevId: 159121999