In filter_dataset, if the element doesn't match, clear the output.
Prior to this change, the output tensors list wouldn't be cleared
if the element didn't match, so output_tensors would be appended to
on the next iteration, violating the number of inputs contract for
the filter function.  The new test exhibited this problem.  By
clearing the output tensors list when the element doesn't match, we
reset the state back to empty, maintaining the invariant at
the start of the loop.

Fixes #10725.

PiperOrigin-RevId: 159242269