[tf.contrib.data] Fix the handling of dict-typed elements in functions.
Previously, we were treating a `dict` as a sequence, which led to
incorrect behavior like passing all of the dict's keys rather than
values as the arguments to a map or filter function. This change
changes the behavior so that the dict is passed as a single argument
to these functions. It additionally fixes the ported version of
`nest.flatten_up_to()` so that `Dataset.padded_batch()` works with
dict-typed elements.

Fixes #11016.

PiperOrigin-RevId: 160548475