Add support for display_name and summary_description to the tf.summary.tensor_summary op.
The display_name will be used to display the series in TensorBoard, in lieu of the tag, assuming that it is specified. (When it is not specified, behavior will stay the same.)
The summary_description will allow the user to write a longform readable description of the summary series for display in TensorBoard. Markdown will be supported.

This will make it possible for TensorBoard-2 summary ops to give the user direct control over the display name, solving https://github.com/tensorflow/tensorboard/issues/59.

PiperOrigin-RevId: 162566261