[XLA:CPU] Switch TF gather's HLO implementation to use dynamic-update-slice in a "while" loop.
Benchmarks results (times in ms):

  nontrivial_gather.axis0_cpu:     0.110
  nontrivial_gather.axis0_xla_cpu: 0.139
  nontrivial_gather.axis1_cpu:     0.093
  nontrivial_gather.axis1_xla_cpu: 0.142
  nontrivial_gather.axis4_cpu:     1.183
  nontrivial_gather.axis4_xla_cpu: 2.658
  slice_gather.axis0_cpu:          0.00388
  slice_gather.axis0_xla_cpu:      0.00397
  slice_gather.axis1_cpu:          0.00421
  slice_gather.axis1_xla_cpu:      0.00427
  slice_gather.axis4_cpu:          0.252
  slice_gather.axis4_xla_cpu:      0.114

As you can see, the pure-XLA implementation is slower in all the nontrivial
cases and as-fast or faster in the slice-gather cases.

The slice-gather cases are gathers that can be implemented as a single XLA
dynamic-slice, and so the speedup here is likely understated: Once we can
simplify the gather to a single dynamic-slice, we should be able to do many
other optimizations to it, ideally fusing it so it has zero cost.

The nontrivial gathers all gather more than one element, and are implemented
with an XLA while loop.  The most important one is the axis 0 gather --
gathering from an inner dimension is so slow no matter what you do that it's
probably not worth optimizing.

It's possible to make this XLA implementation faster -- one option I've
considered is "unrolling" the gather into a series of dynamic-slice's that are
then concat'ed together.  This would be totally fusable, unlike the
implementation in this CL.  Another option would be adding a notion of
uninitialized memory into XLA -- part of what makes us slow is that we have to
initialize the memset our output to 0 before we overwrite it.

But given that the shape we're benchmarking here is totally arbitrary, and
given that we're getting decent performance, I think this is good enough to
start with.

PiperOrigin-RevId: 171883273