[tf.data] Move contrib op registrations to tensorflow/contrib/data/ops.
This change moves the contrib-level Dataset ops out of the core op
libraries (where they were placed for short-term technical reasons)
into a more appropriate location in contrib. This enables us to modify
the signatures of these ops without being subject to the core
backwards compatibility requirements. (We currently must modify the
backwards compatibility ledger each time we make an allowed change to
a contrib op.) For now, the kernel implementations remain in the core
library, because they depend on headers that are not part of the
public API.

The change also removes some code testing contrib features in the core
kernel_tests, since it relies on the contrib op registrations and was
already adequately tested in the contrib tests.

PiperOrigin-RevId: 174912490