[XLA] Don't deemphasize nodes inside of subcomputations in dumped XLA graphs.
Nodes inside of subcomputations (e.g. fusion computations) are always
printed by the HLO graph dumper.  Before this change, the dumper was not
fully aware of this fact, leading it to mark as "deemphasized" (i.e.
draw as gray with a dashed outline) nodes that had no business of being
deemphasized.

PiperOrigin-RevId: 175247474