Do not log a warning when `Rendezvous::Send()` fails in the Send op kernel.
A failing call to `Send()` indicates that the step has been aborted by
a corresponding call to `Rendezvous::StartAbort()`. As a result, the
error logged by `Send()` is not particularly informative, and creates
a non-deterministic amount of extra log spam for each step that fails
as `Send()` calls are being issued. The failure that causes the step to be
aborted is logged separately by the kernel that failed, unless that kernel
deliberately does not log on failure.

In particular, this change reduces log spam when using `Iterator.get_next()`
in a multi-device setting. The `Iterator.get_next()` op deliberately does not
log when an `OutOfRange` error (indicated the end of the dataset) is raised,
because this is common and expected behavior, especially when using an
initializable iterator that is reinitialized at the end of an epoch. Previously,
when running in distributed mode or using a GPU, pending `Send()` calls may
cause unwanted log messages to be printed.

Fixes #12414.

PiperOrigin-RevId: 175716290