[XLA:CPU] Add Hlo profiling support to XlaJitCompiledCpuFunction
Some of the functionality has bled into the generic XlaCompiledCpuFunction, but
there still remains a fair amount of work to do before the AOT side of things
start working.

This CL also fixes a bug I introduced in a previous CL -- when I changed
IrEmitter::hlo_to_profile_idx_ to a value, I changed the signature of the
generated function to always have the "profile_counters" argument when the AOT
client code expects the signature to not have that argument.  In practice this
wasn't an issue for the standard x86 calling convention, but it could easily
have been problematic on other architectures and calling conventions.  After this
change the mismatch is no longer present.
PiperOrigin-RevId: 177481998