Make 'producer_op_list' argument to import_graph_def work with C API.
This changes exposes a minor functional change: unknown attrs cause an
import error with the C API enabled, but only cause an error when the
graph is run in the original Python code (the error will occur even if
the problematic op isn't being run). It seems OK to fail faster in
error cases.

PiperOrigin-RevId: 177615841