Relax the requirements on --input_arrays and --input_shapes.
Relaxing the requirement on --input_shapes is overdue: in many cases, the input nodes contains shape information, which we're reading from the input file anyway, so there is no point asking the user to pass it. Also, we want to eventually fully support arbitrary runtime input shapes, and that is a step in that direction.

Relaxing the requirement on --input_arrays is motivated by the need to get graph visualizations for strange graph where either we don't know beforehand what the inputs are, or they only have constant inputs (which makes no sense from an inference perspective, but is something that we've seen in some graphs).

Finally, this change also drops the requirement that --input_arrays be connected by a path to --output_arrays. This was a meaningful requirement to have, however:
 - it caused friction with some bad graphs that we need visualizations of, and getting a visualization is often useful to understanding how the inputs failed to be connected to outputs;
 - it rarely caught user mistakes in practice. I.e. the mistake that it was catching wasn't really a mistake that people make in practice.

PiperOrigin-RevId: 178395267