Add complex64 and complex128 support for `tf.angle` on GPU
In PR 10643, complex64 and complex128 support have been added
for `tf.angle` on CPU. However, because of the compilation
errors, the complex support on GPU is not enabled yet.

The issue was that, std::arg is not available on nvidia device
for GPU. This fix changes to used atan2 instead, which is
available in CUDA.

The relevant test cases have bee enabled.

This fix is related to 10643.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>