[XLA] Remove a false invariant used in generating test inputs.
This invariant is incorrect; the parameter will never be the shape of the use's
first operand; in fact the param should match operand(1)'s shape for
DynamicSlice and operand(2)'s shape for DynamicUpdateSlice (since start_indices is the argument that we are trying to constrain the bounds for).  This is checked
for in CreateLiteralForConstrainedUses().

PiperOrigin-RevId: 180717508