[XLA:CPU] Count cycles in non-entry computations
This change teaches XLA to maintain cycle counters specifically for non-entry
computations, like computations representing the body of a While.  Without this
change, instructions in such non-entry computations are noted as taking 0.00% of
their execution time which isn't ideal.

Implementation-wise, this just falls out of uniformly using a
std::unordered_map<T, int64> for both the HloInstruction->ProfileIndex and the
HloComputation->ProfileIndex mappings.
PiperOrigin-RevId: 180750463