Say SQLite BLOB protos are ideally temporary
We ideally want a schema that's pure SQL and doesn't embed
protos. However some protos we don't have time to translate
quite yet. So we put a BLOB as the last field so it can be
migrated later.

Please note there might end up being some things highly
specific to TensorFlow and/or Google that end up here. In
those cases we might consider keeping the proto blobs
permanently, if they're difficult to express in SQL.

PiperOrigin-RevId: 182313847