Including common.h with NEON_2_SSE.h (#15743)
* Including common.h with NEON_2_SSE.h

Including common.h to make sure that USE_NEON is defined in case of NEON_2_SSE.h is used; otherwise USE_NEON will not be propagated to this file and `portable_tensor_utils.h` will be used

* Removing spaces

* Using neon is USE_NEON is defined

* Using common.h to set USE_NEON if applicable

* Adding required dependencies

Since tests are more likely compiled on sse4-enabled machine - if they are run with common.h - it will enable sse4, there are two ways - just enable it and let tests using sse4, or add extra flag check which would require USE_SSE4 flag to be defined. I've added all required dependencies for tests to pass, so it could work.

* Rearranging dependencies according to linter

* Fixing sanity check

* trying to rerun, could be cached

* Triggering build

* triggering build - looks like it stuck

no status for more than 24h

* including x86_64 since NEON_2_SSE4 convertion will allow neon to run on x86 with SSE3,4

* Fixing name

* using darwing for MacOS

* Fixing header inclusion rules

* Fix syntax

* Depend on includes

We depend on both types and compatibility. That's in the "types" rule.