Fix the script entry point for freeze_graph.
The wrapper created by `setup.py` calls the entry point
function with no arguments. `freeze_graph.main` expects
the global `FLAGS` to be set, and one argument.

This change adds a `run_main` function to use as the entry point,
which expects no arguments and parses the flags.

It also adds a `flags` argument to `main` so the flags can be passed
directly without using a `global FLAGS` declaration.