PUBLIC: [TF:XLA] Eliminate more copies after HLO scheduling.
After scheduling HLOs it is very beneficial to try more copy elision: The
sequential ordering from the schedule is stricter than the data-dependency ordering used
during copy insertion.

Also, allow more operands to share a buffer with their user. In particular, the user has to be element-wise only wrt to the specified operand, and not wrt to all operands.

These two changes allow more copies to be eliminated.

PiperOrigin-RevId: 200292049