Derivative of tf.igamma(a, x) and tf.igammac(a, x) with respect to a.
Previously, both functions only supported the derivative with respect to x. We add the derivative with respect to the other argument. It is computed using the Eigen function igamma_der_a that performs forward-mode differentiation of the code for igamma. This function is not exposed in the public TensorFlow API.

PiperOrigin-RevId: 201145398