[XLA] Add free functions in the xla:: namespace that mirror XlaBuilder methods. These free functions can be used unqualified if desired (even outside xla:: because of argument-dependent lookup), removing the need to explicitly pass or call via an XlaBuilder for most methods.
The main exceptions to this rule are source nodes (e.g., Const, Parameter, Recv), or nodes with a possibly-zero arity (e.g., ConcatInDim, Tuple).

No tests are added for the new methods; the intention is to switch all existing users of the builder methods to the free functions, at which point they will have good test coverage.

PiperOrigin-RevId: 202167553