Co-allocate call operands with callee parameters
Right now we give these parameters an additional allocation and then don't use
it.

This is interesting for the CPU backend because we use call instructions to
represent fork-join parallelism (i.e. a specially annotated kCall instruction
tells the CPU IR emitter to shard the called computation across CPU threads).

Moreover, I need this for a principled fix to b/111116907.

PiperOrigin-RevId: 204820965