Add new Dockerfile assembler based on partials
This change adds a new suite of TensorFlow dockerfiles. The
dockerfiles come from an assembler controlled by a yaml spec, and are
based on a set of re-usable partial dockerfiles.

The assembler and spec include conveniences like spec validation,
references to other images and specs for minimizing repetition, and arg
expansion.