Add explicit "-lz" linkopt for Android selective registration targets.
tensorflow/core/lib/io/zlib_inputstream.cc uses zlib, and hence, if a binary is compiled
that uses any of the symbols in zlib_inputstream.cc, then the binary needs zlib's
symbols to be available. One way to do that is to specify "-lz" as a linkopt, making the
binary use the system's zlib library.

PiperOrigin-RevId: 205710390