Rename MKL-related feature macros.
The existing feature macros are named INTEL_MKL to indicate that any flavor of
MKL is available, INTEL_MKL_ML to indicate that *only* MKL-ML is available
(i.e. MKL-DNN is not), and DO_NOT_USE_ML to indicate that *only* MKL-DNN is
available (i.e. MKL-ML is not).

This change renames INTEL_MKL_ML to INTEL_MKL_ML_ONLY and DO_NOT_USE_ML to
INTEL_MKL_DNN_ONLY. The meanings of the macros have not changed.

This change also adds a few sanity checks to mkl_util.h that ensures that the
combination of INTEL_MKL, INTEL_MKL_ML_ONLY, and INTEL_MKL_DNN_ONLY is
logically consistent: the *_ONLY macros may not both be defined, and if either
of them is defined, bare INTEL_MKL must also be defined.

PiperOrigin-RevId: 208313735