Handle control dependencies from switch nodes as nonreachable.
In DeleteReachableNodes all the nodes reachable from nodes deleted from the
graph during extraction was considered. But if a node had a control dependency
on a switch, then that node doesn't conditionally execute based on the switch
predicate and is not part of the conditional extracted, so it should be
considered reachable for deletion.

Additionally perform sweep of graph for dead nodes together with deleting the
reachable nodes to keep all dead node deletion together.

Also delete a dead function and ensure all graph dumps from functionalize_cond
has that as prefix.

PiperOrigin-RevId: 212485183