Make worker heartbeats robust to shutdowns.
The coordinator sends worker heartbeats periodically in a loop.  Previously if
one of these heartbeats timed out (due to a lame worker), the entire loop would
exit and force the remaining workers to exit after the timeout.

This removes the coordinator from the heartbeat set and catches any TF errors
that occur during the heartbeat.

PiperOrigin-RevId: 217574934