Clarify concrete function calling conventions
Throws an exception if Tensor inputs have non-unique names. These names become keyword arguments to the concrete function.

A bunch of extra documentation and error checking around the creation of concrete functions, since it will be part of the SavedModel export workflow. The new calling conventions are relatively close to SavedModel signatures, which will hopefully mean fewer surprises when people try to serve a model.

PiperOrigin-RevId: 218886883