Prune unreachable functions in the `GraphDef` objects returned from partitioning.
We currently make a full copy of the `FunctionDefLibrary` for each partition of a partitioned graph, and transfer it to remote workers in the remote setting. This change should reduce the memory consumption of `FunctionLibraryDefinition` objects throughout the runtime, by eliminating functions that will never be executed.

This change moves the implementation (and corresponding tests) of the functions `tensorflow::grappler::ReachableFunctionDefinitions()` into methods `tensorflow::FunctionLibraryDefinition::ReachableDefinitions()`. It also removes the `tensorflow::grappler::ReachableFunctions()` functions, which were unused except in tests.

PiperOrigin-RevId: 221549148