Use if_cuda instead of if_cuda_is_configured
if_cuda_is_configured is now a no-op.  For xla_gpu_jit this does not matter
since its deps are under if_cuda, but it does make a difference for
xla_gpu_device.

We should at some point refactor both xla_gpu_device and xla_gpu_jit to either
uniformly include an if_cuda or to not include it, but there are other hills I
want to die on at the moment.

PiperOrigin-RevId: 222906280