Workaround for PartitionedCall trying and failing to run on TPUs when saving
Just omits the function decorator for now. This is pretty terrible and we should fix it, but it will need some work on the TPU side.

Spoke to iga@. Apparently the CPU annotations don't work because the function captures a resource which is on the TPU (and so the eager placer puts the call op on the TPU). One option is to then XLA-compile the function, although that fails right now because we're trying to save strings and XLA doesn't have a kernel for that.

I should also follow up with TPU+checkpointing integration tests.

PiperOrigin-RevId: 226390521