Add is_dynamic_dimension repeated bool field to Shape. The bool indicates which of the dimensions in the Shape are dynamic. For dynamic dimensions the 'dimensions' int64 value indicates an inclusive upper bound on the dimension bound. This CL adds the field to ShapeProto and Shape and adds support to the parser. It is unused otherwise.
This change also includes a couple small changes to the Shape interface to preserve the invariant that the dimensions vector and the dynamic_dimensions vector are always the same size.

This may not be the final form for dynamic shape representation. Having int64 lower and upper bounds may be cleaner where static dimensions have equal lower and upper bound dimensions. However, this current representation is sufficiently expressive to start using dynamic shapes in other parts of XLA. Shape inference would be a natural first step. Changing the underlying shape representation during this initial process should not be too difficult.

PiperOrigin-RevId: 226434424