[tf.data] Replacing IteratorStateMetadata with a simpler alternative.
Currently, serialization of iterator state uses the `IteratorStateMetadata` proto, which AFAICT is only used to convert a map to and from string. This CL replaces the proto-based conversion with a simpler mechanism. This change is motivated by making it possible to move the VariantTensorDataReader and VariantTensorDataWriter into a header file and depending on protos from TensorFlow header files is discouraged (c.f. b/62899350).

PiperOrigin-RevId: 229802166