Refactor code for listing checkpointable functions of an object.
The main goal is to move the logic of "dir+getattr" from all checkpointable
objects to just "AutoCheckpointable". Similar to how checkpointable base
does not enforces

The refactoring also provides an internal mechanism for a class to provide
its own listing of functions if needed and makes obsolete the need to register
a getter and an attribute extractor.

PiperOrigin-RevId: 232499034