[XLA:Python] Rename CompiledLocalComputation/CompiledXrtComputation to LocalExecutable/XrtExecutable. Rename LocalComputation* to Computation* (there is nothing Local about it any more).
Split the two roles of the Python LocalComputation class into Computation (an uncompiled computation that can be compiled) and Executable (a compiled computation that can be executed).

While in principle this is a significant API change, in practice JAX (the only known user) does not refer to any of these classes by name.

PiperOrigin-RevId: 233772418