Log Keras training and validation summaries to separate runs.
Previously, metrics like accuracy would be logged as `epoch_acc` during
training and `val_epoch_acc` during training. As of this change,
training metrics are logged to a `train/` subdirectory under the
top-level logdir, while validation metrics are logged to `validation/`.
This has the advantage that training and validation metrics can be shown
in the same plot:

![Screenshot of new behavior](https://user-images.githubusercontent.com/4317806/52606214-9e3ddc80-2e26-11e9-9a02-2a5228edc8f6.png)

Tested:
Running a simple MNIST model generates a TensorBoard instance with
summaries as described above, and with a graph under the ?train? run.

RELNOTES: Keras training and validation curves are shown on the same plot.
PiperOrigin-RevId: 234179031