New configuration mechanism for eager mode.
Expose setting various ConfigProto options as individual functions. We
choose to expose these as individual functions as they become easier
to maintain and deprecate if needed.

Most of these options are required to be set at program startup.
However, in the future we can enable setting some of these options at
runtime.

Fixes #25138

PiperOrigin-RevId: 235588442