Detect and report an error when a WorkerService.RunGraph request is delivered more than once.
Currently, we track request IDs for WorkerService.RecvTensor, because if a duplicate message is delivered for that method the system will deadlock. However, a duplicate WorkerService.RunGraph request can also trigger deadlock. Consider the following simple example:

1. Master M divides a simple graph across workers W_0 and W_1, where W_0 sends a tensor to W_1.
2. M calls W_0.RunGraph and W_1.RunGraph.
3. W_1 calls W_0.RecvTensor to receive the tensor, and receives a response.
4. W_0.RunGraph returns successfully.
5. M retries the call to W_1.RunGraph.
6. W_1 again calls W_0.RecvTensor to receive the tensor, and never receives a response because the sender has completed.

To work around this problem, we extend the duplicate request ID tracking to WorkerService.RunGraph.

PiperOrigin-RevId: 239327645