Move communicator_key from instance to group params.
Before this change, the communicator_key in NCCL collective OpKernel was
associated with a collective instance.  This caused a new NCCL communicator
object to be created for each collective instance.

After this change, communicator_key is associated with collective groups.  For
a particular group, defined by a group key and a set of devices, all instances
of the group will reuse the same key, and hence communicator object.

PiperOrigin-RevId: 240638209