[XLA:GPU] Rename ReductionToVector to ReductionFromOrToContiguousDimensions.
In the GPU code emitter, ReductionToVector initially means the dimensions to
keep are contiguous in the input tensor. Later on we extend the code emitter to
support the case whether the dimensions to keep are not contiguous but the
dimensions to reduce are. As such, ReductionFromOrToContiguousDimensions is a
better name.

PiperOrigin-RevId: 241435326