If Const node in outside compilation at head has successor that's not outside compilation at head, do not move it out of TPU computation.
Example: if we put a TopKV2 op in outside compilation, its "input" is from TPU and its "k" is a Const node in outside compilation, original logic will move "k" out of TPU computation and leave TopKV2 in outside compilation. So the shape inference graph for outside compilation will have an XlaRecvAtHost node as TopKV2's "k" input, and we cannot infer output shape for the TopKV2 node any more.

Solution is to keep the "k" Const in outside compilation with the TopKV2.

PiperOrigin-RevId: 242024072