Fix negative axis issue with ragged tensor and reduce_sum
This fix tries to address the issue raised in 27497 where
`tf.reduce_sum` with multiple negative axes and ragged tensor
does not produce correct result.

The issue is that during reduce op, ragged tensor will reduce
one axis at a time. However, for negative axis, sort result is
reversed so order is different.

This fix convert to positive before the sort to make sure
the order.

This fix fixes 27497.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>