Moved EigenThreadPoolWrapper class from eigen_thread_pool.h to local_devices.cc.
EigenThreadPoolWrapper class is no longer needed and to prevent further use of
this class, it makes to sense to remove it. The only place it is used now is
inside local_device.cc, in a fallback code path that is intended to be run when
threads::ThreadPool stops implementing Eigen::ThreadPoolInterface. Therefore
moving this class inside an anonymous namespace inside local_device.cc
PiperOrigin-RevId: 243352899