Make physical device initialization thread-safe
Physical devices are lazy initialized in order for the runtime to be
fully initialized. We must therefore ensure that this initialization is
thread-safe. Also, as there are a number of other functions that utilize
the physical_devices, we ensure they are also initialized appropriately.

Also rename _initialize_devices() to _initialize_logical_devices().

PiperOrigin-RevId: 244124118