Use a linear outer loop in auto-clustering
This removes the existing worklist driven algorithm which can be quadratic in
the worst case.  The algorithm is not fully linear because we still have a
worst-case o(n^2) step to decide whether a clustering decision will break
resource variable semantics.

This required imposing a new invariant over our clustering policies: making a
cluster larger does not make edges leaving it more contractable. That is, if we
have digraph { X->Y; Y->Z; } then collapsing X->Y does not make it possible to
contract Y->Z if Y->Z was not contractible originally.  This invariant was
broken by the existing rule that the root of a cluster could not be a shape
consumer op.  This CL also moves this rule to partially_decluster_pass, which is
also generally a better place for this sort of thing.

This CL also changes mark_for_compilation_pass to VLOG out some extra
information to help debug clustering decisions.  With this CL when we are unable
to contract an edge because it would create a cycle we know for sure that edge
will not become contractible in the future.  So we can print out more
information in that situaton without the risk of the information being
misleading.

PiperOrigin-RevId: 244508182