[XLA:CPU] Fix numeric stability of exp function.
Our implementation of exp(x) decomposes the problem as

 exp(x) = exp(x' + n log(2))
        = exp(x') * exp(log(2))^n
        = exp(x') * 2^n

For an appropriate integer choice of n, we can restrict x' to being in the
range (-0.5, 0.5) and then use a polynomial approximation for exp(x') in that
range.

The old implementation had the bug that it computed the multiplication in
exp(x') * 2^n naively: Calculate the LHS, calculate RHS, and then multiply
them.  This causes an overflow to inf when n == 128 (because the max f32
exponent is 127), even though the final answer does not necessarily overflow
(if exp(x') < 1).

This patch fixes our implementation to handle the case when n == 128.  It does
so by restricting n <= 127 and "shifting" one factor of 2 into the exponent x'.

  exp(x) = exp(x') * 2^128
         = exp(x' * log(2)) * 2^127.

This makes the polynomial used to approximate exp(x') less accurate (because
it's most accurate only on the range [-0.5, 0.5]), but it seems to work well
enough, certainly much better than our previous implementation which would have
simply overflowed.

PiperOrigin-RevId: 247102119