[TF] Check that TF inlined Variants properly call value destructor on move.
This new test mimics the behavior of having a class with a custom destructor
and a std::shared_ptr; if the destructor is not called a memory leak occurs.

Here we test the memory leak by doing simple counting with a static class member
variable; we manually override the move-constructor and move-assignment to
mimic the their default behavior (but with liveness counting).

PiperOrigin-RevId: 247427651