[XLA] In the HLO parser, disallow trailing text after parsing a single instruction.
The HLO parser has two modes, "whole module" and "single instruction".  If some
text fails to parse as a whole module, we attempt to parse as a single
instruction.

The bug fixed in this patch is, we weren't checking that, after parsing the
single instruction, we hit the end of the string.  Thus a string that contained
*two* instructions would parse successfully and return the first instruction,
ignoring the second one.

We made exactly this mistake in at least one unit test.  Moreover the thing the
test was trying to check was also not working!

This patch fixes the test and the parser.  Now that the test is doing what it
intends, it finds an actual bug, so we leave the test disabled for now.

PiperOrigin-RevId: 247512316