Rewrite messages of errors raised at raised graph construction to include a stack trace relative to original code before conversion. This is done by attaching metadata (as new attributes) to errors raised in converted code. This metadata includes the necessary information to indicate the location in the original code that corresponds to these exceptions' actual stack trace. The location is then included in the error message. In the common cases, the type of the error raised is preserved, however unknown error types will be replaced by a dedicated AutoGraph error when leaving the converted code. The error type remains the same within the converted code.
The approach is much less intrusive than wrapping an error, as it allows preserving the exception type within graph code. The risk of conflicting with existing attributes is extremely low, and since the names are internal, they are easily changed.
However, it would be better to reserve standard names for this.

PiperOrigin-RevId: 249109716