Add extra_gen_deps option to tf_gen_op_wrappers_cc
This allows the caller to add extra dependencies to the wrapper
generator if necessary.

The rationale is that I'm adding the ability to register ops in C as
part of the modular tensorflow effort. These ops will live in a separate
location (e.g. //tensorflow/c/kernels/ops/...), but will still need to
be part of wrappers for ops registered in other places.

PiperOrigin-RevId: 252916394