Add experimental `optimize_for_static_graph` option to `tf.ConfigProto`.
The current Session implementation maintains two copies of the graph in a `GraphExecutionState` object:

1. The original `GraphDef` protocol buffer.
2. A pointer-based `Graph*`.

The original protocol buffer is only needed when `Session::Extend()` will be called (or in the rare case that `GraphOptions::place_pruned_graph()` is true). For many users, including most inference serving users, `Session::Extend()` will never be called, and retaining the protobuf wastes memory. By setting the new option to true, the user can declare that `Session::Extend()` will never be called, and the runtime can avoid storing the graph.

PiperOrigin-RevId: 255401474