Log state before it is deleted in collective ops callbacks.
We had previously added VLOGs to track the start and end of each collective
opkernel execution.  However, the VLOG for the end of execution happened after
the DoneCallback from executor.cc finished, which cleaned up the state that was
being logged.  This change fixes this issue by moving the VLOG to before
executing the cleanup callback.

This issue only arose with TF_CPP_VMODULE containing "collective_ops=n", n>=1.

PiperOrigin-RevId: 255680928