Add modeling of NextIteration node as MLIR operations in the tf_executor dialect
The NextIteration node in TF is split in two separate operations in MLIR: a source and a sink. This allows to break the back-edge and keep a linearized DAG to represent the graph.

The two operations are paired by a value produced by the source and consumed by the sink using a newly introduced Token type.

PiperOrigin-RevId: 255965178