Store inferred shape and dtype for VariantType in TensorFlow
TensorFlow variant type is used to support arbitrary custom C++ data types and the underlying type isn't available in the GraphDef. However, many ops consume or produce variant types of a particular underlying type. In such cases, shape inference function can be used to infer a list of dtype and shape for each of the outputs. Interpretation of this list depends on the underlying type. For example, TensorList types infer element dtype and shape. The same mechanism is used for resource dtype as well.

Inferred list of shapes and types are stored as list of TensorTypes in VariantTypeStorage and printed as a comma separated list within angular brackets.
For example, !tf.variant<tensor<3xf32>, tensor<2xi32>>

Specific changes:

* Introduce VariantTypeStorage for VariantType instead of using the default storage type.
* Add support to print, parse and verify VariantType.
* Construct VariatType with subtypes if inferred shape and types are available while importing GraphDef.

TESTED with unit tests

PiperOrigin-RevId: 256448352