[ROCm] Skipping subtests that check bit pattern result from zero division.
ROCm platform currently does not produce the same (as CUDA platform) bit pattern result from zero division.

This commit skips subtests (within python unit-tests) that test this functionality. The "skip" is guarded by the call to "is_built_with_rocm()", and hence these unit-tests will not be affected in any way when running with TF which was not built with ROCm support (i.e. `--config=rocm`)