[XLA:Python] Generalize stream pacing mechanism.
Currently on CPU and GPU there is no limit to how many operations the host can enqueue on the device stream.

On GPU this doesn't usually cause a problem because the allocator is logically synchronized to the tail of the compute stream, and so we can free and reuse memory for operations enqueued on the stream.

On CPU, the allocator is logically synchronized to the head of the compute stream, which means that the allocator cannot reuse buffers between operations enqueued on the stream. This means that the memory usage is proportional to the number of enqueued operations, which can rapidly blow up.

Add a semaphore class and use it to set a moderate limit on the depth of the queue (32). The existing "synchronous" mode, used on TPU at present, is a special case of this support where the queue depth is 1.

This may help with https://github.com/google/jax/issues/928 .

PiperOrigin-RevId: 257606960