Fix support for partitioned while loops in multi-process multi-device functions.
Previously, the function inlining logic would append a unique prefix to the ?frame_name? attr for ?Enter? nodes in each subfunction sent to remote processes. The graph partitioning logic relies on all partitions using the same ?frame_name? for the same loop, because these determine the rendezvous keys used in Send and Recv ops. Therefore, if a `tf.while_loop()` were partitioned across multiple processes, the rendezvous keys would not match, and the function execution would hang in the Recv nodes.

The fix provides an option to disable ?frame_name? uniquification when inlining a function body, and moves the inlining to the client process that partitions the multi-device function (instead of performing it at each remote worker).

PiperOrigin-RevId: 257844481