Add reallocation capability to bfc_allocator.
This commit mitigates external fragmentation in bfc_allocator by reallocation.
That is, although the sum of regions and unallocated bytes is larger than the
requested bytes but the bfc_allocator still fails to allocate a large enough
contiguous region to fulfill the request due to fragmentation. To avoid this
case, a relocation feature is implemented to deallocate free regions so that
a larger region can be formed.