Add executor-to-control dialect conversion pass
This pass convert an MLIR representation of TensorFlow graph from a mix of tf_executor
and tf dialects to TensorFlow Control Dialect (_tf).
This is intended for managing the transition only, the TensorFlow Control dialect is
ultimately intended to be removed after the GraphDef importer is updated to target
directly the tf_executor dialect.

PiperOrigin-RevId: 259133988