[XLA] BF16 propagation: fix bitcast-convert fusion root handling
Although we try to skip changing bitcast-convert, but when it is a fusion root,
we still change it to match the fusion output. In this case, we now add a
convert after the bitcast-convert, instead of changing the shape in-place. This
way we still get the benefit of reduced memory write in the fusion.

PiperOrigin-RevId: 259609312