[ROCm] Fix for the broken `--config=rocm` build
The following PR/commit breaks the `--config=rocm` build

https://github.com/tensorflow/tensorflow/pull/30997

It introduces a declaration of the variable "data_layout" , within only the #else section of a #if-else block, and then adds a reference to "data_layout" outside of that #if-else block. Since the ROCm build takes #if path, the declaration for "data_layout" is missing, leading to the compile error on the reference to it.

The fix is add the corresponding declaration for "data_layout" in the #if section.