Open function.Defun's attribute whitelist to include two more attributes.
Currently, Defun restricts the attributes that can be set on the function_def to a whitelist of keys that start with "experimental_". I don't really know the background to that. In order to support new attributes like "implements" and "reference" as part of RFC https://github.com/tensorflow/community/pull/113/files, and to have Defun also support these use cases, there are two options I can see:

1. Remove the restriction on attributes being "experimental_". I tried this but looks like we make assumptions in the code around this and not sure what can break as a result (an initial attempt had quite a few failures).
2. Extend the whitelist to include the newer attributes. This limits the extent of the change and hence has much less risk of breaking things.

PiperOrigin-RevId: 260532386