[tf.data] Extending the TF 2.0 support for `shuffle(..., reshuffle_each_iteration=True)` to work across different Python iterators for the same dataset.
To achieve its objective, this CL creates a `RandomSeedGenerator` resource and ops for creating and deleting the resource which is used to manage state for seeding the shuffle order of different Python iterators for the same dataset.

Note that the new functionality is not yet supported for (multi-worker) distribution strategies that clone the input pipeline graph created by user programs. To support this use case, we need a mechanism to clone the `RandomSeedGenerator` resource (and in general other resources).

Fixes: #27680
PiperOrigin-RevId: 260765646