Add new dtype behavior for keras layers.
No functionality added by this change is enabled by default, but it will be enabled very shortly in a subsequent change.

With the new behavior, Keras layers will cast inputs to the layer's dtype. In addition, the layer's dtype now defaults to floatx (which is typically float32) instead of defaulting to None and being inferred from the inputs.

In addition, mixed precision Policies have been expanded (which are now referred to as dtype Policies). Each Policy has a compute_dtype and a variable_dtype. Layers no longer store a dtype in self._dtype, but instead a policy in self._dtype_policy. Layer still accept a dtype or a Policy in the "dtype" constructor argument, and Layer.dtype still refers to the variable dtype.

This change to policies was needed to avoid breaking mixed precision as otherwise it would be impossible to determine the dtype to cast to when using the "infer_float32_vars" policy.

Within TensorFlow, I fixed all tests that break when the new behavior is enabled. I will fix internal tests outside TensorFlow after this change is submitted, before making the new behavior on by default.

A few docstrings and comments are slightly incorrect as of this change, as they assume the new behavior is on by default. Once the new behavior is turned on, all docstrings/comments should be consistent again.

PiperOrigin-RevId: 260980313