NFC: Refactoring Importer to create GraphDefImporter
This CL extracted the logic for inferencing the main function
signature out of the original Importer into a new subclass
GraphDefImporter. The original Importer class is renamed to
ImporterBase at the same time. The purpose of this change is
to prepare for a future SavedModelImporter subclass, which
will reuse lots of the logic already in the existing Importer,
but it does not need to infer the main function signature,
given that a SavedModel contains a list of functions.

This CL does not introduce functionality changes; it just
moves code around and adjusts function signatures for the move.

PiperOrigin-RevId: 262248246