XLA: Remove no-op reshapes
When reshaping to the exact same shape, these reshapes do nothing and clutter up the IR.

Note that this is only necessarily true because the xla_hlo reshape op does not have a dimensions attribute, which allows transposing the dimensions. That is not part of the HLO reshape, but it is included in the XlaBuilder client. If the ReshapeOp acquires that attribute, then this pattern would have to check for it before eliding the redundant reshape.

PiperOrigin-RevId: 263168172