Fix allocator in cases of sizes overflowing 32bit integer arithmetic in size_util.
Part of it was AllocateFast not checking if ptr_ is null
before using it (null deref with offset, so didn't look like a null deref).

Part of it was using round_up_pot with a large size_t value that got implicitly casted to int as round_up_pot took an int argument. This showed how it's safer to just templatize the helpers in size_util.h, make them accept either int32 or int64 (guarded in floor_log2, which is the only of these functions that cares).

I just changed allocator to use only signed types (std::size_t --> std::ptrdiff_t) because I didn't want to deal with the extra complexity of dealing with both signed and unsigned in size_util.

PiperOrigin-RevId: 264668215