Optimize FrameState::GetIteration() (in the PropagateOutputs() critical section).
Accessing the size of and indexing into a fixed-size InlinedVector is more expensive than indirecting via the raw data pointer. We call FrameState::GetIteration() many times during ExecutorState::PropagateOutputs() while holding the FrameState lock, which serializes concurrent node completion. This change reduces the time spent under the lock, and additionally optimizes for the common case where we are not executing in a loop (avoiding a DIVQ instruction when the iteration index is 0).

PiperOrigin-RevId: 267414728