Add SaveOptions.save_debug_info and associated plumbing for SavedModel V2.
(second try: original rolled back due to path sep mismatch in a unit test for Windows)

The main behavior change is that I chose to mangle the trace key to be:
  op_name "@" func_name

(global ops use func_name = '')
Originally it was just a simple concat: func_name op_name

I don't have a strong opinion on the specific mangling to be used, but I do believe that how it was done is collision prone and best to correct now. Specifically, I chose this form because:
a) func_name does not seem to have strong validation on its naming (and in practice can be quite varied in the presence of lambdas, etc)
b) op_name does have strong validation on its syntax and specifically excludes '@' (matches regex: [A-Za-z0-9.][A-Za-z0-9_.\\-/]*)

Given these points, what I propose should be collision free. I'm not sure I should be making this decision, though. Please advise.

The test coverage was pretty sparse for all of this and I tried to buff it up around my modifications.

PiperOrigin-RevId: 267674315