[XLA GPU] [NFC] Further unnested reduction emitter simplification
The 3D tile loop in EmitTilingKernel is not actually 3D, since block size of
all blocks except for the first one is hardcoded to be 1.
Using this assumption allows considerable code simplification,
at which point EmitTilingKernel becomes so simple that the function it's using
can be inlined.
Dramatically reduces the number of passed callbacks.

PiperOrigin-RevId: 268563920