Add SavedModelBundleLite, which saves memory by not storing a MetaGraphDef.
Most users of SavedModelBundle only use the embedded MetaGraphDef to access the nested SignatureDefs, and the remaining memory used by MetaGraphDef is wasted. This change adds:

1. A SavedModelBundleLite class that provides most of the same functionality as SavedModelBundle without storing a complete MetaGraphDef.
2. A pure-virtual SavedModelBundleInterface that SavedModelBundle and SavedModelBundleLite now implement, to facilitate migration.
3. An implementation of LoadSavedModel() that creates a SavedModelBundleLite.

Note that the Session returned in a SavedModelBundleLite does not support Session::Extend(), since that feature is rarely used in inference, and it increases the memory requirements inside DirectSession. It also does not support "partial run", which is a seldom-used and deprecated feature, but could be enabled if there is future demand.

PiperOrigin-RevId: 268904828