Lower TensorFlow StridedSlice op to HLO
Currently, this only handles the simple case with constant begin and end operands and no mask attributes except shrink_axis_mask.

StridedSlice op is lowered to HLO Slice op along with Reverse op to handle negative stride values and Reshape op to update the output shape to handle change in the output rank.

Added a new build for the HLO Slice op that computes output type based on the operand type and attributes.

PiperOrigin-RevId: 270360518