Get per-channel bias quantization parameters
The bias quantization parameters are derived from the inputs and weights. When
the input is per-tensor and weight is per-channel, the scale of the inputs is
broadcasted to each channel. The quantization dimension indexes and dimension
sizes from the per-channel quantization parametes need to be matched.
Otherwise, nullptr is returned.

PiperOrigin-RevId: 270840445