Transpose and reshape per-channel quantization ops
When transponse and reshape are passed through the per-channel quantization
ops, the quantization dimension of the per-channel quantized type needs to be
adjusted as well.

For transpose op, the new quantization dimension is set to 0.  For reshape op,
the new quantization dimension is set to 3, at the same time, the scales and
zero points need to be broadcasted to match the length of the new dimension.

This is based on the assumption that the transpose and reshape ops are from
preparing conv2d and depthwise_conv2d ops.

PiperOrigin-RevId: 272011891