Define the data format to import quantization information from different sources
We are expecting the quantization information of different tensors in the graph
can be imported from different sources (calibration tools, manual inputs,
etc.), so this CL provides a standard format for these tools. When the
quantization passes are applied, a separated pass can read in the information
in this format and annotate the graph with its quantization representation
(e.g.  inserting Quantized/Dequantize ops).

PiperOrigin-RevId: 272240228