Move part of Layer docstring to Policy docstring.
Also make other improvements to Policy docstring.

I reduced the size of the float64 section, since it's mostly redundant with the mixed precision section. Also, instead of recommending to use floatx for float64, I recommend setting the global policy. Admittedly, this is a bit weird as the full name for a policy, tf.keras.mixed_precision.experimental.Policy, has the words "mixed precision" in it. But I want to be consistent, and using floatx for float64 and policies for mixed precision would not be consistent.

PiperOrigin-RevId: 272471221