Add ConfigProto.experimental.disable_output_partition_graphs().
Currently, DirectSession retains a copy of the Graph for each partition of a subgraph, in order to be able to respond to a Run() call with `RunOptions.output_partition_graphs == true` or build a cost model (which is explicitly enabled via `ConfigProto.graph_options.build_cost_model > 0`). This experimental option makes it possible (in conjunction with not enabling cost models) to release those Graph copies.

PiperOrigin-RevId: 272658318