Merge islands that independently feed into fetch.
The approach we take here is to take all the fetch inputs and pass them through a dummy island that just yields them. This allows our use-def based coarsening algorithm to merge the independent islands.

Thanks to Andy for the idea of doing it this way.

We still can't merge an island that doesn't feed into a fetch at all with this technique.

PiperOrigin-RevId: 272907508