Fix tf.assert_equal issue when one tenor is empty and another is non-empty
This fix tries to address the issue raised in 32082 where
tf.assert_equal([], [1.0]) doesn't raise error.
The reason was that in assert_equal `[1.0]` was broadcasted
as `[]` and equal was in place in that situation.

This PR updates the _binary_asesert so that it will check if
x, y are both empty or both non-empty. If one is empty and another is
non-empty, then assertion throws exception. This change is to not impact
other ops that depends on the broadcast behavior.

This fix fixes 32082.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>