[XLA] More robust handling for input/output aliasing in mem space assignment.
In this CL, we substitute HloBuffer-based allocation decisions with
HloValue-based ones so that we don't unnecessarily couple allocation decisions
for values that have input/output aliasing relationships. We introduce required
memory assignments for inputs and outputs which prohibit prefetching and can
force evictions to ensure inputs/outputs are properly assigned in the default
memory space. In a future CL, I will also add support for inputs/outputs that
are pre-set to be in the alternate (fast) memory space.

PiperOrigin-RevId: 273337348