Add the function RunDatasetOp()
Different from `MakeDataset(...)` which returns a `Dataset` object,
`RunDatasetOp()` executes the kernel and returns the produced
outputs as a tensor vector. It is helpful for some dataset operations
that do not have an internal customized `Dataset` class (
e.g. `ReduceDatasetOp`). This function can also be used to run other
types of operations and produce outputs.