Add `strict=True` to tf.cond()s to avoid unpacking single-element tuples and lists.
This was causing weird behavior in the (unusual) case where a dataset was mapped to a one-element tuple:

    dataset = dataset.map(lambda features, columns: (features,))

wherein depending on the divisibility of the batch size and whether we're dropping remainder in the dataset, we would take one of two code paths internally and only sometimes return a tuple back (return type was either [(PerReplica,)] or [PerReplica]).
This unpacking was likely unintended, so we'll revert back to the (less surprising) behavior of returning the same shape as was passed in.

If we want this case to return a nest of PerReplicas instead, we could add a line to get_next_as_list_deprecated() to flatten single-element lists and tuples instead.

PiperOrigin-RevId: 275970835
Change-Id: I056dc79590ee264f9ea295c8c4d3df7478e78c44