Create MultiProcessRunner class in multi_process_runner module.
multi_process_runner.run() is currently a blocking call that returns after all subprocesses return or it times out. To implement partial process termination (where some processes are killed while others are not), we'll need to make run() asynchronous, and it will be programmatically natural to have run() belong to an object that has states, and the object will be later invoked methods such as termination of some processes.

With this I hope to finalize the api of multi_process_runner before applying it to more tests.

PiperOrigin-RevId: 276132334
Change-Id: If791681192c7e74477188bebf466e059cae966e9