Support tuple tracking for checkpointing and SavedModel
As usual this isn't terribly neat or pretty, but the ratio of people annoyed that this hasn't been happening to ugliness is pretty favorable.

We'll only wrap tuples if they have a mutable sub-structure or a trackable element. This works because the tuples themselves are immutable, so no late-added dependencies.

Also wraps namedtuples and subclasses, although I'm happy to drop that if there's a case where it changes semantics.

From SavedModels tuples revive as lists since we need to re-create the object graph and then add the edges (mutate) as things are structured now. I don't imagine this will be a huge issue; mostly we need somewhere to put variables that are tracked inside pre-saving tuples.

PiperOrigin-RevId: 276361351
Change-Id: I3bc80daba823b2a7ee848eea2f1bdfa0a99a33b9