[TF:XLA] Implement MatrixSolve in terms of QR decomposition.
This implementation is effectively just a MatrixInverse followed by a batched
dot, so all of the same disclaimers apply as for MatrixInverse (QR decomposition
likely to be slower than LU, doesn't support complex numbers). This is a
short-term solution to avoid having to use workarounds for tf.linalg.solve on
TPU. Performance can be improved in future (e.g. when LU is implemented).

PiperOrigin-RevId: 276557074
Change-Id: I2eb2420eb24a057f3ea6a0e9b340d86552432077