Improvements to TFShouldUse
(1) Add an option to raise an exception if an object is not used by the end
    of tf.function tracing.

(2) Clean up support in eager mode (by improving introspection).

This requires adding a slow path for checking if an object is an instance
of a subclass of EagerTensor created by TFShouldUseWrapper.
In this case, pull out the wrapped EagerTensor and operate on that.

I also had to move a method from func_graph.py to ops.py to avoid a circular
dependency between func_graph and ops/tensor_array_ops (ick!!)

PiperOrigin-RevId: 278627366
Change-Id: Idb4bb9bb0783f5d3ae66b568f101bd4b0be27b57