Add TF_SetStatusFromIOError to the C API.
This is needed so that plugins can set Status (TF_Status) from an error code without needing to directly call tensorflow::IOError. We don't want to call tensorflow::IOError from plugins because that would enlarge the space that we have to watch out for ABI compatibility.

Part of work for modular filesystem plugins. For more details, consult the RFC at
https://github.com/tensorflow/community/blob/master/rfcs/20190506-filesystem-plugin-modular-tensorflow.md

PiperOrigin-RevId: 278921539
Change-Id: I8ac3e3a0d8147f901aefa1e301d774fff6055a99