Cache type test in ZeroDimArrayToScalar and avoid TensorShape::RemoveDim.
This change implements a ConverterState structure that is used in the
conversion process in to ways:

The first is as a place for caching a check performed by the
ZeroDimArrayToScalar function, which is called repeatedly when one
calls tf.constant().  It turns out the most expensive part of this
check is testing for types, and most of the time the type does not
change between calls.

The second is as a place for storing the inferred tensor shape and
data type. The conversion code was modified to use those values
instead of repeatedly constructing new shapes (with
TensorShape::RemoveDim).

This speeds up all conversions to various degrees. In one example, the
constant_20x20x20_double_list_to_float64_tensor benchmark went from
78?s/example to 54?s/example in my machine, a 30% gain.

PiperOrigin-RevId: 279876889
Change-Id: I8363f31501f080c95f10c49bb6cfef2ccb91762d