Check for JSON conversion subprocess failure in visualize.py
The flatc subprocess call was quietly segfaulting on my .tflite model, but visualize.py wasn't checking the return code and was generating out-of-date HTML using stale JSON from an earlier, successful conversion.

Instead of os.system(), use subprocess.check_call() to check that the flatc invocation succeeds, and raise an exception if it doesn't.

PiperOrigin-RevId: 280010077
Change-Id: I6bb2526d5ce6a67fff44dae250f840fa72521646