Isolate small zeros/ones optimization to graph mode
Trying to create a small constant instead of a fill op turns out to be
more expensive in eager mode due to the dim calculation. This is true
even if we change np.prod to instead TensorShape::num_elements(). It is
simply faster for us to perform the fill op instead. Note that using the
fill op is still faster than simply returning a constant in eager mode.

PiperOrigin-RevId: 280954363
Change-Id: I7e3c5a7ab6e52ee2d1029469871579eb38bf2502