Made minor changes to make code more suitable for GCI challenge. Refer #34005
The changes have been made in order to make the code suitable for a Google Code-In Documentation and Research challenge.
The task now is to clean the documentation so that it is more clear. There are too many code blocks in-between the text.
Some examples are placed in the args section for illustration. A better way to handle that would be to add those examples in a separate "Examples" section (consistent with other docs) and add pointers in the "Args" section.
Similarly, some other code elements could be eliminated from the text by adding pointers.
Clarify the crops part in the "Args" section further.