Split PRun-related fields from DirectSession::RunState into DirectSession::PartialRunState.
Each DirectSession::Run() call must initialialize a RunState(), which currently involves constructing fields that are only relevant in the rare PRun case. This change creates PartialRunState as a new subclass of RunState, and moves the PRun-related fields out so that they are only constructed during PRunSetup(). The practical benefit is that we can avoid the unnecessary Notification calls (and attendant atomic operations) in RunInternal() and ~RunState().

PiperOrigin-RevId: 281338058
Change-Id: I347f077ed5edee5461bd398e8280c745125f09c7