Always use real exp when scaling in linalg.expm
Previously, for matrices of complex type, the linalg.expm implementation
would compute 2**squarings as complex type (despite being an entirely
real computation). However, there's no GPU kernel for complex
exponential.

Instead, we can compute 2**squarings as a real type. We still do the
same number of casts as before (one), but now the cast happens after
the exponential instead of before.

As a result of this change, the entire linalg.expm computation can run
on GPU.