Introduce tfl.external_constant op.
tfl.external_constant op models constant values that live in the flatbuffer (outside of the MLIR representation). The location of the constant within the flatbuffer is represented as the buffer_index attribute.

The op will be useful to reference constant values located in flatbuffer within an MLIR program. The user need to guarantee (and make runtime checks if necessary) that the constant pointed by the buffer_index actually exists.

PiperOrigin-RevId: 284914069
Change-Id: I90b01b11dbe18190b8f54cce04cfda5416145700