Remove `device_set` check from collective param resolution.
This change fixes a subtle bug in `CompleteGroupLocal`, described below.  After
this change, `collective_ops_gpu_test_gpu` is no longer flaky.

During group resolution, the first device to be resolved creates a `GroupRec`
in `group_table_`, and initializes this `GroupRec` with the group's key, size,
and device type.  Subsequent devices access this group rec and check that the
group attributes match.

Because of the way mutexes are locked, before this change it was possible to
skip the device type check as illustrated in the following example.  Imagine a
group with 2 devices, CPU:0 and GPU:0.  CPU:0 creates the `gr` and initializes
its type to CPU.  It releases `group_mu_` but before it can lock `gr->mu` GPU:0
locks it first.  GPU:0 then skips the device type check because `device_set` is
empty.  This is incorrect, and the desired behavior is to return an internal
error for a collective group with different device types.

This change removes the `device_set` is empty check because it is not needed.

PiperOrigin-RevId: 286024716
Change-Id: Ib54defe5326426ab5cd1b692067b2b207f7a1d7e