Optimize the "standard order" case in SparseTensor::IndicesValid().
Most paths to IndicesValid() implicitly assume that the indices are in the standard, row-major, order. For example, SparseToDenseOp makes this assumption, and it accounts for the majority of time spent in IndicesValid().

In addition, this change caches the `InlinedVector<int64>::data()` call implied by indexing into `SparseTensor::shape_`, which avoids a branch every time the vector is accessed.

This change improves the throughput of the sparse_to_dense_op microbenchmark, which stresses IndicesValid(), by between 1.1% (BM_SparseToDense/977k) and 15% (BM_SparseToDense/3M).

PiperOrigin-RevId: 286196888
Change-Id: I79b3075b8325f5b7345c134de6c4a3028a6d33e2