[tf.data] Removing capturing cancellation manager inside of InstantiatedCapturedFunction.
We previously captured the cancellation manager to use it for `InstantiatedCapturedFunction::RunInstantiated` which only needs to be used in the GeneratorDataset destructor.

However, if the `GeneratorDataset` has not been finalized before calling its destructor, the cancellation manager captured in `InstantiatedCapturedFunction` would have the cancellation bit set by the time the destructor is called -- which means that the finalize method won't be invoked.

The effect of this change is that the finalize method will be invoked in the destructor but will not be cancellable and it also silences the warning previously issued when the GeneratorDataset destructor's execution of the finalize method is cancelled.

Fixes: https://github.com/tensorflow/tensorflow/issues/35100
PiperOrigin-RevId: 286484274
Change-Id: Ic5cb1430df80f92f52c00549d9534b5b4b295f30