Fix issubclass error in metrics_serialization_test.py.
On Windows and Mac, the following error would occur:

TypeError: issubclass() arg 1 must be a class.

This is because a function would be passed to the first argument of issubclass. Strangely, this error did not occur on Linux. Perhaps a Python bug? I will investigate further if I have time.

PiperOrigin-RevId: 286503059
Change-Id: I714211cd582a09cde252662000b0f30bff2abbeb