Simplify control flow in ops/functional_ops.py when creating PartitionedCall
The current branching is confusing: reviewers glancing at the code assume that
the first branch corresponds to the eager mode, and miss `or len(tout)`.
Handling two separate cases in this branch requires a correlated if- at the
end, which can only take the empty branch in eager mode.

Moreover, the logic is subtly incorrect: the user-provided `config` is not
considered in the second branch.

PiperOrigin-RevId: 286595076
Change-Id: I29fb3533f19923a8b63fe1cd7ea94848a74ac6ec