Replace --xla_gpu_disable_autotune option with --xla_gpu_autotune_level
XLA performs autotuning for convolutions and GEMMs. for each algorithm
it runs some checks to check for out of bounds access or functional errors.
The latter in particular can take a lot of time, increasing compile time
by orders of magnitude. This hurts end-to-end execution time.

--xla_gpu_autotune_level enables clients to set a level.
        0: don't autotune
        1: autotune with uninitialized data
        2: autotune with initialized data
        3: autotune with initialized data, and reinitialize for inplace case
        4: autotune with initialized data, reinitialize, and check
The deafult is 4, not changing the current behaviour.

Change some tests accordingly.