[XLA] Don't special-case bitcast uses in CopyAllocation; insert bitcasts as needed when processing.
Previously, we were special-casing bitcasts when creating CopyAllocations. In
CopyAllocation::Process, we were changing the bitcast operand with copy-done.
However, if the same original bitcast has two uses; one in default, the other in
alternate memory, this would have erroneously placed the first use in the
alternate memory as well. Now, we insert bitcasts if the shapes mismatch in
CopyAllocation::Process.
PiperOrigin-RevId: 288373548
Change-Id: I61ced1d74f599711f4cd6dab3970296b0742108b