[MLIR:TF/XLA] Variable runtime reformatting pass.
A pass that takes advantage of a loop to add ops that allow the execution to
avoid repeatedly formatting variables back and forth. The desired formatting
is determined by TPU program compilation, so this pass does not include how
to reformat the variables, but only inserts general TPUReshardVariablesOps in
proper places, and TPUReshardVariablesOps interpret the compilation.

PiperOrigin-RevId: 288633379
Change-Id: I6c52c63c56fe5151da1bd291996d141b11192f54