Use DepthwiseConvolutionConverter before ConvolutionGroupConverter.
A recent change in shape_inference required the usage of ConvolutionGroupConverter
instead of DepthwiseConvolutionConverter. This meant that the filter
shape got expanded before we called into cuDNN, which is less efficient
than handling depthwise convolutions directly with cuDNN.
Now that this change is reverted, go back to using DepthwiseConvolutionConverter.
However it cannot handle cases with batch_group_count > 1 if input batch
is not equal to batch_group_count. For this, we still need the
ConvolutionGroupConverter.

PiperOrigin-RevId: 290037172
Change-Id: I5b4a1f8eea92392e39ae9cce8b4122f86f7e992e