Provide mechanism for registering custom resource tensor resolvers for ACD.
ACD only looks at the direct resource inputs of stateful ops. This doesn't work for cases where ops access resources indirectly e.g. consumers of TPUReplicatedInput and in tf.data where the MapDatasetOp may be touching a resource but we need to add control dep from the ReduceDatasetOp. This mechanism will provide a way to notify ACD of the indirect resource accesses of an op.

PiperOrigin-RevId: 290063112
Change-Id: I329007eb99fce2dee9dda03593651992086d0b18