Fix tf.cast issue when python native data type is combined with dtype=tf.float64
This fix tries to address the issue raised in 35938 where the following issue exist:
```
>>> tf.cast(0.2, tf.float64)
<tf.Tensor: id=37, shape=(), dtype=float64, numpy=0.20000000298023224>
```

The issue was that, in tf.cast, 0.2 was first converted to a float32 dtype tensor,
then cast to float64 tensor.

This fix adds dtype_hint=dtype to convert_to_tensor so that native dtype is preserved.

This fix fixes 35938.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>