Add pass that populates empty tf_executor.island ops with a TensorFlow op.
When an tf_executor.island is empty and has no data results, it can be mapped to a tf.NoOp. Otherwise a tf.Identity/tf.IdentityN can be used. This is possible as an empty island can forward a data result and have control inputs. By populating a TensorFlow op in an empty island, it makes the Graph exporter a bit simpler.

PiperOrigin-RevId: 290771107
Change-Id: Ie0101a6d1933f0165560b36ce03eb2d7800a45e0