Add memory management to plugin interface.
Since on Windows memory allocated by one shared object must be freed from the same object, we need to provide a mechanism for plugins to provide their own memory management routines.

A previous change made core TensorFlow provide an allocator for plugins to use, however it is more cumbersome to get this allocator to be used in every function in the implementation provided by the plugin. Plus, we still needed to also provide a deleter.

This approach simplifies the interface and also allow one to alter these routines to debug issues with plugin's memory allocation.

Tested manually, both on Windows and in Linux.

Part of the work for modular filesystem plugins. For more details, consult the RFC at https://github.com/tensorflow/community/blob/master/rfcs/20190506-filesystem-plugin-modular-tensorflow.md

PiperOrigin-RevId: 291408627
Change-Id: I4fc919701e2bb39928453a02dd6019b559edbbb1