Fix flatbuffers link error
This commit:
https://github.com/tensorflow/tensorflow/commit/adf6e22e4af83afd55e0da3caa7e7959def1e6b6#diff-4bdae93a605044a27f9df49a92bcf398L129

Removed the linkopts section from the cc_binary 'flatc' from
third_party/flatbuffers/BUILD.bazel. That caused builds to fail with:

/usr/bin/ld: bazel-out/host/bin/external/flatbuffers/src/libflatbuffers.a(idl_parser.o): undefined reference to symbol 'cos@@GLIBC_2.17'
/dt7/lib/powerpc64le-linux-gnu/libm.so.6: error adding symbols: DSO missing from command line
collect2: error: ld returned 1 exit status

(This is GPU builds on ppc64le)

I tried passing `--linkopt=-lm` on the bazel build command but it still failed.
Restoring the original linkopts section does resolve this problem. (and it
seems better to have the linkopt here than globally for the entire build)

A simple `linkopts = ["-lm"],` works for ppc64le, but I figured for windows and
other platforms it was better to restore the original values.