Expose control ret nodes when converting a MLIR module to TensorFlow Graph.
This adds an option to capture control ret nodes which then can be used later and not be accidentally pruned away. Users of control ret nodes may need to remap nodes as control ret node names are not preserved. This is necessary as Graph does not store information defining which nodes are control ret nodes.

PiperOrigin-RevId: 292199204
Change-Id: I4074cc6b2d514ff89435b9f419a71f7417526ab0