Roll tensorflow/platform/path.h into tensorflow::FileSystem class.
Path separators, joining and parsing is dependent on the fielsystem. Even on
windows, if someone tries to access GCS, they will need to use "/" as path
separators. Therefore, these functions need to be coupled with the filesystem.

This is an update for the old filesystem classes. Modular filesystems will be
updated by mihaimaruseac@

PiperOrigin-RevId: 292625664
Change-Id: Ide619cc2d3b3fa05af5e0907b873cca96033ebc5