Introduce a data type parameter to EagerContext::SelectDevice.
This is in preparation for fixing a couple of bugs related to ConvertToEagerTensor,
one of them about device placement of string tensors. The final fix will have
ConvertToEagerTensorUncached calling SelectDevice, which will then need the data
type parameter introduced in this change.

Since the method is currently called in only one place, and that place has a
DT_INVALID hardcoded parameter value, the change has no effect until
the changes in ConvertToEagerTensorUncached come in. I created a context_test
anyway to make sure the method is doing what it is supposed to do. This turned
out to be valuable enough to justify this otherwise small change.

PiperOrigin-RevId: 292934896
Change-Id: I4aa1e91904fb14b3e60f6f1b86ea699dd66c03dd