Add separate registry for function graph optimization passes in function runtime.
This adds a separate registry similar to OptimizationPassRegistry but for invoking function graph optimization passes. Other fields that are necessary but not exposed via OptimizationPassRegistry are exposed in this new registry (e.g. control ret node names). This is also a workaround to moving MLIR dependencies into the runtime as the TensorFlow Graph -> TensorFlow MLIR converters depend on the runtime.

PiperOrigin-RevId: 293692159
Change-Id: If616fe4d8080f08f4d1513429fec222417d5fcfe