Extract GpuAsmOpts struct into its own header file.
This will make it possible for libraries to depend on just the library
exporting that header if they want to use GpuAsmOpts, but not anything else
from asm_compiler.h. This makes it possible to compile some libraries that
don't need a GPU, but currently depend on asm_compiler target which is only
available if a GPU is configured.

PiperOrigin-RevId: 293774059
Change-Id: I234278cc4402d9934616a38a2902a6078b48299c