Add indicator in GraphOptimizationPassOptions whether not the Graph was generated from a Function.
By having some indicator for if a graph is derived from a function, graph optimization passes can then selectively choose whether not to run. For example, it may be possible to have a function optimization pass that when ran (only on function graphs), an associated graph optimization pass should not run.

PiperOrigin-RevId: 294478764
Change-Id: Ic407e81d9b528ab1b532c05ff157654851812566