Enable splitting out TFL runtime verification from op verification.
This enables use cases where the elemental types of the operations during
conversion/optimization may differ from what the TFLite runtime supports. The
approach followed here is pretty local and it could be changed to something
more general, but this allows verifying TFLite ops without verifying supported
runtime until such time.

Made the change in as small a way as possible and so the op verification could be tightened, but no reduction in verification on the path to TFLite runtime.

A downside to this local approach is that it makes the autogenerated
documentation less useful for folks interested in seeing the runtime supported
types/constraints.

PiperOrigin-RevId: 295857615
Change-Id: I5fa1bb8d83b740fe359e8551e22907fcf6fdfeb7