Support waiting on local async mirrors
The primary change is merging the local mirror maps and having the value
being a pair of an EmptyLocalTensorHandleData and LocalTensorHandleData.
The EmptyLocalTensorHandleData contains the mutex necessary to perform
waiting.

We further optimize sync execution by adding the AddLocalMirror API to
TensorHandle. This avoids creating an empty mirror just to set it a few
moments later. We also avoid a heap allocation of CopyToDeviceNode when
executing in sync mode.

As part of this change we also make poisoning mirror-aware for both
local and remote mirrors.

PiperOrigin-RevId: 297327607
Change-Id: I97f004cc6e4b608857e1f54ebb8be34921a15e5f