[MLIR:TF/XLA] Stack decomposition
Turns a stack to two variables (represented a a new pseudo op tf.MlirLocalVarOp) for the buffer and the size. Then changes the pushes and pops into operations with these variables.

After this pass, a resource lifting pass can remove the local variables as well.

I added this pass in the ConvertMLIRToXlaComputation() because it needs to infer the full shape of a stack.

Also relax the resource lifting pass' checks for stack-type resources, to allow the stack pass to run after the bridge-time lifting pass.

PiperOrigin-RevId: 297954689
Change-Id: Ie87e5255975f4213b48f1d1a226e2afcb10a4963