[TF:MLIR] Enable source information in the MLIR importer.
Change the MLIR importer to use SourceMgrDiagnosticHandler to get the location
for graph nodes. This enables source information in the error message when
source debug information is provided to the MLIR importer. If source debug
information is not provided, the error message won't be changed.

Add a test case to test source information in the error message.

PiperOrigin-RevId: 298358883
Change-Id: Ia92b185d1b83dfca2d3fbe700e30cc5ca80f48e3