Refactoring of TensorFlow runtime support for executing multi-device functions.
The main change is that rendezvous factory is moved from FunctionLibraryRuntime::Options to ProcessFunctionLibraryRuntime constructor. This avoids the complexity of piping the factory method through the executor to the runtime (which is not always done correctly in the current implementation) where it is actually needed.

This CL also makes tf.data's GeneratorDataset functions to use multi-device function backend.

Both of the above changes are important steps towards making it possible to support Python generator + Keras on TPU.

PiperOrigin-RevId: 299715147
Change-Id: I53ab7bf9e3a9d3a1a202ce96d54b78ba43500409