Fix <unknown> shape issue for sparse.transpose
This PR tries to address the issue raised in 37638 where
the shape after sparse.transpose is <unknown> even though
the input shape's rank is known. This PR relexed the shape
to only use rank (vs. is_fully_defined) to populate the shape.

The other issue raised about `sparse.reduce_sum` is a little challanging
as the shape is inference from C++ `SparseReduceSumSparse` ops which
will output unknown shape anyway. For that reason this PR does not
address the `sparse.reduce_sum` issue.

This PR fixes 37638.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>