Add multi worker mirrored strategy combinations to enable easier testing
It's required to use it with @run_in_each_worker.

When a test method is annotated with @run_in_each_worker, we call
multi_process_runner.run() under the hood. The test method is thus executed on
*each* worker, and is not executed in the original process. The original process
will assert on the success of the test on all workers.

This is very slow at this moment. Do not use it for a large number of tests.

PiperOrigin-RevId: 304516887
Change-Id: I6006cb53354c00f5f2cbedf94ff843a498bce5b4