Fuse Relu(X) to the affine ops
We don't use tablegen source patterns to match the Relu(x) ops following the
other ops. Instead, some logic is added to the existing patterns to fuse the
ops for Relus and append the activation function to the kernel name. This will
increase the number of kernels, but on the other hand, it makes kernel scale
function definition simpler.

PiperOrigin-RevId: 305400470
Change-Id: I664bdd8fbd01a5d1e9b9d15383e44bd4e44f4828