Fix issue in tf.image.extract_glimpse
This PR is to re-apply PR 12829. While 12829 was merged before,
for some reason it was reverted at one point. The guess is that
there are some internal testing that caused the revert. This
PR will try to submit again, and fix any internal tests that fails.

This fix tries to fix the issue raised in 2134 where
`tf.image.extract_glimpse` does not work as expected
when `centered=False` and `normalized=False`

This fix fixes 2134.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>