LSC: Allow targets to pass with `-fsanitize=null` under `--config=asan`.
This is a manually-generated change. Ensure that behavior and style are
preserved by carefully inspecting the diff.

This CL fixes one of the many targets in TAP that test with ASan but fail
with `-fsanitize=null`. We would like to enable the Null Pointer Sanitizer
in builds that use Address Sanitizer globally, and this change gets us one
step closer to that goal.

PiperOrigin-RevId: 306704264
Change-Id: I15fc65a7e3c7b3ed81e5b72038b070c6b847ff3f