Preserve dict ordering after regroup() with best effort
dict ordering may or may not be guaranteed, depending on Python version and
which dict implementation it is. But we still would like to preverve the
ordering on a best effort basis, so that problems caused by incorrect
assumption on the ordering doesn't look that it works without distributed
strategy, but breaks with strategy, which is very hard to debug.

PiperOrigin-RevId: 306898958
Change-Id: I660f0cd19e23f7991f5d0420e7232a94701a076f