[FunctionLibraryRuntime] Optimize single-component "multi-device" function dispatch.
This change enables the (single-device) `FunctionLibraryRuntimeImpl` to dispatch a multi-device function directly, if (i) it has a single component, and (ii) that component is local to the `FunctionLibraryRuntimeImpl` instance. This avoids the (microsecond-scale) overhead of preparing and remapping the inputs to and arguments from a multi-device function, which is important for clients (like tf.data) that invoke many fine-grained functions.

PiperOrigin-RevId: 307090208
Change-Id: I21820aa5b84360c2595b49c2e22d7a2b037819c4