[Build cleanup] Split "core_cpu_impl" into fine-grained targets (1/n).
This change starts the process of splitting up the monolithic
"//tensorflow/core/common_runtime:core_cpu_impl" target into fine-grained targets
that can be depended on individually. The change makes some small header and
forward declaration changes to avoid unnecessary transitive includes and
dependency cycles. It also revises the dependencies of tests in the
"common_runtime" directory so that they depend on fine-grained targets where possible.

It leaves in place the "interesting" dependency cycles of the
FunctionLibraryRuntime, Placer, and Executor, which I plan to clean up in a
subsequent pass.

PiperOrigin-RevId: 307133103
Change-Id: Id054e7929fa86b159539135c09f0023813f23f5a