Add builtin len() support of tf.data.Dataset for autograph
While working on autograph was looking for calculating the average
of a tf.data.Dataset, noticed that `len()` in autograph
only support tensor, tensorarray and tensorlist, but not tf.data.Dataset.

I think it makes sense to add tf.data.Dataset as well as `len()` is too widely
used in many cases.

This PR adds `len()` support of tf.data.Dataset for autograph.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>