Rolling forward "[Build cleanup] Split "core_cpu_impl" into fine-grained targets (1/n)."
The previous version of this change caused linker failures when linking pybind modules on Windows. The "symbols_pybind.txt" file, which is only used on Windows, lists the additional symbols that should be exported from pywrap_tensorflow, and (crucially) associates them with the name of the cc_library target in which they are defined. Since several targets moved from "core_cpu_impl" to different libraries, I needed to update this file with the new library names. In addition, I updated the target "//tensorflow/python:win_lib_files_for_exported_symbols" to list the new targets, so that they would be scanned during .def file creation.

PiperOrigin-RevId: 307834347
Change-Id: If15f70e67e173da7273c837474047c9d8aa7a87c