Add Python bindings and some tests for the parallel device
Still quite experimental; no public API and lots of TODOs.

Since the parallel device uses the C API, the pybind extension needs to re-include the parallel device sources so it can use the copy of the C API in pywrap_tensorflow. This is pretty ugly, but I don't see a way around it until pywrap_tensorflow relies on libtensorflow.so as the single source of the C API.

PiperOrigin-RevId: 307840967
Change-Id: Id8e7e72b14e2e5a2886c6025c7ef6f92f71a156c