Add partial ExtractOutsideCompilationPass for extracting CPU computation from TPU cluster.
Some operations are marked with `_xla_outside_compilation` attribute and should be extracted to run on the CPU rather than TPU device.  This is done by extracting the outside compilation clusters to separate parallel_execute regions and run on the CPU host.  This change focuses on identifying and moving the outside compilation clusters to separate parallel_execute regions.

PiperOrigin-RevId: 308178946
Change-Id: I3855321225a244c5826f2c8e52790a3291ba1f9c