Refactor SyncOnReadVariable update methods to be consistent with MirroredVariable.
This brings the behaivor of SyncOnReadVariable consistent with MirroredVariable.
E.g. previously it always returns a tf.Operation in cross replica context.

After the last refactoring SyncOnReadVariable only needs to override _update_replica to have the desired behavior.

This change also moves assign* and scatter* override to DistributedVariable
level.

This is part of the effort to make DistributedVariable.assign* returns a variable.

PiperOrigin-RevId: 308384727
Change-Id: If936ca5589bcffd9b77653657d6ac2b860b577d7