[WhileV2] Fix potential stack overflow when kernels execute inline.
The previous implementation of the `WhileOp` uses callbacks to handle the
result of the cond and body functions. While this is reasonable (and necessary)
when the kernels execute on the inter-op threadpool, it can lead to stack
overflow when inline execution is specified and the number of iterations is
large.

This change adds a non-callback-based version of `WhileOp` that will be
dispatched to if either (i) the `OpKernelContext::run_all_kernels_inline()`
method returns true, or (ii) the `WhileOp::Compute()` method is invoked.

PiperOrigin-RevId: 309264480
Change-Id: I4697d6b080caca75a5c3f555063bea32a0de9987