Partial rollback of "[Single-threaded executor] Optimize the handling of `OpKernel::const_tensor()`."
The previous version had a bug which allowed the value of an argument tensor to be mutated if buffer forwarding were applied to that tensor. This change switches from using `Entry::const_tensor` to `Entry::val` (which makes a shallow copy and increases the refcount on the argument tensors, inhibiting buffer forwarding). It also updates a test to include assertions that were failing with the broken code.

PiperOrigin-RevId: 309317692
Change-Id: Iac0f991b9bf44431ae0247e4907b6e87fa116ccb