Do not fail if local mirror already exists
It is possible for multiple threads to trigger a mirror copy at the same
time. If this occurs we need not error out. In async mode we can we a
bit careful and avoid duplicate copies when adding the empty mirror. For
sync mode we do not need to add much protection as the only thing that
would happen is an extra copy.

PiperOrigin-RevId: 309743954
Change-Id: Ic9106e6ae712a2caa40a958eaa90b5e267e588f6