Support XlaCompiler arguments of type kConstant in XlaCompilationCache
These arguments types are generated in two ways:

1) XlaCompileOnDemandOp, based on the xla op kernel requirement, provides inputs with constant constraint during compilation. This allows the op to get compiled to HLO and then the actual HLO op can get executed on the required device.
2) _XlaCompileOp gets some of the inputs as "constants" variadic input. These are the constants coming from GuaranteeConst op. TPUCompileMlirOp doesn't support these constants yet. I also found a note in the old bridge to revisit this idea as it has limited utility. (Couldn't find the note now.)

kConstant type arguments are supported by inlining the constant arguments after import and dropping those arguments from the signature. Mapping from the old set of arguments to the final list of arguments is set in the input_mapping of compilation result.

Some of the compilation tests are disabled with this change as those were going through the old bridge and requires further changes to support those.

PiperOrigin-RevId: 310040060
Change-Id: I121bb717e0376728bd727c7646116a04c77dcb53