Add  "//tensorflow/compiler/mlir/tensorflow:graph_optimization_pass_registration" to tensorflow/python/_pywrap_mlir.so
This commit adds build target //tensorflow/compiler/mlir/tensorflow:graph_optimization_pass_registration
to tensorflow/python/_pywrap_mlir.so, so that graph optimization could be available in tf-nightly pip wheel.

In the past while graph_optimization_pass_registration is available, it is not packaged with
tf-nightly so it is not possible to access graph optimization if installed from pip wheel.

In order to enable graph optimization pass for pip wheel, graph_optimization_pass_registration
target has to be included in "somewhere" that will be loaded when pip installed tensorflow
is imported.

A natural place would be libtensorflow_framework.so which is the core .so always loaded with `import tensorflow`.

It is possible to include graph_optimization_pass_registration to target libtensorflow_framework.so, see
last attempt on this route:
https://github.com/tensorflow/tensorflow/pull/39231

However, this caused many test failures like:
```
: CommandLine Error: Option 'help-list' registered more than once!
LLVM ERROR: inconsistency in registered CommandLine options
```

The reason is that many tests as a binary also have a copy of the LLVM in its binary, thus causing
multiple copies (one in binary, another one in libtensorflow_framework.so where many tests depends on).

Because there are so many tests, it is really hard to make all the needed changes without break somewhere else.

This commit takes a different approach to include graph_optimization_pass_registration to
tensorflow/python/_pywrap_mlir.so. This shared object is dedicated to mlir related APIs. The current exposed one
is `tf.mlir.experimental.convert_graph_def`.

Because this tensorflow/python/_pywrap_mlir.so already depends on llvm, place graph_optimization_pass will avoid
multiple copies of llvm in multiple locations. This tensorflow/python/_pywrap_mlir.so is also loaded with
`import tensorflow` as part of the python binding.

Ideally it probably would be still preferrale to get the graph_optimization_pass into libtensorflow_framework.so.
That will be investigated further later.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>