[Executor] Optimize `PropagatorState::FindOrCreateChildFrame()`.
At present, the `PropagatorState` must look up the "frame_name" and "parallel_iterations" attributes on the `Enter` node's `NodeDef` each time it propagates its outputs. This information is mostly cached already in the `ImmutableExecutorState`.

This change makes the following optimizations:

1. Cache a `FrameInfo*` for each enter node and provide O(1) lookup using an `std::vector<FrameInfo*>`.
2. Add `FrameInfo::parallel_iterations`.
3. Perform `frame_name` and `parallel_iterations` resolution once at `ImmutableExecutorState` construction time.
4. Avoid building and storing `std::string FrameState::frame_name`, which is only used for verbose logging. Instead use the `uint64 FrameState::frame_id` as the key in all data structures, to optimize lookup. This is safe because we already depend on a lack of collisions between frame IDs (since the frame IDs are used in rendezvous keys when tensors are sent between devices).

This change also modifies the "executor_test.cc" loop microbenchmarks to cover different numbers of loop variables (which, in the lowered case, directly translates to the number of "Enter" ops).

PiperOrigin-RevId: 310467528
Change-Id: Id70d284e99fd7537156a3a5e10da827aea4791f9