Allow single element padding_values to be broadcasted to a structure for tf.data.Dataset.padded_batch
This PR is a follow up based on comment in 35900.

In 35900, the issue was raised where if padding_values is a single elment
and the dataset has a structured shape, then tf.data.Dataset.padded_batch
will return and error.

This PR adds the support to "broadcast" padding_values to match the same
structure as the dataset, which could be convenient in many cases (avoid
repeat the padding_values into the stucture manually).

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>