Add basic support for top level function/NameAttrList attributes in Graph -> TF MLIR importer.
Currently when experimental_implements is used in a tf.function, an _implements attribute of NameAttrList is added to a function. Top level function/NameAttrList were not supported in the general case. This adds a custom attribute to model NameAttrList, where NameAttrList.name will be a StringAttr and NameAttrList.attr will be a DictionaryAttr of <string, converted AttrValue>.

e.g.
```
attr {
  key: "_implements"
  value {
    func {
      name: "embedding_matmul"
      attr {
        key: "key1"
        value {
          i: 2
        }
      }
      attr {
        key: "key2"
        value {
          b: false
        }
      }
    }
  }
}
```

will generate
```
_implements = #tf.func<@embedding_matmul, {key1 = 2 : i64, key2 = false}>
```
PiperOrigin-RevId: 312393667
Change-Id: I15e866ee6ea2bfdeb7642e810dc1244d4d858e36