[TF MLIR SI] Don't constant fold, only consider result of folding
This results in less changes to the module during shape inference (e.g., only
shapes are changed, no constant nodes are created). Effectively this computes
the folded result and then just uses that information locally. Which is
conceptually more wasteful (as a subsequent canonicalize pass may need to
recompute these) but is less surprising and avoids dropping attributes during
this part.

This adds an option to allow constant propagation from caller to callee and vice versa to retain current behavior. But these should be subsumed by more general constant propagation.

There is still additional changes that need to be made to avoid doing needless
computations here, this mostly focuses on decreasing graph mutations.

PiperOrigin-RevId: 312509934
Change-Id: I8252d73395a5d3a129e80e96b7c16f7abbbad97f