Reduce Functional.call internal per-Layer overhead by 95% when single Tensor passed.
Node now has a performance optimization for the common case where a single
Tensor is passed to a Layer during the Functional API construction phase.
This means there is almost no overhead imposed by Functional.call in this case,
and the remaining per-layer overhead is all from Layer.__call__.

For a Model like ResNet, this optimization will be active for all Layers except
for the Add() layers.

PiperOrigin-RevId: 313085525
Change-Id: I0b9821e954d17f1f101617449fd5f0851d6ba9c3