Reduce Layer.__call__ overhead by ~5%.
Uses ops.name_scope_v2 directly when in eager mode, since it does not increment
in eager mode.

Also when clearing losses, only checks for is_in_tf_function when build_graph=True.
When build_graph=False in the new Layer class, we know that we are in eager mode.

PiperOrigin-RevId: 313085589
Change-Id: I081ab1c592137445f918403c1293ae6a05758b38