[MLIR] Add ConvOp conversion from xla_hlo -> xla_lhlo -> linalg
Add conversion for the convolution op from the xla_hlo dialect
to xla_lhlo, and from xla_lhlo to linalg. The two conversion passes
in context here are -hlo-legalize-to-lhlo and -lhlo-legalize-to-linalg.  A part
of the code for the LHLO to Linalg conversion is borrowed and adapted from
IREE's xla_hlo to linalg conversion (https://github.com/google/iree/).

Signed-off-by: Uday Bondhugula <uday@polymagelabs.com>