Add canonical TestConv model to the memory arena threshold test.
The conv model has a different head/tail profile than the keyword benchmark model. This is due to the use of conv layers inside the model. Currently, conv/depthwise-conv are pretty memory intensive due to the quantization per-channel data (TfLiteAffineQuantization). Many of the models used in TFLM use conv/depthwise-conv and we should have coverage for these types as models on top of the keyword model.

Total memory usage on x86-64:
[RecordingMicroAllocator] Arena allocation total 12096 bytes
[RecordingMicroAllocator] Arena allocation head 7744 bytes
[RecordingMicroAllocator] Arena allocation tail 4352 bytes
[RecordingMicroAllocator] 'TfLiteTensor struct allocation' used 1680 bytes with alignment overhead (requested 1680 bytes for 15 tensors)
[RecordingMicroAllocator] 'TfLiteTensor quantization data allocations' used 1216 bytes with alignment overhead (requested 1216 bytes for 36 allocations)
[RecordingMicroAllocator] 'NodeAndRegistration struct allocation' used 560 bytes with alignment overhead (requested 560 bytes for 7 NodeAndRegistration structs)
[RecordingMicroAllocator] 'Operator runtime data allocation' used 136 bytes with alignment overhead (requested 136 bytes for 5 OpData structs)

PiperOrigin-RevId: 316154172
Change-Id: I3869dbff9dc22375e1d4fec50fd9dce0f6c14bc6