Fix error caused by incorrect rank assumption in tf.image.flip_left_right
This PR tries to address the issue raised in 40580 where an error
was thrown out when tf.image.flip_left_right process a tensor of unknown
rank.

The reason was that tf.image.flip_left_right assumes rank == 3 in case
of unknown rank.

This PR adjust to use array_ops.rank(image) to obtain the true rank
when unknown rank is present.

This PR fixes 40580.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>