Prepping for switch to ROCm 3.5+
Starting with ROCm 3.5 the underlying compiler used by hipcc will change from HCC to hip-clang. There will be a corresponding change in the HIP Runtime as well. This commit is part of a series which are intended to make the transition to ROCm 3.5+ easier.

Disabling some unit-tests (by adding the no_rocm tag to them) because they regerss after the switch to ROCm 3.5+

```
# XLA related failure
//tensorflow/python/eager:def_function_xla_jit_test
//tensorflow/python/keras:backend_test

# eigen packet then scalar JIRA ticket 236756
//tensorflow/python/ops/ragged:ragged_dispatch_test

# genuine miscompares
//tensorflow/python/kernel_tests:bias_op_deterministic_test
//tensorflow/python/kernel_tests:array_ops_test

# Memory access fault by GPU ... Reason: Page not present or supervisor privilege.
//tensorflow/python/ops/parallel_for:math_test
```