Loosen restriction on TPUReplicatedInput indices in TPUClusterFormation and TPUDynamicPaddingMapper passes.
The exact ordering is not necessary, but knowing what indices the TPUReplicatedInput ops map to originally is, for TPU Dynamic Padder. This updates TPUClusterFormation pass to store an attribute keeping track of the original indices of TPUReplicatedInput ops and TPUDynamicPaddingMapper pass to use the same attribute when remapping padding_map paddings (arg_index, padding_arg_index). Missing nonnegative indices restriction is also removed.

This is necessary to move graph pruning before island coarsening. In addition, this cleanup is necessary to make adding support for packed inputs (TPUReplicatedInputs with only one operand) simpler.

PiperOrigin-RevId: 318086700
Change-Id: I9d0373d23c5b5335b43b91f13afa50604344b65c