Prepping for switch to ROCm 3.5+
Starting with ROCm 3.5 the underlying compiler used by hipcc will change from HCC to hip-clang. There will be a corresponding change in the HIP Runtime as well. This commit is part of a series which are intended to make the transition to ROCm 3.5+ easier.

This commit makes some minor tweaks. It is already present in the `master` branch