Don't trace OpLayer in SavedModel.
OP Layers wrap a single Tensorflow op in a Layer class. Previously, SavedModel would wrap every internal layer call in a tf.function, so that the user can inspect individual layers in the loaded model. For TensorflowOpLayer, this is unnecessary because (1) wrapping a single op in a tf.function is very inefficient (2) the user is unlikely to individually inspect the autogenerated op layers in the loaded model.

This change also resolves the saving issue that occurs when a user builds a functional model while using the eager-computed results of `tf.shape(x)` as the input shape to another op layer.

An example to help illustrate:
```
x = tf.keras.Input((2,))  # Shape is (None, 2)
state = tf.zeros(4, tf.shape(x)[0])  # Expected shape is (4, None)
LSTM(inputs, initial_state=state)
```

Prior to this CL, the TensorFlowOpLayers generated for tf.shape and tf.zeros would be separately wrapped in tf.functions when saving. This results in `state` having a shape of `(None, None)` instead of `(4, None)`, causing potential problems when saving the rest of the model.

PiperOrigin-RevId: 318311978
Change-Id: I15099d8ba29c1d4facd3f88630f8e2651f22ae83