Only insert TPUCopyWithLayout when resource generator is on CPU.
The placement of iterator is not determined by the device attribute of tf.IteratorGetNext but by the device attribute of the alias resource generator op. The generator ops are either function arguments or tf.VarHandle ops. So, we modify the TPU dynamic layout pass to check device attribute of generator ops.

This CL modifies the Resource Alias Analysis to return the set of aliases of a given value. For efficiency, it now keeps a map of both resource value --> unique resource ID and vice-versa.

This improves the performance of few models running with MLIR TF/XLA bridge. For example, Resnet50 step time is reduced from 180 ms to 125 ms (on par with the old bridge).

PiperOrigin-RevId: 319255334
Change-Id: I4df9f26f480b580b8277caae981f06c3189e7bf4