Add error message for outfeed shape mismatch.
This error is triggered when the XLA program outfeeds data with a shape that is
not compatible with the shape expected by the runtime. This error can be
triggered by the end user. Currently, the error is not very helpful.
This change adds the two mismatched shapes to the error message,
in the spirit of the error message reported on CPU for the same error.

(see tensorflow/compiler/xla/service/cpu/cpu_runtime.cc line 289)

PiperOrigin-RevId: 320162796
Change-Id: I723e70f740a2b0734b7be575ea20d5bed19f55e7