Display verbose diff if API goldens changed.
Before this change, we only displayed a summary of the diff but this makes debugging harder. Displaying the verbose diff makes it easier to understand what changed, at the cost of increasing verbosity in logs. Since it is usually rare to have the API goldens changing, we might be ok with this change?

PiperOrigin-RevId: 321169903
Change-Id: I0a450356af3082b02e73517a858925e89cd5fee3