Improve error message in shape check
This PR tries to address the issue raised in 41504 where the
error message of:
```
InvalidArgumentError: Shape [2,2,2,2,2,2,2,2,2,2,2,2] would have more than 2**63 - 1 elements [Op:BroadcastTo]
```
is not intuitive.

The issue is that the construction of error message in tensor_shape.cc
did not copy the complete shape dims.

This PR fixes the issue.

This PR fixes 41504.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>