Use standard type conversion before trying default
When converting a non-tensor argument to an op we were using the default
dtype for conversion purposes. This leads to inconsistent behavior with
the type that is inferred during regular tensor conversion. For example,
python integers are either treated as int32 or int64 depending on their
size when converting with tf.convert_to_tensor.

Instead, we should try to use standard tensor conversion with the
expectation of getting the correct type for the argument. There are
cases where this doesn't work such as with empty tensors. In that case
we should fallback to use the default_dtype to ensure we have a valid
dtype.

PiperOrigin-RevId: 321874236
Change-Id: If3c6abb7eb15e1743fc5aec4a71c51ff964ef527