Dynamic Concat: Correctly replace root instruction.
It's very easy to make the mistake where we replace a bunch of uses of a instruction but didn't really set the root. Add a new function to do that (probably need another round of renaming to be explicit about where a function will replace the root or not).

Also use the new dynamic literal feature to simplify the test.

PiperOrigin-RevId: 322221542
Change-Id: I314c89118555b897d575f8b418a2ba187d579b07