[XLA] Allow copy-done to be scheduled earlier to avoid copy ordering issues.
We previously required that copy-dones to be scheduled right before the use.
We also require asynchronous copies to maintain pipelining order (no nested
copy-start/copy-done pairs). This could mean that a smaller buffer that has
shorter copy-start/copy-done duration may block a larger buffer due to copy
ordering. E.g. this situation might arise which is not allowed:

  small tensor in default mem---------------------->CS----->CD->use
  large tensor in default mem------------------>CS----------------->CD->use
====================================================================> time

This CL checks if there is an already committed asynchronous copy that violates
the pipelining behavior. If so, we attempt to move the copy-done earlier:

  small tensor in default mem---------------------->CS----->CD->use
  large tensor in default mem-------->CS----------------->CD----------->use
====================================================================> time

PiperOrigin-RevId: 322621813
Change-Id: I8287c11c96a6d71de86a5f2e22cf1846c26ef4f3