[TF:TRT] The converter should only access a cuda device when creating static engines.
When creating a static engine for an TRTEngineOp, the converter needs to use the
memory allocator associated with the cuda device that will be used to execute
the engine. As such, the converter saves the current cuda device id before
doing the conversion and restores the cuda device id before returning. These
steps are not necessary for dynamic engines because the creation of dynamic
engines happens at graph execution time and outside the converter.

PiperOrigin-RevId: 322640028
Change-Id: Ib0a1f8af1a333dc310822eda24881cab74c0845e