Simplify reshape handling in dynamic dimension inference.
When the code was first written last year I was still in the mindset that XLA has to infer dynamic dimension on its own, thus creating some fairly complicated logic.

Over the last few months I've been adding more hints in the bridge (by setting the dynamic output dimension in reshape's shape directly) so now instead of letting xla handle this itself, we can rely on the info passed from the client. This way we can reduce the complexity of the logic inside xla, which appearnly has CHECK failure.

PiperOrigin-RevId: 322870657
Change-Id: I7751b03e89af86342c12538261e43efdd97dfbe6