Add tf.data service journal reader and writer.
This change adds reader and writer implementations for journaling dispatcher state. The implementations read and write protocol buffers representing state changes. The records are written in TFRecord format so that we can detect corruption (TFRecord takes a CRC checksum for each record).

Later CLs will update the dispatcher to write to the journal during state changes and read from the journal on restart.

PiperOrigin-RevId: 323398037
Change-Id: I62eaab40219992c26df5029075a5875f212f92da