Resolve to handle of the primary variable when saving a loaded function
If a model is loaded under strategy, we wrap the loaded function so that when used in strategy.run(), the captured variables can resolve the ones in the current replica. We placed the guard that such loaded functions cannot be called outside of replica context, because the function may mutate the variables. However this breaks saving which happens in the cross replica context.

This change checks if we're saving a non distributed version of the model, and if so, resolve the primary variable.

PiperOrigin-RevId: 323926989
Change-Id: I5c24395b82e9601214e969f5889655b93839ba20