[tf.data service] Write dispatcher state changes to a journal.
When a work directory is configured, the dispatcher will write journal entries for its state changes to a journal within the work directory. If no work directory is configured, the dispatcher uses a NoopJournalWriter, which writes nothing.

We don't yet read from the journal on dispatcher start. This support will be added in the next CL.

PiperOrigin-RevId: 324749061
Change-Id: I0baf10cef05a53e0fa94139d84d5cfd284550acf