[XLA] Reserve for all pinned buffers first them before starting allocations.
We were previously reserving for pinned buffers in alternate memory as we were
allocating. This meant that we could discover a pinned buffer after we have
other buffers, possibly going beyond the size of the alternate memory. Now we
split the logic to reserve buffers from the rest of allocations so that doesn't
happen.

PiperOrigin-RevId: 325249612
Change-Id: I9fa36b832d86701cdbcb61a808fad763cf664634