[XLA] Implement mechanism to repack allocations to reduce fragmentation.
This CL defines an interface for allocation repackers. A repacker can be
specified in MemorySpaceAssignment::Options. If an HloValue couldn't be
allocated due to running out of alternate memory, we now export the allocations
done so far to the repacker, run the repacker, and import the new offsets back
into memory space assignment for better memory packing.

PiperOrigin-RevId: 326040207
Change-Id: Icc518874781eb74e38701514d8f6fb20d23a4124