Extend TPUClusterFormation pass to use resource alias analysis when deciding which interleaved ops to moved after the cluster.
Ops not in the cluster may be interleaved with ops in a cluster due to ops being in a topological sort order. This ordering though takes into account control dependencies prior. When working with resources, as ordering of such ops are not done via data dependencies, resource ops may be moved before a cluster, when in fact they should be after. Resource alias analysis is used then to decide whether a resource has been used in a cluster before deciding to move an op to after the cluster.

PiperOrigin-RevId: 326263053
Change-Id: I625bdfdb9208703cd8918cd76ed01b1a151bbdba