Add a check alive thread in MWMS behind a flag
This is now disabled by default but can let us test the idea. The thread keeps checking the cluster and aborts the collective if any worker is not reachable.

Currently the only way to recovery is to restart, since once collectives are aborted, all subsequent collectives fail immediately.

The thread uses a RING all-reduce as a check mechanism, since we don't have a check alive op yet.

PiperOrigin-RevId: 327493026
Change-Id: I9ab9e5be1f5c1a15b3741a9f26e42f25b9a59a12