[TF:TRT] Initialize TensorRT plugin registry before deserializing cuda engines.
When a TF-TRT converted graph with static cuda engines is executed, we call the
TensorRT runtime to deserialize cuda engines without initializing the TensorRT
plugin registry. This causes TensorRT runtime failure when the cuda engines
contain plugins.

Move InitializeTrtPlugins to common/utils.cc and replace the use of mutex with
absl::call_once.
PiperOrigin-RevId: 327679056
Change-Id: I5e50a01aa06f3b5a22a3114a2c54e3712461bd6b