Fix Windows GPU build failure in resize_blinear_op.cc.
I broke it in 67d15573a776119d5a544ed266dc2514ae13c3b5.

Before, I called a function under a `std::is_same<Device, CPUDevice>` condition which cannot be linked if Device is a GPUDevice. I would expect the function not to be generated if Device is a GPUDevice due to dead code elimination, but apparently it still is on Windows.

PiperOrigin-RevId: 327887418
Change-Id: Ib97e1abf1680c75dc072850cc69c761e10ac3e1e