Add Python binding for CheckPeerHealth
This will be used to check health of peers for MultiWorkerMirroredStrategy. I chose to add a python binding instead of an Op because I don't it to be exposed in tf.raw_ops, nor be subject to GraphDef compatibility requirement.

This difference between this and the existing ContextCheckAlive is that this is specific to multi client training. We need to tell if a worker has failed and restarted. EagerContext uses context_id for this purpose, but in MWMS we don't create EagerContext on peers. Collectives tracks device incarnations of peers, so we use that instead.

For now the check health will be in an all-to-all style, we can use more complicated topology in the future if needed.

PiperOrigin-RevId: 328843814
Change-Id: Id1281074cf0a79d1339bb4fd73caa6f6e68f518f