TFLu: Fix Ethos-U build issue
The issue is that a binary need to be built twice, since it
depends on recursive_find. The driver is not fully downloaded when
recursive_find is called. The solution is to call the download script
immediately via the make's shell function.

Also add missing exit to error case in download script.