[XLA] Disambiguate allocations that need to be colocated for repacking.
Previously, we were assuming allocations for a given HloBuffer that had the same
offset needed to be colocated. This is overly restrictive. Sometimes, buffers
can get the same offset even though they don't need to. This will increase the
freedom that the repacker has in terms of allocating the buffers.

To disambiguate which buffers need to be colocated and which don't, we introduce
a new struct called AliasedOffset. Instead of relying on the actual int64 value
of the offset for specifying preferred_offset, we use this object and associate
colocations with each other.

PiperOrigin-RevId: 329941439
Change-Id: Ia402c8afd950c9c6d4e4a340643627b0071c4f2a