Don't prune the __nv_relfatbin section
Pruning seems to strip PTX for NCCL as well, which defeats the purpose of
shipping PTX with TF at all.

Note: it is expect that this will increase the TF pip size by ~ 33MB.  There may
be a way to claw this back but fixing the bug takes higher priority.
PiperOrigin-RevId: 330657787
Change-Id: I3fac11e504326e08428ae2ac4dd899783519c2bc