Rolling forward disabling failed test since it wasnt using API correctly
When adding a function to the EagerContext, we check the cache whether we have added that function before by comparing the name. If its there, we just take another ref on it without checking whether the bodies are actually the same or not. This CL fixes that by ensuring that the function defs are the same. If they're not, it reports an error.

PiperOrigin-RevId: 331728315
Change-Id: Ib289996a28d553e2b2b9d30730288692a9a69640