Fix a memory leak when creating UninitializedVariables (as SavedModel loading code does)
This is the same shared_name kernel caching issue we fixed for regular tf.Variable.

Adds a test using the C++ memory checker; the Python checker complains about a weakref being allocated each time, so I'm just using the C++ checker for now.

Does not test on GPUs, since that appears to still leak memory (could be an infra issue).

PiperOrigin-RevId: 331798116
Change-Id: I59f7912621b7d20b6fcf19f87d3f27e60d56fec3