Remove autocast warning.
This warning was issued when inputs to layers will autocasted and the mixed precision API was not used. The intent was to warn users familiar with TensorFlow 1 that passing a float64 tensor to a layer will run the layer in float32 by default in TensorFlow 2.

The warning was confusing to users who never intended on running the layer in float64 in the first place. Since TF 2 has been out for a while now and the warning is confusing, it is removed.

PiperOrigin-RevId: 332135086
Change-Id: Idf264b1f7a90a9c05e4bc8a9a14cea27a1ec4a79