Mixed precision Embedding layer now outputs float16 tensors.
Before, the Embedding layer did not use the global dtype policy by default, but instead set its dtype policy to floatx. The net result is with mixed precision, it would output float32 tensors instead of float16 tensors. Typically this wasn't an issue as the next layer would cast to float16, but this caused issue when the Embedding output was passed to an op. This change fixes the layer so it uses the global dtype policy, making it output float16 tensors.

To avoid a performance loss from having to cast the entire embedding variable, the layer casts the output of the embedding instead of the variable itself.

Fixes https://github.com/tensorflow/tensorflow/issues/42631

PiperOrigin-RevId: 332317478
Change-Id: Ia71f4a0fc252ac6a158322ef2cabe1e404266cd5