Remove _auto_cast_variable_read_dtype from Graph.
This change removes a dependency from Keras to a private API of TensorFlow. In particular, this removes the _auto_cast_variable_read_dtype property of the Graph and adds a thread-local variable in Keras that plays the same role.

This also causes a behavior change. Now, calling `MirroredStrategy.run` or `ReplicaContext.merge_call` within `Layer.call` will reset the value of the autocast dtype to None because these methods use threads. This means layers which call `Strategy.run` may not work correctly with mixed precision unless they start manually casting variables. In practice, I do not think this is an issue, as I don't think any layer calls `Strategy.run` or `ReplicaContext.merge_call`.

Another behavior change is that switching a graph in `Layer.run` will no longer reset the value of autocast dtype to None.

PiperOrigin-RevId: 332508065
Change-Id: I3fb081a4e5f8e8ba275c20c304980c14ad522032