Make softmax and sigmoid activation backtracking more robust. Logits are now cached on the post-activation Tensors and used whenever available. This allows logits to be used in eager mode as well as in the middle of a tf.function, whereas previously they were only able to be used in the middle of a tf.function.
The only time this mechanism will not work (and the fallback, existing mechanism will be used) is when the activation and the loss function are separated by a tf.function boundary.

Also fixes Loss class to only autograph the loss function when in eager mode. This allows Python control flow to work correctly in the loss function.

PiperOrigin-RevId: 333367543
Change-Id: I64454b9acb6a77247122d72e1302aa5110415d57