[lite] Update the label_image example to use StatefulNnApiDelegate.
This changes the behaviour when the '--accelerated' flag is enabled to
use acceleration via StatefulNnApiDelegate, which is not deprecated,
rather than via NnApiDelegate, which is deprecated.
It also sets the allow_fp16 field in the StatementNnApiDelegate::options
to the appropriate value based on the setting of the --allow_fp16 flag.

PiperOrigin-RevId: 333695351
Change-Id: I3917ff8f79f73323635bb79f4af22df96f4ec655