Split env library so that there is only one instance of `Env::Default()`
While working on modular file systems, noticed that tensorflow's pip package
includes two copies of `Env::Default()` in two shared objects libraries.
One is in `libtensorflow_framework.so`, another is in `_pywrap_tensorflow_internal.so`.

Because of that, `Env::Default()` could return different instance depending on when
a modular file system is called. The Lookup vs. Register discrepancy may confuse the file
system to believe the scheme (e.g., `s3`/`gcs`) is not registered (actually registered
by another `Env::Default()`).

The duplication of `Env::Default()` is caused by the fact that `tensorflow/core/platform:env`
dependency is included in both `libtensorflow_framework.so` and `_pywrap_tensorflow_internal.so`.
in bazel indirectly.

There are two many dpendency graph component in between `tensorflow/core/platform:env`
and `libtensorflow_framework.so`/`_pywrap_tensorflow_internal.so` in bazel.
As a result this PR tries to address the issue by split out the `Env::Default()` implementation
out and only link it with `libtensorflow_framework.so` to guarantee `libtensorflow_framework.so`
is the only place to include `Env::Default()`.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>