Adding bsdmainutils to the install list for ROCm docker containers.
For the ROCm TensorFlow build, the "hexdump" utility is required for the tool flow used by MLIR generated GPU kernels. The "hexdump" utility is part of teh "bsdmainutils" package and that package is currently not installed in the ROCm docker container. This change simply adds the "bsdmainutils" package to the install list