Remove `CHECK` in `TensorShapeBase::end()`.
Since we use `kUnknownRank` to mark partial shapes with no rank information, without this `CHECK` or this code change the following program would result in accesses out ouf bounds

```cc
PartialTensorShape s;
for (auto d : s) {
  std::cout << d << "\n";
}
```

With the `CHECK`, the above program results in an abort. Without the `CHECK` and without the rest of the changes, `TensorShapeBase::dim_size` will be called with arguments from 0 to 255 (`kUnknownRank`) which would result in reading outside of bounds of the `ShapeRep` buffer (`dim_size` has `DCHECK`s). With the fixes in this change, the `for` loop will not iterate at all, but there won't be a program crash.

PiperOrigin-RevId: 334466121
Change-Id: I96252933b67cb64a27c1648157f87ccff3a5561a