Disallow MEAN non-floating distributed variables.
MEAN aggregation always produces a floating number. We ran into issues when assigning to a MEAN ON_WRITE variable which was caused by the dtype mismatch. The conclusion at that time was that we're going error instead of casting the number to an integer to avoid potentially surprising precision lost.

I recently found that SyncOnReadVariable.read_value() has a similar issue. If aggregation=MEAN, it returns a floating number, instead of a value of the dtype of the variable. Based on the same rational, this changes disables MEAN aggregation for non-floating variables completely.

PiperOrigin-RevId: 334682155
Change-Id: Ib0f96c2a90f9e5f0b4bb4e255f2622e3dd4670bd