[ROCm] Fix for ROCM CSB breakage - 201007
The following commit introduces a new subtest that is failing on the ROCm platform

https://github.com/tensorflow/tensorflow/commit/f2ebefba658fb4f424842f03d15f694cb917110f

The new subtest ( `NcclManagerTest.Abort` ) requires simulating a multi-node environment on a single-node with mutliple GPUs. This functionality is currently not available on the ROCm platform, and other subtests that require this functionality have already been disabled on the ROCm platform. Doing the same for the newly added subtests as well.