[TF2XLA] Check 'case' op compilation from autoclustering codepath, but not from tf.function
This is a bit complicated: we do allow uncompilable code under
tf.function(compile=True) in 'case' statements (and, in the future, under
'if/while' statements), provided the input is constant-foldable, and the
uncompiled branch is guaranteed to be never executed.

We can not allow the same during autoclustering though, as we don't know the
value in advance, and we have to be more conservative.

PiperOrigin-RevId: 335928032
Change-Id: I84a4683079d1871672f429029e279522c258d66c