Add the Python API and AST translator for authoring TF op composition
The new API allows the user to define a new op with the composition of existing
TF ops. For example, the following code defined a new op called 'NewConv2D'
with three tensor inputs and one output:

```
@Composite(
    name = 'NewConv2D',
    inputs=['input_: T', 'filter_: T', 'bias: T'],
    attrs=[
        'stride_w: int', 'stride_h: int', 'dilation_w: int', 'dilation_h: int',
        'padding: {"SAME", "VALID"}', 'act: {"", "RELU", "RELU6", "TANH"} = ""'
    ],
    derived_attrs=['T: {float, int8}'],
    outputs=['o: T'])
def _composite_conv_add_relu(input_, filter_, bias, stride_w, stride_h,
                             dilation_w, dilation_h, padding, act):
  res = tf.raw_ops.Conv2D(
      input=input_,
      filter=filter_,
      strides=[1, stride_w, stride_h, 1],
      dilations=[1, dilation_w, dilation_h, 1],
      padding=padding)
  res = tf.raw_ops.Add(x=res, y=bias)
  if act == 'RELU':
    return tf.raw_ops.Relu(features=res)
  elif act == 'RELU6':
    return tf.raw_ops.Relu6(features=res)
  elif act == 'TANH':
    return tf.raw_ops.Tanh(x=res)
  else:
    return res
```

The translator can be invoked to translate this function directly to an MLIR
module, which can be used to (optionally) expand the 'NewConv2D' in the graph
to a sequence of leaf nodes with kernels and bridge support. Optimization
(constant folding, loop unrolling and branch condition evaluation, etc., which
are chcked in seperately) can be applied to simplify the result graph.

PiperOrigin-RevId: 335976996
Change-Id: Ie4d641c0e86b234cd94d0a6d36e26269659676f9