Change default value of loss_scale in Policy.__init__.
This has no functional effect except the string 'USE_DEFAULT' can longer be passed to the Policy constructor, but I don't think anyone does that. The default value is changed from 'USE_DEFAULT' to 'auto', but the default value as the same effect as before.

This change is described in the RFC: https://github.com/tensorflow/community/pull/293. If an alternative approach ends up being used in the RFC, I will revert this change and implement the other approach.

PiperOrigin-RevId: 336012719
Change-Id: I339a29305276bf1e52af555df9e84090a96db6b8