[XLA] Add back required assignment for conditional outputs after every repack.
We currently require conditional outputs to be placed in the default memory. And
before we start allocating a colocated interval, we add a required assignment
for conditional outputs to be placed in the default memory. However, a repack
erases all of the required assignments for the colocated interval, since some of
the required assignments are inserted as a mechanism to propagate the allocation
decisions to the called computations. We never actually added back the required
assignment for conditional outputs, causing these to sometimes get allocations
in the alternate memory.

PiperOrigin-RevId: 337608583
Change-Id: I28225f49ac7d78f6c08efb6dd45903c6e50c68eb