Support ops with regions in tensor-array-ops-decomposition pass.
Since regions support implicit capture, there is no need to explicitly capture resource operands. It is expected that explicitly captured resource arguments have been canonicalized away for region based ops. Now, we only need to run the pass on the regions associated with such ops.
This change is primarily meant to support WhileRegion, IfRegion and CaseRegion ops in this pass.

PiperOrigin-RevId: 337923974
Change-Id: Ic4b6e0b66df7e548df69647eef6a91ed1c07ad44