Rely on cancellation in collective V2 kernels
For collective v2 kernels we stop aborting collective ops if they're cancelled.
Most componenets except param resolution is able to respond to cancellation.
Param resolution is not large concern in practice, since group resolution is
likely not needed, and most instance resolution do not block.

Technically we can do this to v1 kernels as well, but it doesn't seem safe since
we reuse instance keys in v1 collectives.

PiperOrigin-RevId: 338508607
Change-Id: Iab2f4e1061d7b384b83bc2712b849e42ba3677fc