Addressing large number of outputs in a consistent way across our various C APIs (Eager fastpath, Eager slowpath, graph).
Previously in graph mode, we would error out for somewhat the wrong reason but
the error message was decent. This CL makes it more precise.

In fastpath eager mode, we tried parsing an int64 attr to an int32, failed and
unnecesarily hit the slow path.
In the slowpath, we used to just try and allocate a huge output number of
handles causing a crash.

This CL just checks whether the output size is greater than int32 and errors
out before trying to do any allocation.

Fixes https://github.com/tensorflow/tensorflow/issues/42281

PiperOrigin-RevId: 339150482
Change-Id: I6d9d1541ccfa9881f92a06348567562f79b0a963