[tf.data service] Fix bug with infinite repeat not working with distributed_epoch.
To perform repeat when using a split provider (currently only done when using distributed_epoch  processing mode), we need to reset the split provider in addition to making a new iterator. Before this CL we did this correctly for FiniteRepeat dataset, but not for ForeverRepeat dataset. This CL fixes the issue for ForeverRepeat dataset.

Without the change, the added test fails with a StopIteration exception.

PiperOrigin-RevId: 339184323
Change-Id: Ib669ca7ad4c6a1ce87599288587a769256de1a14