Use worklist for function type updates
Switch to using a global worklist for functions to enable updating callers of functions. This enables propagating result types and so avoids the restriction on refining return types. Functions are added back into the worklist if they reference a FuncOp that was refined. Move the helper functions (and their callers) to the ShapeInference class to enable easier enqueuing/querying of these. Also cache the function call mapping during processing.

PiperOrigin-RevId: 340451923
Change-Id: I7ee4f9d90f8eb259768a1678f292690d602085a9