Update metrics to better support tensors that are not fully defined
If a tensor is not fully defined, broadcast_weights calls
DenseToDenseSetOperation which isn't supported by TPUs. In that
case, metrics should fallback to a less performant but supported
path.

Also disable the mlir bridge presubmit testing because the new
tests fails on the mlir bridge. It isn't possible to use
@test_util.disable_mlir_bridge() because it conflicts with
@ds_combinations.generate.

PiperOrigin-RevId: 340676883
Change-Id: I965fee2185605542a177599ccf47238f10d15da2