Fix xla_compilation_cache's op-by-op-mode enable mlir bridge logic
The xla_compilation_cache was incorrectly checking to see if the
MLIR bridge was enabled in op-by-op mode. It only looked at the
ConfigProto's enable_mlir_bridge flag. It did not check the other
flags that can be used to enable the bridge so it missed several
cases where the MLIR bridge was supposed to be enabled.

Also, disable tests that are now failing because they are actually
running the MLIR bridge.

PiperOrigin-RevId: 341054280
Change-Id: Iceddd1edb5e11e8940d1acd0a69cd31959533d43