[TFRT:Compiler] Add the cluster-tf-ops-by-host pass
This pass clusters the TensorFlow ops by host so that ops placed on different hosts
will be in different functions. Each result of the per-host function will have a "tf.device" attribute which specifies the device assignment of the result.

The pass currently assumes that there is no circular dependency among the per-host functions. For example, if there exists an operation in functionA that consumes the result of an operation in functionB, then there does not exist any operation in functionB that conumes the result of any operation in functionA.

PiperOrigin-RevId: 341145682
Change-Id: I2ab1468fb3acc374e3272bd8aed1c2ed016af6e7