Avoid risk of including conflicting implementations of monitoring classes.
The classes in
  tensorflow/core/lib/monitoring/{counter,gauge,percentile_sampler,sampler}.h
have two implementations.  This is achieved by using a #if directive to provide
the implementation directly in each header files, or alternately, a no-op
implementation on mobile platforms from one of the header files
mobile_{counter,gauge,percentile_sampler,sampler}.h.

I believe the intent is that the mobile*.h be used only via the main header
files, and never be included directly, but nothing was preventing this.  If
someone had included one of the mobile*.h files directly, and the resulting
object file were linked with one that used the primary header files on a
non-mobile platform, it may cause problems.  There would have been no error at
compile or link time, yet the classes would be defined in two different ways,
leading to an unchecked ODR violation and undefined results.  For example, the
linker potentially could pick an arbitrary version of each routine in the
class.

This change tries to avoid the potential problem in two ways:
- by restricting the visibility of the mobile_*.h variants (for bazel builds,
  at least); and
- by causing the mobile_*.h files to use #error if they appear not to have been
  used on a mobile platform, or not included from their respective primary
  header files.

Also, include {mobile_,}percentile_sampler.h in the mobile_srcs_only_runtime
build rule.

PiperOrigin-RevId: 341680724
Change-Id: I1de71dd209f2769e162c0c5522cf0cf5006ef5ff