Remove decompose types patterns.
These were only needed because ReturnOp is marked illegal if it has an illegal operand. But in a full conversion that isn't needed, because all values (results/bbargs) are replaced, so all operands are automatically legal at the end.

The previous code was effectively requiring a pattern (even a dummy pattern) for the ReturnOp, when it wasn't even needed.

PiperOrigin-RevId: 343032055
Change-Id: I2067016a2c94519df58325151cebcf18ac0ab059