[TF:TRT] Remove duplicated implementations for converting between TF data types and TRT data types.
Remove the implementations from convert_nodes.cc and convert_nodes_test.cc and use the implementations in utils.cc instead.  In doing this, we also fix a problem in convert_nodes_test.cc, which is related to the use of EXPECT_EQ in the two routines and causes the test to continue running in release build even when the data type has a problem.

PiperOrigin-RevId: 343563121
Change-Id: Ia8dabeafe4f0d75d7da7d74f628678cbae951bac