Update logic for compute_output_shape in keras.layers.Bidirectional.
The existing logic was not able to handle the list/tuple correctly, as well as the rnn layer with multiple states. It wasn't a big problem before since the output shape was calculated by the call() method, and compute_output_shape() is only used in a very cornered case (eg. dynamic model).

Also fix https://github.com/tensorflow/tensorflow/issues/43816.

PiperOrigin-RevId: 344840193
Change-Id: I5a59389be57212e453804f9ad10c85cac58dc8c9