Remove "+code-object-v3" from the target feature_string used during XLA codegen.
Starting with ROCm 4.0, "+code-object-v3" is no longer recognized as a valid feature-string from ADGPU targets, and result in the following warning for each HSACO codegen

```
'+code-object-v3' is not a recognized feature for this target (ignoring feature)
```

Since the hipclang compiler generates code-object-v3 objects by default now, we no longer need to the explicitly specify "+code-object-v3" as a feature-string. Removing taht feature string also gets rid of the warning.