[tflite] map leaky relu to prelu to use NNAPI
Leaky ReLU is one of the widely used activation functions.
However it's not supported by NNAPI, so it's unlikely to
accelerate a model with many Leaky ReLU ops. Fortunately,
NNAPI 1.2 supports PReLU, which is a general case of Leaky
ReLU. So we can map Leaky ReLU to PReLU. With that, we
can fully delegate models, such as the TFLite super
resolution example, to NNAPI.