Trace generic cudaMemcpy / cudaMemcpyAsync
These copies are "generic" as in, I don't know a priori what memory spaces the
src and dest are in.

Previously we traced them as "UNSUPPORTED" ops, and didn't fill in the op's
name.  So consumers of the trace can't tell that it's a copy.  Now we trace it
as a MemcpyOther, which seems close enough.

Also tweak some logging inside cupti_tracer itself.

PiperOrigin-RevId: 346106575
Change-Id: I0d87aea73d6ca6fa7b0ce0ec54f56c0b5d286e54