Add option to enable the MLIR bridge if safe
The MLIR bridge currently can run many models without any issues.
There are some features that it does not support or only partially
supports.

Currently, the public interface supports forcibly enabling
the MLIR bridge, forcibly disabling the MLIR bridge or unspecified
which may or may not enable the MLIR bridge depending on where we
are in the launch process. At the moment unspecified does not enable
the bridge.

Some customers have asked for a fourth option: enable the MLIR
bridge if an analysis determines it is safe to run the MLIR bridge.
They would like to opt in to the MLIR bridge early instead of waiting
for the main launch. The MLIR bridge can run some of their models but
not all. Since they support many models that all use the same
framework to launch their jobs they can't easily selectively enable
the MLIR bridge for just the models that the MLIR bridge can currently
support.

PiperOrigin-RevId: 346854739
Change-Id: Ia7ad2117712c7ca1c2fea3c6e7be9d172c482ae4