Add a pass to drop `shape_invariant` attribute from TF While/WhileRegion ops.
`shape_invariant` attribute relaxes the constraint on While/WhileRegion ops to have separate operand and result shapes. This is to support changing shapes in each iteration of loop body. For cases when this relaxation is not required (for example when compiling to XLA) this pass can be used to drop the attribute.

PiperOrigin-RevId: 346856163
Change-Id: I91477151be4c28e77d39ff62e715e0822526c3b7