Add a custom device tensor handle type
This will replace TensorHandles holding custom device tensors, which will clean up the runtime a bit, provide flexibility for custom devices (especially for shape handling and copies), and morph custom devices a bit more toward op handlers.

I'll do the migration itself in a followup.

For now I'm leaving the naming as "custom device" for consistency. Once custom devices are migrated to this type it will be easier to bump the abstraction level a bit (AbstractTensorHandles instead of ImmediateExecutionTensorHandles) and re-name things.

PiperOrigin-RevId: 346871040
Change-Id: Ic1d4791b27d37b1078bd985be344c54748ae49b8