[XLA:GPU] Add --xla_gpu_deterministic_ops to force run-to-run determinism.
Fail if there is no deterministic lowering for HLO ops in the module. The env variable `TF_DETERMINISTIC_OPS` is related, but serves a somewhat different use case.
First, it is best-effort in the sense that deterministic lowering will be preferred if the exist. However, Scatter with non-unique indexes and SelectAndScatter do not have a deterministic implementation at present, which is silently ignored. In contrast, --xla_gpu_deterministic_ops will cleanly fail compilation for these ops.
Second, the `TF_DETERMINISTIC_OPS` env variable is not convenient for compiler frontends other than TensorFlow.

PiperOrigin-RevId: 347672477
Change-Id: I62c3093784560394b0aa173747afd350a299ed64