Speed up tf-tpu-outside-compilation-cluster pass by not processing dependent ops multiple times
Currently, this uses stack to maintain the op worklist using SetVector which will cause some ops to be visited multiple times as the ops are not necessarily popped in the topological order. This will have exponential complexity.

Visited ops are now tracked exclusively to make it linear time. Added a TODO to use GraphCycle data structure that should help further improve the performance.

PiperOrigin-RevId: 347715110
Change-Id: Iba936427f2a7b8e3552dc8ae72e24b947201d496