Add a test env object to allow passing additional information to worker processes
Sometimes we need to prepare the test environment in the main process and pass the information to worker processes. For example to test tf.data service, we need to start the tf.data service cluster in the main process and pass dispatcher address to the worker processes.

We hardcode the fields of this object to give us a better idea of its usage. There shouldn't be too many of them.

PiperOrigin-RevId: 348852814
Change-Id: Idef7e2659b11623abb46f06139e385047b73cb2f