Add quantization spec for LSTM, so bias can be handled properly.
This logic wraps GetOpQuantSpec and adds special handling for LSTM specs. Instead of adding new traits to tfl_ops.td, this approach was picked due to following features not supported in `AccumulatorUniformScale` trait, and tricky to implement cleanly.
1. LSTM bias scale depends on variant - it can either depend on inputs, or normalization value.
2. Intermediate tensor scale is not supported

Will reconsider adding this to trait when need for similar use cases arise in other ops.

Minor changes:
* Changed name of power_of_two_bound to SnakeCase.
* Factored out commonly used codes
* Fixed dequantize op's return type to float, instead of quantized type.
* Skip input processing when input if from ops with same scale requirement.

PiperOrigin-RevId: 349524402
Change-Id: I392998215abbfd394368d289543cad774baeb057