[ROCm] Fix for breakage in XLA Conv Op functionality
The following commit breaks Conv Op functionality (in the XLA backend) for ROCm platform.

https://github.com/tensorflow/tensorflow/commit/8684c6b2e95601542c6c5c006bde5dd50f589a50

The cause seems to be that the `scratch_size` field in the new `GpuConvDescriptor` is not getting correctly populated in the new MLIR path. It is being used correctly in the convolution runner code.

declaration:

https://github.com/tensorflow/tensorflow/commit/8684c6b2e95601542c6c5c006bde5dd50f589a50#diff-6453912dbc4ee715a56da9d7b218b52795dea2aa631a482101fc6d58c573d9ccR122-R135

use (get access) in conv runner:

https://github.com/tensorflow/tensorflow/commit/8684c6b2e95601542c6c5c006bde5dd50f589a50#diff-a01181d08b28a9c7432f22439622f16725126184283a73822c70b2151098a8adR277

set access in non-MLIR(?) based path:

https://github.com/tensorflow/tensorflow/commit/8684c6b2e95601542c6c5c006bde5dd50f589a50#diff-a01181d08b28a9c7432f22439622f16725126184283a73822c70b2151098a8adR450

This commit merely adds the missing "set" in the MLIR based path