PSv2: Shutdown the worker preemption thread in ClusterCoordinator's destructor.
When ClusterCoordinator's destructor is called, mark the worker preemption thread as finished to stop the thread.

Notices the Cluster and Worker objects will not be destroyed due to the threading.Thread object holds a reference to the cluster object.

There will be 2 following CLs to address the Cluster/Worker deletion:
1. Close the worker threads in ClusterCoordinator's destructor.
2. Use weakref as target function for threading.Thread objects, to ensure Cluster/Worker objects can be garbage collected in memory.

PiperOrigin-RevId: 350639066
Change-Id: I23b54912f9b450fb2ccbe6b1da69d2bf5bc55aa5