[ROCm] Fix for breakage in ROCm support - 210112
The following commit introduces a new unit-test `//tensorflow/compiler/xla/service/gpu/tests:mlir_sorting_test`, which fails on the ROCm platform.

https://github.com/tensorflow/tensorflow/commit/70502be4a5941b6ffa2b8c33bb549657b33976da

The test fails on the ROCm platform because the underlying code for it is CUDA-centric. This PR/commit updates the test to make it work on the ROCm platform as well.