[XLA:SPMD] Adding support for detecting potential parallel dimensions in gather.
Some gathers have indices driven by IOTAs that cause the access to be
partitionable in the dimension indexed by the dimension with of the IOTA.
An example is:

  %iota.1 = iota()
  %indices = concatenate(..., %iota.1, ...)
  ... = gather(..., %indices)

which is a pattern followed by tf.reverse_sequence for example.

This patch adds support to detecting such situation and allowing sharding
of the gather across devices.

PiperOrigin-RevId: 351602320
Change-Id: If6be7a51ee1bbb31cadb6eef9d7b01975137e12e