[tf.data service] Support num_consumers being a Tensor.
The `if num_consumers >= 0:` check causes an error when num_consumers is a Tensor in graph mode:

OperatorNotAllowedInGraphError: using a `tf.Tensor` as a Python `bool` is not allowed in Graph execution

Comparing to `None` accomplishes the same thing without this issue. Once the forward compatibility window expires we can remove the check completely.

PiperOrigin-RevId: 351894589
Change-Id: I0f6e8b93f6ae14b92b08c7c438d6a111ae0efff7