Rewrite the mhlo->TF legalizer for mhlo.gather
The new legalizer targets TF.GatherNd instead of TF.GatherV2 and as a
result can support a wider selection of ops then the precious version.

If a consumer prefers TF.GatherV2 over TF.GatherNd then we should add a
canonicalizer from TF.GatherNd to TF.GatherV2 to handle that usecase
instead of dealing with it during legalization.

PiperOrigin-RevId: 352780338
Change-Id: Ib65342b002fce2bf306d1fcbd54d89e0a775ccc2