Handle ReplicateOp op with packed inputs in following passes
* TPUSpaceToDepthPass
* TPUDynamicLayoutPass

Packed inputs were added to ReplicateOp after this pass was written so it only handled replicated inputs.

Also, added helper method GetOperandsForBlockArgument to ReplicateOp and using it for other helper methods.

Updated TPUSpaceToDepathPass existing test by adding devices attribute to trigger this codepath in the test.

For TPUSpaceToDepathPass test, I had to copy the replicate the existing complex test case instead of having a minimal test. I tried some things but the pass seems to be having many assumptions and will require closer look to trim the simplify it.

PiperOrigin-RevId: 353307553
Change-Id: I8c1ed8ca396da9bb28ffa4e479fb81813b01b17d