Fix crash when invalid keepdims value is being passed to tf.math.reduce_prod
This PR tries to address the issue raised in 46700 where
tf.math.reduce_prod will crash if keepdims is being passed
with a non-boolean value (e.g. numpy value)

The issue was that keepdims is passed through pywrap
which can not interprete numpy values, thus crashes.

A way to detect the type mismatch before being passed
to pywrap is to use `bool(keepdims)` to give python a chance
to convert to bool (and throw out error when appropriate).

This PR also fixes all reduce_ ops.

This PR fixes 46700.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>