[XLA] Fix implementation of igamma and igammac for f16 and bf16.
The implementation was supposed to work for f16 and bf16 by upcasting
to f32. The problem was that when computing constants list `Epsilon` and
`MaxFiniteValue` for the precision of the original type (e.g., `f16`) they
are used in computation along with upcast operands, leading to error:

```
Seen floating point types of different precisions in %compare.198 = pred[1,1,3]{2,1,0} compare(f32[1,1,3]{2,1,0} %divide.195, f16[1,1,3]{2,1,0} %broadcast.197), direction=GT, but mixed precision is disallowed.
```

The fix is to use Epsilon for the new type of the computation (F32). This produces values that are more accurate w.r.t. scipy than trying to use Epsilon at F16 and then upcasting it to the type of the surrounding computation.

PiperOrigin-RevId: 354271949
Change-Id: I009929e0072c901a75e6091f823745dfe082554f