Move handling of variant input/output to outside compilation operation to MarkOpsForOutsideCompilationPass.
Variant types can not be communicated to/from TPU device.  Instead we should pull the Op providing the variant input or the op consuming the variant output of the cluster.  This is not done if the variant source/sink ops are inside of an already outside compiled region(for example control flow).

PiperOrigin-RevId: 354326932
Change-Id: I44002cf14c192571492a7b2a03aa05c8cb40ac67