Fix vocab indices when calling set_vocabularly with special tokens
This applies to TextVectorization, IntegerLookup, and StringLookup. Vocab indices would
be off by the number of special tokens when you set a vocab with explicit special tokens, e.g.
layer.set_vocabulary("", "[UNK]", "foo")

This meant you would get incorrect results when trying to set_vocabulary with the results of
get_vocabulary for any of the affected layers.

PiperOrigin-RevId: 354561683
Change-Id: I19858328e33717a30181f85a56510b8ebe1060db