Fix clang-format installation on the TFLM Docker container.
The TFLM CI is currently failing with the following error message:
```
The following packages have unmet dependencies:
 clang : Depends: clang-12 (>= 12~) but it is not installable
 clang-format : Depends: clang-format-12 (>= 12~) but it is not installable
E: Unable to correct problems, you have held broken packages.
The command '/bin/sh -c apt-get install -y clang clang-format' returned a non-zero code: 100
```

This changes ensures that we download a specific version of clang. We can still
upgrade if there is a new release, but that should be deliberate instead of
breaking the CI.

PiperOrigin-RevId: 354953821
Change-Id: I693ffef2975a116ffac8f27c409cd5ca7f527790