[PJRT] Use ExternalReference as the return type of PjRtBuffer::ReleaseDeviceMemoryOwnership.
Rename PjRtBuffer::ExternalReferenceHold to PjRtBuffer::ExternalReference. Move implementations into their corresponding .cc files. There is no reason for them to be in headers at the moment.

Slightly devirtualize the class: we must always support a single pointer, so that may as well go in the base class. It's possible that in the future we might want to extend the reference class to multiple pointers; that change could also go in the base class.

PiperOrigin-RevId: 355074230
Change-Id: I745daf39f660fdb0749e729ddb0ca8c77b11dfeb