Remove `--force-reinstall` from tb and tfe. Update tfe to latest.
Having `--force-reinstall` brings in `numpy-1.20` which has a backwards incompatible change. This is a stop gap to fix broken builds but we should convert to installing from a requirements file instead of expliclty listing dependencies (in 6 different places!!!!!)

Also, update estimator to the latest release. It seems we forgot this step after last release.

PiperOrigin-RevId: 355173614
Change-Id: Ia5b235d3df34631a80dcae9b4a101fdd17575295