Disable Eigen docs builds by default.
Eigen documentation builds compile tools with the host toolchain which
it then tries to execute. When cross compiling doc tool execution
fails as the tools are built with the target toolchain. This change
disables Eigen docs builds by default to facilitate cross compilation.
Unfortunately, CMAKE_CROSSCOMPILING can't be used to enable docs
builds when cross compilation is disabled as it does not work on Apple
platforms.

PiperOrigin-RevId: 355646699
Change-Id: Ic9bf789037a0abf9c40676f5f12998466d4978bc