Change API contract for the private add_function_callback and allow graph edits before finalizing a func_graph.
Currently, function_callbacks were invoked after the graphs were finalized, primarily used to prevent dereferencing. While the API seemingly allow graph edits, they are futile as the graph is already committed to the C++.

With this change, we change the semantics of the API a little bit and explicitly allows graph edits.

PiperOrigin-RevId: 355877980
Change-Id: If43e37d2fb1032085048b7810aa4e387714e2f88