[tf.data service] Fix replay issue in round-robin reads.
Previously we indexed requests by (1) round index and (2) pointer addresses. This is a problem when we get multiple requests for the same round, e.g. due to retries. This could cause the same consumer to be counted multiple times, potentially producing a round of data before all consumers are ready.

PiperOrigin-RevId: 356858844
Change-Id: I2511fabb769bb5211b4d7d9c4185df4f3ce0e4cc