Fix the bazel build as part of the github continuous integration.
To keep the bazel build short, we maintain a copy of the subset of
packages that are needed for the TFLM (+ shared TfLite) bazel targets.

Eigen was updated for TF with b26252b38334ab18dc7be2547c41e91e78257d60
and we make the corresponding change in TFLM's pared down version of
workspace.bzl with this change.