Have tf.nn.relu of NaN output NaN.
Before it outputted 0.

In rare cases, it is possible this may cause some models using float16 to start getting NaNs if they were overflowing, but such models were not working correctly anyway. Previously, an overflow on the forward pass could get turned into NaN by BatchNormalization, which could then get turned into 0 by relu, hiding the NaN. Such models were not working correctly, as all the work done in the forward pass before the relu was discarded.

Also, on CPUs, relu(-0) is now 0, not -0. This is unintentional and arbitrary, but relu(-0) is already inconsistent across different devices and XLA and so -0 vs 0 cannot be relied on.

I tested performance on a Titan V with the following program:

    import tensorflow as tf
    import time

    def bench(dtype):
      x = tf.random.normal((2 ** 28,), dtype=dtype)
      p = tf.constant(0.)

      # Warmup
      tf.nn.relu(x)

      start = time.time()
      for _ in range(100):
        tf.nn.relu(x)
      # Synchronize GPU by sending result of computation to CPU
      p = p + 1.
      p.numpy()

      end = time.time()
      print('time for %s: %s' % (dtype, end - start))

    bench('float32')
    bench('float16')

I ran 3 times with and without this change and took the average. The results show no difference in performance:

Before:
float32: 0.381
float16: 0.190

After:
float32: 0.380
float16: 0.190
PiperOrigin-RevId: 357290491
Change-Id: Icc01f3bfb35ddfce97d4fd347167f0c41322e3ae