Implement resource and variant tensor sharing across multi-subgraphs
* Sharing TensorFlow Resource and Variant Tensors across multi-subgraphs

The existing Flex delegate handles the cases for the multi-partitions in a
subgraph well but when sharing tensors across subgraphs, we need to be careful
to implement since resource and variant tensors are RAII objects, which means
that the memcpy does not work on them. Other value-based tensor formats, numbers
and strings, can be easily a format of memory data and easily copied around
subgraphs by just copying data. Sharing the pointer of resource and variant
tensor should be done instead.

* Sharing TensorFlow tensor pointer across multi-subgraphs
To do that, Flex resource and variant buffer format is just a container for
storing the TensorFlow tensor object, which refers to TensorFlow resource or
variant tensor like the below code snippet.

struct OpaqueBuffer {
  // Store a TensorFlow's tensor pointer. The life cycle of the pointer will be
  // managed by the reference counting in the TensorFlow world and the pointer
  // will be freed when all the buffer maps, who own it, are gone.
  const tensorflow::Tensor* tf_tensor;
};

* Life cycle of the shared TensorFlow tensor object

When TFLite runtime requires those tensors to be transferred to the other
subgraphs, TFLite runtime invokes the delegate interface, the
CopyFromBufferHandle method. The Flex’s CopyFromBufferHandle method will
create the above OpaqueBuffer structure and store the corresponding TensorFlow
tensor pointer. Later, the other subgraphs will find the TensorFlow pointer
inside and insert it to their buffer map to track the life cycle in the
subgraph.

Even though the same TensorFlow tensor object can be found in the multiple
buffer maps in the Flex delegate. They will be freed when the Flex delegate is
finalized.

PiperOrigin-RevId: 358553465
Change-Id: I7cb056bf8f216851c771d7b2f26e69821a7cbf6a