[XLA:CPU] More accurate expm1 when x is small
We approximate it with:
expm1 = tanh(x/2)*(exp(x)+1)

Additional care is taken to handle the case when x/2 underflows but x does not
by simply approximating the result with x itself.

This suffices to get us within a relative error of 5e-7 or about nine ULPs when
compared against libm.

PiperOrigin-RevId: 358736039
Change-Id: I53a49e929edab1e3ba671b62b3d27495888c5011