[XLA:CPU] More accurate expm1 when x is small, take two
We approximate it with:
expm1(x) = tanh(x/2)*(exp(x)+1)

Additional care is taken to handle the case when x/2 underflows but x does not
by simply approximating the result with x itself.

Yet further care must be taken to handle the case when exp(x) would not be all that close to 1, in which case we simply use:
expm1(x) = exp(x)-1

The pseudo-code for this is roughly:
  if x/2 == 0:
    return x
  exp_x = exp(x)
  if |x| > .5:
    return exp_x - 1
  return tanh(x/2)*(exp_x+1)

The actual code sequence emitted preserves vectorization in the case where different lanes observe inputs where the magnitudes are entirely different.

This suffices to get us within a relative error of 4.76e-7 or about eight ULPs when
compared against libm.

PiperOrigin-RevId: 358861023
Change-Id: I4a51ec8e2a16a95b6cbaa2af3305ce3a16201c54