Add temporary fix to include gcs_file_system builds in the TPU build
There is an implicit assumption here that Env::Default() is loaded after InitializeCreateGcsFileSystemFnPtr is called, but this seems to hold out in practice. We can also do it the other way around if it changes (e.g. just write out all the Env* variables and then have the initializer register GCS file system to them later too).

PiperOrigin-RevId: 359315295
Change-Id: If40bdf07d6d58d0dd354e8ec1405e79340aa1d2d