Move pre-export verification to helper function and pass
This enables reusing in a pipeline and so getting an error reproducer if export pipeline doesn't produce expected output. Also change to return errors in context rather than use a status so that we can get more exact errors.

Currently this would result in verifying this invariant 2x, so we could add a ConvertVerifiedModuleToGraph entry point too to avoid this or just during export do it while exporting to avoid walking over the module again.

PiperOrigin-RevId: 359946650
Change-Id: Ia82ee95d27330842fbc83b7c6d1dd55439a6b2a4