Remove a callback to Python to get the value of some flag.
This is done to simplify the code, and not at all for performance, because it's only executed during the compilation phase.

One possible design question: should we let the user access the value of the flag if it has not been set? Right now, the Python code allows it I think (meaning the behavior may not match the flag value, which has not been parsed yet).
We could raise an error, by setting the flag value to absl::nullopt, and check it's not null. But it would be a breaking change, so I am a little reluctant doing so.

PiperOrigin-RevId: 360407549
Change-Id: I1bcf6787d4d6680b975f73dbe3cef9037931e264