[tf.data service] Report an error on consumer restarts in round robin reads.
In round robin read mode, we store the latest round of data in case a network error causes a consumer to need to retry. Before this CL, we would provide the data for the latest round N for any round request R such that R <= N. If a consumer was restarted with the same job name, the first N elements it gets from a worker would be identical. After this CL, we will report an error instead.

This CL also fixes an IteratorContext lifetime issue in data_service_dataset_ops that was causing the added test to fail.

PiperOrigin-RevId: 362099156
Change-Id: Id19837e4a403a6a7c40d2062f45e4d859bda991c