Update the tflite model "buffers" field checking rule.
If we don't use "--force-empty-vectors" flag[1] for flatc, the buffers
might be a null ptr if we serialize a model with zero buffers size(e.g.
all ops in model doesn't use const weights in model).
This commit relaxs the "buffers" null ptr checking for this situation,
and also updates the "subgraphs" checking for null ptr dereference.

[1]
https://github.com/google/flatbuffers/blob/fc4fffea41f5b682198edfc72eca536d33fd848c/src/flatc.cpp#L167