lite: nnapi: Check memory mapping errors after NNMemory()
There are no error checks in the constructor `NNMemory::NNMemory()`.
When creating a shared memory fd, `NNMemory::fd_` could be set to a
negative value due to errors. Thus, when trying to `mmap` this,
`NNMemory::data_ptr_` will be set to `MAP_FAILED` (and which in turn
will yield an invalid call to `ANeuralNetworksMemory_createFromFd()`) in
the constructor.

Ideally, one should check for errors early. However, since Google's C++
style guide prohibits the use of exceptions, we need to check this
explicitly after the constructor instead. We check `NNMemory::data_ptr_`
for any errors (since `NNMemory::fd_` is private).

Signed-off-by: Waqar Hameed <whame91@gmail.com>