Add DeviceAttributeToLaunchOp pass before DecomposeResourceOps pass.
Some resource ops such as AssignAddVariable are decomposed into multiple ops and the device assigment is lost. Ops with explicit device assigmnment are now wrapped in a launch op with the same device attribute.  When the op is decomposed it is within the LaunchOp.  The ops are later hoisted in the LaunchToDeviceAttributePass and the device attribute written for each decomposed op.

PiperOrigin-RevId: 362528023
Change-Id: Id03d375b6a54333b30bff9ff214614df6043d96e