Generalize tf-rewrite-tpu-embedding-ops pass to handle embedding ops within regions
Currently, the pass assumes embedding ops to be directly in a function and was working only because functional control flow to regions pass introduces call functions so the embedding ops ended up being in a separate function.

The above won't be true for this pass's usage to rewrite these ops in MLIR module before converting it to Graph for compilation using the old bridge second phase.

PiperOrigin-RevId: 362586719
Change-Id: Ieaabcc0db8a44f6c721d1842da0720e8d9389797