Explicitly delegate learning_rate in LossScaleOptimizer.
This allows LossScaleOptimizer.learning_rate to be accessed and set when LossScaleOptimizer wraps another optimizer wrapper, such as tfa.optimizers.Lookahead.

Normally LossScaleOptimizer finds hyperparameters to delegate by checking the attribute `inner_optimizer._hypers`.  However, the _hyper attribute for wrappers don't have the hyperparameters, so LossScaleOptimizer does not correctly delegate them. Handling this in general is tricky and likely requires adding a new API, which we want to avoid. Instead, we special-case learning_rate by unconditionally delegating it. Other hyperparameters are still not expose when LossScaleOptimizer wraps another wrapper, but in practice usually the only hyperparameter that is accessed on an optimizer is learning_rate, so this is OK.

Fixes https://github.com/tensorflow/addons/issues/2403 and https://github.com/tensorflow/addons/issues/2332.

PiperOrigin-RevId: 362589168
Change-Id: I989771ad505fa9ec85ece05fb7adfdf95f521537