Use old bridge in _TpuCompileMlir op unless MLIR bridge was enabled explicitly
This allows incremental rollout of the bridge.

Old bridge is invoked by,
* Deserializing MLIR
* Using mhlo.is_same_data_across_replicas arg attributes in the main function to update metadata
* Exporting the MLIR module to a temporary function library after running a pass pipeline
* Compiling the function using the old bridge code path

If the conversion to graph or compilation using the old bridge fails, then compilation is aborted instead of trying out the MLIR bridge. This makes the expectations clearer. One possibility could be to evolve this similar to first phase of the bridge.

PiperOrigin-RevId: 362599636
Change-Id: I4f1fc0620637ab1bb9a9a8663774b3ad4442a8be