Export StructuredCoder as an internal API.
StructuredCoder is the only thing that decodes the input signatures from the SavedModel concrete functions, and Keras uses it to build the layers loaded from config. The layers must be built before loading the other functions/objects from the SavedModel in order to prevent them from creating new sublayers (which would not have the correct weights). Thus, we can't remove the use of StructuredCoder from Keras right now.

PiperOrigin-RevId: 363539996
Change-Id: Ie1a0083adb11ccb16ac729d1e0c701152de4cd93