[XLA:HLO] Cleanly destruct fusion instructions and their computations.
If a fusion instruction dies before its computation, the computation becomes unreachable, but it may not be destructed until later. In this case, we need to clear its back-pointer to the fusion instruction, otherwise it will point to freed memory. Similarly, when a fusion computation dies, we must clear the fusion instructions's called_computations_.

We also add a boolean field in HloComputation to determine if it is a fusion computation, since with this CL we can have (unreachable) fusion computations whose fusion instruction is null.

Last, do a nullptr check during module cloning, to avoid runtime errors due to unreachable computations.

PiperOrigin-RevId: 363541665
Change-Id: I9229083bd647c7b6d98cf0751558bc8b45bc7887