Add _handle_dtypes and _handle_shapes attributes on export.
The dtype and shape information for resource arguments has to be handled
specially: GraphDef encodes that information in _handle_dtypes and
_handle_shapes attributes while MLIR encodes that information using
!tf.resource subtypes.

When exporting MLIR to GraphDef, add _handle_dtypes and _handle_shapes
attributes for resource arguments. This is consistent with import, which
already has a special case for reading the _handle_dtypes and _handle_shapes
attributes on import from GraphDef to MLIR.

This fixes RET_CHECK arg_shape.handle_type != DT_INVALID errors that happen
when converting GraphDef to and from MLIR before passing the graph to the old
TF/XLA bridge.

PiperOrigin-RevId: 363957495
Change-Id: Ie28e88acd6789e289b69c988317db77459966615