Support serialization for dataset-initialized lookup tables.
This required changing how we represent serialization logic. Previously, we used `InitializerAsGraphDefFunc`, a typedef of std::function. This caused issues because serializing dataset-based initializers requires taking a reference to the dataset, to prevent the dataset from being destroyed. std::function doesn't give us a way to release this reference. To handle the lifecycle correctly, this CL switches to using a `InitializerSerializer` class for defining serialization logic.

PiperOrigin-RevId: 365597208
Change-Id: I4fefbbdf8fcde9e53bf80a87f6123a921bf07054