[TF:TRT] Minor code clean up.
Previously, when we use an AsyncHelper object in the asynchronous execution
macros, such as OP_REQUIRES_OK_ASYNC, we only use the noop operator() of the
object. On the other hand, the same AsyncHelper object is also used to call
DoneCallback to signal the completion of the TRTEngineOp execution. This CL
adds DummyAsyncHelper class with a noop operator() and use it in those
asynchronous execution macros. The can help simplify the AsyncHelper class and
make it clear that the asynchronous execution macros should not to call the
DoneCallback for TRTEngineOp.

Revise some comments. Reduce nested levels in some code.

PiperOrigin-RevId: 365674961
Change-Id: I705cc144732b53c042fcb952a3d85c83ce595620