Switch TF filesystems to keep backwards compatibility.
Recent changes to enable modular TF can result in breakages or require majority of users to take quick steps to fix their broken code. Instead, we should enable the modular filsystems only if requested by an environment variable and use the legacy behavior in all cases.

In a future TF release, the cloud filesystems will come from different pip packages but we have to design the path towards that to ensure that no one in the TF ecosystem is broken by this. As we stand now, without this change most TF users will need to take action in a P0 basis and this is not desirable.

CC @yongtang, @kvignesh1420, @vnvo2409. Note that I also changed the name of the envvar to match the logic and you'll have to do the same for SIG IO code. Let's sync on SIG IO meeting to discuss further

PiperOrigin-RevId: 365940168
Change-Id: Id03b793613e521999af47fd6017b70ff659bd23b