Canonicalize after resource op lifting pass to eliminate pass through results
Alternatively, we could only run canonicalization patterns only for cluster ops within resource op lifting pass but that might be premature optimization and can be done later on.

Added a simple mlir test for end to end testing of the pipeline but this may have some maintenance overhead if the output of the frontend changes. We can drop if this require manual updates.

PiperOrigin-RevId: 366881337
Change-Id: I803283308e718a99dc437be2536ae8a296ec5d10