Enable experimental SavedModel support for SummaryWriter using TrackableResource
This introduces a new `experimental_trackable` boolean argument to `tf.summary.create_file_writer()`. If True (default is False), the created `SummaryWriter` will be a `tf.saved_model.experimental.TrackableResource`, allowing the writer to be saved in a SavedModel as a property on a `tf.Module`, and then used within @tf.function methods on the module.

There are a couple significant limitations to this support:

1) The logdir passed to `create_file_writer()` will be baked into the resulting SavedModel and cannot be changed later. One workaround is to specify a relative path, and then ensure that any code loading the SavedModel is using a working directory into which logs can be written.

2) Initializing the SavedModel resources (e.g. when loading it back into Python) will recreate the writer and open a new file, which may be a surprising side effect. Possible future solutions would be deferring either the writer creation or the opening of the file.

PiperOrigin-RevId: 367347580
Change-Id: Ib6b614aaf5edb248e7b6de690d9f3d367a4c97cd