Mark packages with layering_check violations.
This is a no-op for OSS builds because currently we do not use layering_check.

Support for the layering_check feature has been added to bazel's autoconfig cc_toolchain (see https://github.com/bazelbuild/bazel/pull/11440), but TensorFlow uses custom toolchains for most builds. We might enable layering_check sometime in the future.

Google-internal builds traditionally disabled layering_check for cuda-adjacent code. This change is a preparation for enabling it.

PiperOrigin-RevId: 368264711
Change-Id: Ibdd94f09fe3841368430032e4f84e493660913a2