Modify v1 island coarsening to sort islands between first op and last op merged into the island cluster.
Earlier we sorted ops from the first op merged into the island and any of the special TPU outputs ops that would be present at the end of the vector of islands to be merged. Note that it did not consider the order between special TPU output ops. It can lead to failure if two consecutive special TPU output ops are added to the island vector out of order.

This change adds documentation as to why the sorting is required and does it between the first island merged and the last island in the block that was merged.

PiperOrigin-RevId: 368469534
Change-Id: Iacf77a96d5421531352630123482c3fd753d5ca7