Unconditionally set _xla_propagate_compile_time_consts attribute in region_control_flow_to_functional pass
This is to fix one of the failures in old bridge phase two legalization expecting a constant. This happens because phase one of the MLIR bridge moved operand of the op outside of the loop body because it was loop invariant. Also, _xla_propagate_compile_time_consts attribute was set to false for the op so the constant didn't get propagated during legalization.

We need to set _xla_propagate_compile_time_consts to true while going from region to functional control flow given the possibility of transformation for LICM. The attribute was supposed to be enabled globally anyway and MLIR bridge phase two does that unconditionally.

PiperOrigin-RevId: 368783012
Change-Id: I26751014e73c172bba6ea02e7c1aa2ad41bcfd30