This allows failures when searching for executables. The previous attempt to fix this did not succeed, because searching with `which` on Windows outputs "INFO: Could not find files for the given pattern(s)." to stderr.
I thought the best course would be to update the variable name, because I don't believe this will negatively affect those functions already using it. Let me know if you would rather me add a new parameter for this instead.

Also, the which command returns the empty string (which is not None), so I updated that conditional to check for blank strings as well.

PiperOrigin-RevId: 369272869
Change-Id: Ibe85795482e78de40ced14001e85198881866256