[tf.data] Add `stop_on_empty_dataset` param to `{sample,choose}_from_datasets`.
If `True`, sampling stops if it encounters an empty dataset. If `False`, it
skips empty datasets. It is recommended to set it to `True`. Otherwise, the
distribution of samples starts off as the user intends, but may change as input
datasets become empty. This can be difficult to detect since the dataset starts
off looking correct. Default to `False` for backward compatibility.

PiperOrigin-RevId: 369281872
Change-Id: Ibbb769092f5c06933561358c1a2860f4b751bb78