Add a cache object to avoid repeated materialization of while loops.
Previously when we need to trace a single value from a gradient while to the forward while, we need to materialize the loop once to look inside. If a while loop has multiple values, we repeat the process multiple times. This cl adds a cache to speed the process up.

PiperOrigin-RevId: 369489919
Change-Id: I52add2d6f73a307921548a27c7832853caff2692