Remove all _xla_outside_compilation attributes after extracting outside compilation from device cluster.
After running TPUExtractOutsideCompilation, there shouldn't be anymore outside compilation ops inside of the device cluster.  Sometimes ops with the _xla_outside_compilation op are lifted out of the cluster and the _xla_outside_compilation attribute is meaningless.  However it can cause errors when running the old bridge as the attribute is not expected.

PiperOrigin-RevId: 369960705
Change-Id: Ieb20ff5bb6c5d82b1db3843fa7691c926664ea61