[TF:TRT] Handle Shape OP outputs as TRTEngineOp outputs.
The GPU Shape OP has its output in the host memory. When the Shape OP is inside
an TRTEngineOp and the TRTEngineOp segment executes natively, TF assumes all
the outputs of the TRTEngineOp are in GPU. To fix this problem, we add a node
right after the Shape OP to copy its result to GPU.

Add a test case.

PiperOrigin-RevId: 369960959
Change-Id: I36b04f7c6330d1e69e204ef8448285b2f6ab3aea