Add /usr/local/cuda to the list of CUDA data directories that XLA searches by default.
My installation of JAX chose /usr/local/cuda-11.2 as its default CUDA data directory. That's a fine first choice, but the problem with it is that it names a specific CUDA version, when CUDA 11.1+ are backwards compatible between minor releases. So we would also like to try looking in an unversioned path, and /usr/local/cuda is a very popular choice.

This change is in preparation for JAX reducing the number of CUDA wheel variants that it ships.

PiperOrigin-RevId: 369961339
Change-Id: I45a22ecc465a169112337292071f24f29866a3e7