Remove one outdated TODO
Recently we changed the operation visit order in the prepare_tf pass to be
top-down. This means the tf.Identity and tf.IdentityN between the min/max
constants and fake quant op has been removed when the fake quant op is visited.
So we don't need to skip the tf.Identity and tf.IdentityN ops to retrieve the
min/max constants.

PiperOrigin-RevId: 370103306
Change-Id: I1c58f34348156c7e41840bf96931c4c0b9a8170b