Optimize the implementation of Island Coarsening
The current implementation creates islands greedily whenever a merge pair is found. This leads to very problematic behavior due to the fact that splicing a region is O(N) with the size of that region, and island coarsening often results in lots of islands being merged together. This revision refactors the implementation to first analyze which islands to merge, and then merge them all together as one step at the end of the process. This leads to a ~50x speedup. There are still a few bits of improvement, mostly related to reducing malloc traffic, but this revision is already such a big reduction that I figured this should go in first.

PiperOrigin-RevId: 370511515
Change-Id: Iad5042a340a9fcaf8a4a92fc767b6702eee0508b