Modify the check for model parallelism and outside compilation combination.
Model paralleism and outside compilation are not supported together.  However in order to convert to/from outside compilation attribute to launch op, we need the host device.  Getting the host device fails when there is model parallelism.  To fix this, this CL checks for model parallelism when converting to the launch op representation of outside compilation.  When converting from launch to outside compilation attribute, we just skip processing if model parallelism is encountered because the error should have been triggered earlier.

PiperOrigin-RevId: 370564512
Change-Id: I6e8e9875506bc8c6583e1cb670daa33afe94ad14