Fix for SWDEV 277899 - hipLaunchKernel called with threads per block value greater than launch bounds attribute.
Prior to the fix for SWDEV 252801, if the application code launches a GPU kernel with a threads per block value which is greater than the `__launch_bounds__` attribute value for that kernel, HIP runtime only issues a warning for this scenario.

After the fix for SWDEV 252801, HIP runtime issues an error message instead (for this scenario), and reports an error status back to the application code.

This change results in the following four TF unit test failures

```
//tensorflow/compiler/tests:depthwise_conv_op_test_gpu
//tensorflow/compiler/tests:depthwise_conv_op_test_gpu_mlir_bridge_test
//tensorflow/python:nn_grad_test_gpu
//tensorflow/python/keras/distribute:saved_model_mixed_api_test_gpu
//tensorflow/python/kernel_tests/signal:fft_ops_test
```

This commit fixes the failure in the first three of the above tests. It was caused by an incorrect calculation of the "threads per block" value. The upper bound for that value (i.e. the `__launch_bounds__` attribute value) was not getting passed to the `GetGpuLaunchConfig`. Passing that value, results in `GetGpuLaunchConfig` returning the correct "threads per block" value and the unit tests start passing