Fix divide_no_nan for near-zero complex inputs (#48350).
The Eigen packet `pdiv` operation for complex types divides by the
squared-magnitude of the divisor. If this underflows to zero, the
result is a NaN.  To prevent NaNs, modify condition to check for this,
setting the result to zero if the magnitude-squared of the divisor is
zero.

NOTE: Eigen's underflow behavior for complex division is consistent with both
gcc's implementation and with pytorch.  Clang does special-handling to prevent
such underflows at the expense of additional cycles. If comparing to numpy
built with clang, we may see differing results. We also see different results
when comparing between CPU and GPU due to different handling of subnormals.

Fixes #48350.

PiperOrigin-RevId: 370687832
Change-Id: I9b50a637c8db02685ab23bcbe433b7ca83f96d23