[tf.data] `sample_from_datasets` skips datasets with 0 weight.
When `stop_on_empty_dataset` is `False`, `sample_from_datasets` may hang if a
weight is zero. It is better to skip datasets with 0 weight to support these
use cases.

If `weights` is a Dataset, this check doesn't apply. The caller needs to set
`stop_on_empty_dataset` to `True`.

PiperOrigin-RevId: 370930858
Change-Id: I3d55dd9e0e9a1d41ecc9194ab873e64ee5867a53