[tf.data] `sample_from_datasets` skips datasets with 0 weight.
When `stop_on_empty_dataset` is `False`, `sample_from_datasets` may hang if a
weight is zero. It is better to skip datasets with 0 weight to support these
use cases.

If `weights` is a Dataset, this check doesn't apply. The caller needs to set
`stop_on_empty_dataset` to `True`.

PiperOrigin-RevId: 370970351
Change-Id: Ide67ae6d29f46ac8b9cd3e3515434a8dd0631e57