Fix bug with `deprecated_arg_value` when the value is None.
The current behavior crashes if the user attempts to pass a tf.Variable, because testing `variable == None` raises an error under eager mode / TF2 tensor equality semantics. (debatably, this is a bug in the semantics of `Variable.__eq__`, but my naive attempt to fix it there broke lots of other stuff, so I'm hoping to at least work around it here).

This includes an unrelated change, replacing `inspect` with `tf_inspect`, that I had to make to pacify a presubmit check.

PiperOrigin-RevId: 371808872
Change-Id: I22d4940088061fb96f794240a0db031025270ff5