[MLIR] Add verification for XLA HLO -> LMHLO, and fix a verification failure.
Specifically, sometimes the arg type doesn't equal to the type of its use, even if the arg is an entry computation parameter. This is because program_shape in the computation may not match the operand shape. Add a MemrefReinterpretCast if two shapes don't equal.

The verification failure is discovered in https://github.com/google/jax/discussions/6645.

PiperOrigin-RevId: 372179297
Change-Id: I0865ff030217410cb5913e6386a53d7a68d3b7dd