Update keras metrics to use memory efficient alternative when collect values for evenly distributed thresholds.
This implementation is based on the example in tf.slim. It exhibits a run time and space complexity of O(T + N), where T is the number of thresholds and N is the size of predictions. Metrics that rely on standard implementation instead exhibit a complexity of O(T * N). It could save a lot of memory when N is large.

Added a unit test to verify the memory consumption. Under eager context, the ratio of memory between old and new approach is between 80 and 500. Set the limit to 50 to avoid the flakiness.

PiperOrigin-RevId: 373469207
Change-Id: I95ed482dc4abde5de829d3018e445c9ecb62d2d6