Allow localized disabling of tests for asan/msan/tsan/ubsan.
Currently we can disable entire test targets via the bazel tags `noasan`,
`nomsan`, `notsan`, `noubsan`.  However, in some instances, we may expect
one particular edge-case to fail a sanitizer test (for known reason), but
still want the rest of the test suite to be analyzed. By disabling individual
test methods or parts thereof locally, we allow for this without the need to
separate the offending case into a different target.

There already was some functionality in tflite, which seems to currently
be unused (though has been previously to locally skip tests).
Refactored and moved to tf so can be more widely used.

Also added corresponding defines to the `.bazelrc` file.

PiperOrigin-RevId: 374231232
Change-Id: I62ca36ebe9edec18e0d6d358496b6d07efb43574