Remove const_cast from IntArrayFromInts
 * The primary change is in micro/test_helpers.h and micro/test_helpers.cc
 * This change is accepting the fact that the parameter can not be const and we should treat is as such.
 * Changed all the callers of IntArrayFromInts to pass in non-const pointers

One of the downsides of this bulk find and replace is that many of the variable
names no longer follow the appropriate style. For example, what used to be a
`constexpr int kShape[]` might now be `int kShape[]`.

This change has made the decision that such deviation from the style is
acceptable since we should instead focus effort on improving the kernel testing
infrastructure.

See http://b/188459715#comment2 for additional context.

PiperOrigin-RevId: 374506016
Change-Id: I9fed688353b4a39ae6537e0b260af420ddbe771c