[tf.lite] Add GPU manifest for targeting Android 12 ("S")
Targeting Android 12 requires explicit declaration of any referenced
native vendor libraries. Include the necessary tags in a GPU-specific
manifest, such that downstream apps inherit them when adding the GPU
library as a dependency, enabling use of vendor-provided OpenCL libraries.

Note that for apps not targeting Android 12 ("S") explicitly, there
should be no change in behavior (with or without this CL).

Also note that this manifest cannot be used by default for OSS builds,
as it requires AGP 4.2.0; this will be addressed in follow-up work.

PiperOrigin-RevId: 374959792
Change-Id: Ie10fff602cbc1f031113204e8a7c53a7df24ee4a