Fix the elemental emitter for `add` on `pred`.
Previously it would emit integer addition instructions on the `i8`
representation of a `pred`, which resulted in out-of-range values that caused
havoc in various places (e.g. made `Literal::ToString` print `pred[]
tru\0\0Ca)`, and made `not` fail to actually negate the result).  Now it emits
a bitwise or instruction instead, which maps in-range values to in-range
values.

This also comes with equivalent transformations for other permitted arithmetic
ops on `pred`, which were not incorrect, but needlessly expensive.

This fixes the immediate miscompile on the linked bug (and
https://github.com/google/jax/issues/5406), but since that bug revealed other
inconsistencies with MLIR conversions, it's not yet marked as fixed.

PiperOrigin-RevId: 375520234
Change-Id: I90be5e6c5b9532a3a65b5a22419142ac1b5df407