Account for the difference between HLO `add` and MHLO `AddOp` on `pred`.
Using the HLO interpreter's behavior as a reference, addition of booleans
should behave equivalently to `or`.  The MLIR mhlo/lmhlo `AddOp`s do not behave
equivalently to `or` on `i1`, so this special behavior needs to be implemented
explicitly by lowering to `OrOp` instead.  Since there are two independent code
paths for translating HLO ops into MLIR depending on whether they're fused or
not, they both need to be updated and tested separately.

In the other direction, the MHLO `AddOp` corresponds to XLA `xor`, and since
that conversion is in generated code, it's most expedient just to intercept the
problematic `AddOp`s beforehand.

PiperOrigin-RevId: 375594217
Change-Id: I79cad36035b15bb8ff2f1d4aa890a53996a30a81