Add uint support to tf.sort and fix skipped tests.
For sorting unsigned ints ascending, we cannot negate values, so we need to
subtract from the max unsigned value.

Also fixed a bug for signed ints if the list contains dtype.min.  We cannot
simply negate, since that would overflow and result in an incorrect sorted
order.  Instead, we need to convert to unsigned.

PiperOrigin-RevId: 378072881
Change-Id: I097e068e86bdbf4f37b60d16a13361f05f0999ec