Modify ReluGrad pattern to enforce gradients and features have the same shape.
By deriving the zero constant with ConstantLike from features, HLO_SelectOp will
indirectly enforce that features and gradients have the same size. In the
previous pattern, this is not enforced, because HLO_SelectOp allows a scalar
first operand that would be broadcasted to match the second and third operand
shapes.

PiperOrigin-RevId: 378348036
Change-Id: I3c24feb19a133a58277bd2c4510cdb595006cea6