[tf.data] Remove spurious warning.
The mechanism used for checking for the anti-pattern of creating resources within tf.data user-defined functions is incorrect as it does not account for a) the resource creation happening within an `init_scope` (which leads to the warning being emitted spuriously) b) the resource being created in a nested tf.function which has been traced earlier (which leads to not emitting a valid warning).

This CL removes the current mechanism as resources commonly created within user-defined functions -- variables and lookup table -- are hoisted automatically.

PiperOrigin-RevId: 378921610
Change-Id: I21b244264dd57d90edf8a387bec734818e2ecdc2