Be explicit about the type signature of map_func
Highlight tf.data.Dataset such that the reader doesn't miss it. The typing was mentioned at the top of the doc but not immediately visible at the params table.

Motivation for this change is that after studying this doc, a user still came up with code calling `interleave` with a function that takes element and returns element.