ValueInference -- Simplify expression before evaluating.
Previously we evaluate the expression `upperbound(a+(b-a))` as
`upperbound(a) + upperbound(b - a)`, which is then evaluated as
`upperbound(a) + upperbound(b) - lowerbound(a)`

However this bound can be tighten to upperbound(b) if we do a quick
algebraic simplification on the expression before evaluating.

This is even more useful when lowering a pad op to pad to a static value.

```
size = tf.size(T)[0]
pad_width = MAX_SIZE - size
tf.pad(T, pad_width) # The result size is a static value size + MAX_SIZE - size = MAX_SIZE
```
PiperOrigin-RevId: 383955840
Change-Id: Ie6c19080b0829dd27efe36fd0b6c92002d3399b9