Fix fused convolutions that don't use relu with the new cudnn api.
Previously, all fused convs applied relu, if you use the new cudnn API.

Unfortunately the unit test still does not catch this.  Even if you try to
force the conv to run on CPU, it ends up running on GPU, because there's a CPU
implementation only for int8 convs.

The unit test (badly) needs to be improved, but I don't think that should stop
us from making this change.

For now, don't use the new cudnn API for fused convolutions that use the
identity function as their activation.  It seems that in order to do this, we
have to disallow certain algorithms, same as with the old API.  But with the
new API it's unclear how to tell which algorithm we're using, so it's unclear
which plans we should disallow.

PiperOrigin-RevId: 385030346
Change-Id: I34e4acd7d1c26cc49cb309982e911bde4c0fb381