Make resource alias analysis less conservative
Now we handle ops with `MemoryEffectOpInterface` less conservative which makes
side-effect analysis more efficient. In particular, for a resource-allocating op
with empty or anonymous `shared_name` attribute we can assume that the resource
being created is unique which wasn't utilized before. One implication of this is
less control dependencies in the exported GraphDef.
Also added tests and improved some comments.

PiperOrigin-RevId: 385231241
Change-Id: I9614d050cb1a521dc34befa3e4189c0a2794493a