[XLA/GPU] Generate `lmhlo.fusion { mhlo.some_op }` instead of `lmhlo.some_op`, for most of the ops.
This simplifies the emitter for quite a bit, as kCopy and kReduce emitters now don't have to deal with unnested cases.

This also simplifies XLA HLO -> LMHLO conversion, since now there are much less kinds of LMHLO ops to be converted.

Per chat discussion between sanjoy@, jurahul@ and me, we think it's a good idea to move away from LMHLO and only use lmhlo.fusion + MHLO. This way, lmhlo.fusion effectively just acts as a composable bufferization annotation.

This CL is a preliminary experiment for that plan, as it removes some uses from XLA/GPU. Once we are confident about this approach, we can officially propose the deprecation and removal of LMHLO (except for .fusion).

PiperOrigin-RevId: 385561461
Change-Id: I6e91c598788363f10f954529eea822aeec523c17