Remove quantization parameters from input and output values
After the decomposition is finished, quantization parameters are already decomposed into proper scaling ops in the graph, thus quantization parameters are no longer needed, nor can be understood by the TF graph. So the quantization information is removed from input and output tensors, and only the storage data type is used for the tensors' element type.

PiperOrigin-RevId: 386029733
Change-Id: Ic8ea799d5d8df66d7539acfcad995bb450d6c71f