Gracefully handle encoding WAV files with zero length.
Removes checks for positive frames and non-nullptr audio in EncodeAudioAsS16LEWav.

When creating audio summaries for zero length tensors, the backing data pointer for the tensor of samples is nullptr which causes EncodeAudioAsS16LEWav to return an InvalidArgument. When the number of frames to encode is zero, it is ok for the audio pointer to be null.

TESTED:
- unit test
PiperOrigin-RevId: 386041567
Change-Id: I327f07574d49dcfe93c5588e0626bf4f7e12fdd8