Legalize mhlo.not to TFLogicalNotOp
In fact, there was a pattern in dealing with mhlo.not. However, that operation included a mistake. It required the operand to be TF_IntOrFpTensor and it also demanded SameOperandsAndResultType. Under this circumstance, the pattern became a vacuous one as the result type of NotOp must be Bool and the input operand is IntOrFp. I fixed this bug by simply replacing TF_IntOrFpTensor with TF_BoolTensor in the definition of the pattern.

PiperOrigin-RevId: 387274966
Change-Id: Iaad360036ff5dbc7ece4598789bbd618baccdd69