Raise an early exception when TensorSpecs are missing in input_signature.
The input_signature is verified at runtime instead of declaration time here because at declaration time a method is just a function, not bound to an object, and the verification will fail due to lack of TensorSpec for `self`. At runtime, if it is called as a method, the `self` is excluded when comparing to input_signature.

Update the is_method function to handle methods wrapped in functool.partial.

PiperOrigin-RevId: 387674999
Change-Id: Ic18c51b9145ed1b3912466d9d263eb2e87e4e898