Start a new dialect to model TensorFlow graphs
This dialect has the capability to model accurately GraphDef. We expect to be able
to roundtrip perfectly the protobuf serialization back and forth to this dialect.
This does not make it easily usable as-is, since GraphDef does not include all the
information to describe an operation without having the OpDef available.
More infrastructure will be added later to inject state from the OpDef into Graphs
expressed in this dialect.

This dialect supports V1-style graphs: flat "soup of nodes" where the feeds and
fetches aren't specified, as well as functions where the "soup of nodes" is
contained in a Function which defines a "signature" fixing the inputs and outputs.

PiperOrigin-RevId: 387827427
Change-Id: Ia5b24e7676767a82d77b7c5803c193d04ba350cd