(lite) Extract VerifyFlatbufferAndGetModel to an internal target
Rename VerifyFlatbufferAndGetModel, which was previously a
private subroutine of verifier.cc in an unnamed namespace,
as tflite::internal::VerifyFlatBufferAndGetModel, and move it into
a separate file verifier_internal.cc with a separate header file
verifier_internal.h, with restricted visibility, so that it can
also be used from the Java API JNI code.

This change is in anticipation of adding this function to the shims and changing the Java API code to use the new shims function, rather than just directly calling the
FlatBuffer routines.

PiperOrigin-RevId: 387926502
Change-Id: Ifc69ae4b46a1799553688734259aa47b3997fdc9