Remove "at least 2D" rank expansion in fit/predict/evaluate.
This is potentially a breaking change for subclassed models that process scalar data, as well as for certain Sequential models that were called with incorrectly-shaped data.

Previously, if Model.fit/evaluate/predict saw a batch of data of shape (batch_size,), it would uprank it to become (batch_size, 1). After this change, no upranking is applied.

In addition, this change will cause Sequential models to start enforcing the input shape assumptions they were created with. Previously, Sequential would enforce the input shape assumptions of the first layer in the model. If the first layer was a Dense layer, input rank would not get checked, meaning that a model such as `Sequential([Input(shape=(dim,)), Dense(...)])` could get called with data of any rank as long as the last axis had the correct dimension. After this change, only data of shape (batch_size, dim) would be accepted (i.e. rank 2). If you were in this situation, you need to update your Sequential input shape so that it has the same rank as the data you're going to pass to the model.

Note that this does not affect Sequential models defined with an explicit input shape (as long as they're called on correctly shaped data), nor Functional models, because such models will standardize their inputs to what they expect. For instance a model built with `Input(shape=())` that receives data of shape (batch_size, 1) will reshape it to (batch_size,) and inversely a model built with `Input(shape=(1,))` that receives data of shape (batch_size,) will reshape it to (batch_size, 1).

PiperOrigin-RevId: 388347064
Change-Id: I71e286db58d8dc6bf22132899abb251b248b0385