Add CopyCleanup pass to Kernel Generator.
This pass will be needed once the canonicalization which wrongly casts affine
maps away is fixed.
Currently, BufferCast(TensorLoad) is canonicalized to Cast even if the
TensorLoad operand has a affine map and the BufferCast doesn't. The fix will be
to copy the data. This pass will remove the copy and use the source of the copy
in the linalg::GenericOp.

PiperOrigin-RevId: 388868812
Change-Id: I55ce46d4b279a227d83752f56a4bb07680c6845d