Document that alias renaming could lead to failures
Selective op registration uses a textual representation of the C++ type which could mismatch with renaming to underlying type if the header is pre-generated and checked in. In which case the checked in header file would need to be updated else execution would fail with missing kernels (e.g., kernels would have been filtered out as the allowlist has different spelling).

PiperOrigin-RevId: 390416355
Change-Id: Id59220597d609f630145e8b5296e5498d519dead