Add bias adjustments when the scale for the bias would cause overflow.
In full-integer quantization, bias_scale is determined as input_scale * filter_scale. When the two dependent scales are too small, the INT32 quantized bias values will overflow and break the model. This change is to introduce a logic implemented in TOCO quantizer to MLIR's QuantizationDriver.

Deleted temporary patch that used fixed threshold for this adjustment.

PiperOrigin-RevId: 392550728
Change-Id: Ie7654655181ce083703b6add72d0a8389669f386