[XLA] Adjust CollectiveMatmul heuristic to be more aggressive.
When checking if an operand to the collective matmul would be
re-used by multiple users check if the other users share the same
sharding or need different sharding/no resharding at all.
Only in the case the same sharding is needed by the two op not doing
collective matmul will actually save time , because in the other cases
we generate different resharding ops anyway.

PiperOrigin-RevId: 394047164
Change-Id: I05923132983334befb5a55e3805136f303d6f169