Move DelegateTrackableMixin to core TF (and add __internal__ api symbol)
As SavedModel refactors are going on, there are many updates to the methods in Trackable. Since Keras lives in a different repo from core, it's hard to update the loss_scale_optimizer file at the same time as the refactor is happening. This CL makes it so that the two files live in the same repo.

Given the side effects in SavedModel, this class should only be used by loss scale optimizers (which is registered and loaded with Keras).

PiperOrigin-RevId: 394073420
Change-Id: I23bc83f7ff059db35c3b3865058dedfe0ac9c36b