Introduce a cache to avoid O(n^2) behavior with GPU MOF.
Some models have many multi-output fusion opportunities.  When we evaluate
these, we do an O(n^2) iteration over all pairs of nodes in an equivalence
class.

Before this change, we would recompute the shared memory required for each
fusion, and the number of unnested reductions in each fusion, O(n^2) times.
Computing these is tantamount to iterating over the whole fusion node, so is
especially expensive when the fusion is large.

Now we cache these values, so we compute them only O(n) times.

PiperOrigin-RevId: 394539476
Change-Id: I9b8dba935ee9aa6735b952977091f1c2b62c5311