[MLIR][KernelGen] Fix GPU module cache and make it a kernel resource
This fixes a subtle bug in the GPU module caches, which assumed that no two GPU
modules are loaded from the same memory location. When JIT-compiling kernels,
however, this can (and did) happen, which previously lead to a false cache hit,
resulting in the use of an incorrect GPU kernel.

Making the GPU module cache a kernel resource binds its lifetime to that of the
kernel context and also to that of the JIT cache....

PiperOrigin-RevId: 395508555
Change-Id: I4f3e3be47d643a57c4e7a1237ac5f95b2c143c53