When determinism is enabled, raise error in tf.timestamp.
A FailedPrecondition error is raised. InvalidArgument is used for random ops when the seed isn't specified, but that doesn't apply here as there are no arguments.

tf.timestamp is mostly used to seed random ops. It is also occasionally used for measuring the performance of some part of the model. The latter case is a legitimate use case when determinism is used, but we must raise an error to guard against the former case.

PiperOrigin-RevId: 395516238
Change-Id: I2ca33fede24a674793d128bbbd73179a9b3c3d16