Rewrite ParallelMapDataset in make_deterministic.
This makes calling Dataset.map deterministic when stateful ops are in the map_fn (assuming the user doesn't pass deterministic=False to map or tf.data.Options).

Unfortunately, this can make a model's input pipeline very very slow when determinism is enabled. For example, on the official resnet50 model with 8 V100s, performance drops from 5199 images/sec to 331 images/sec, making performance only 6% of what it is without determinism. The input pipeline is probably even slower than 6% of it's original speed, as the model is bottlenecked by the GPU without determinism and by the input pipeline with determinism. In the future, I hope to improve performance when determinism is enabled.

PiperOrigin-RevId: 395542882
Change-Id: If76a3d6b2508717cf72bb2f410b100df462f6105