Fix buggy original implementation of the sizetracker git version workaround
My original version of this fix last week was a bit buggy; I forgot that the
git_pretty function can be given a range of commits, and that the CL would be
different for each result. This new fix is much simpler, and does not act
differently for different versions of git, so I'm not going to miss it again.

The data is still correct, but results with the original fix do not correctly
include a full range of CLs, and they instead repeat the same one many times.

PiperOrigin-RevId: 395577412
Change-Id: I6e014343761b2f7ee4f72682183ab46d5e2eccfd