Roll out refcounting ResourceHandle to AnonymousResourceOp.
Only one user, AnonymousIteratorOpV2 is converted.

Associated fixes:

1. Release GIL when Eager dealloc calls into TF, which may call into Python from a different thread (Executor::Run always create a new thread).

2. Further trivialize TF_TensorFromTensor and TF_TensorToTensor, stop special treatment of DT_RESOURCE outside of the session Ops. (a step towards more cleanups near the TF_Tensor API)

The original motivation was to avoid calling into the (currently) undefined behavior of serializing a ref-counting ResourceHandle during SummarizeValue, which calls into these functions as part of TensorHandle::Resolve.

wangpeng and I agree that eventually upon / before serialization, Refcounting ResourceHandle shall register itself to the ResourceManager, with a unique name.
This will be done soon and the error message can be removed after that.

PiperOrigin-RevId: 395803716
Change-Id: Id5df8c64d9587302273b52bdc257931f23b4f88e