Adding `async.execute` lowering to `tfrt_test.do.async`.
This allows scheduling GPU work on multiple streams.

The implementation splits the `gpu-to-tfrt-gpu` into 4 separate passes. This is necessary because the (gpu.token to chain+stream within blocks and to chain+event across blocks) conversion of doubly-nested (func and async.execute) code became too complex, with the different patterns (including ones rewriting the op and the tail of the op's region) interacting in unexpected ways.

The passes now adhere to the recommendations in the 'Type Conversions the Not-So-Hard Way' [talk](https://mlir.llvm.org/talks): the individual passes rewrite a set of ops and insert source materialization (cast from the illegal type to the legal one) and target materialization (cast the other way) before and after. This keeps the rewrites in flight smaller, gets back to valid IR quicker, and makes the conversion pipeline more controllable and understandable. It has the [disadvantage](https://llvm.discourse.group/t/partial-lowering-with-type-conversions/1166/6) of inserting a significant amount of intermediate casts, but that should be fine for converting those relatively infrequent gpu ops.

Passes that make up the `gpu-to-tfrt-gpu` pipeline:
- `func-tfrt-streamify`: add chain and stream to func
- `async-tfrt-streamify`: convert `async.execute` to chain+event
- `gpu-tfrt-streamify`: convert gpu ops to chain+stream, synchronize streams
- `async-to-tfrt`: rewrite `async.execute` to `tfrt_test.do.async`

No more `tfrt.new_chain`: the chains are new correctly threaded through the stream synchronizations (previously a new chain was created in each `async.execute` region), which is necessary because `tfrt_test.do.async` is non-strict.

Only one `tfrt_gpu.stream.get_context`: getting the context is lifted to the function body (previsously there was one per `async.execute` region), which is sufficient because all streams use the same context.

PiperOrigin-RevId: 395861586
Change-Id: I74a1008fc251b892cc873d1f8a7d76864c2cd9aa