Refactor SavedModel SaverDef implementation so that the save/restore functions are generated in the Eager context.
This change prepares the SavedModel code to allow users to register advanced checkpoint saving/restore functions.

The SaverDef is a SavedModel field that allows TF2 SavedModels to be compatible with TF1. Before this CL, the SaverDef is generated by copying all objects and resource tensors from the eager context/default graph to the SavedModel exported graph.

The object-copying will not work well when we start introducing custom objects and classes, so this change defines the save and restore functions in same context that the user is saving in.

The extra arguments in `GraphView.frozen_saveables` are removed, and there is a slight reshuffling of code to make things the functionality partitions clearer.

NEW:
The traced save/restore functions now create a dummy SaveableObject to be saved into the checkpoint when there are no saveables to save or restore.

The test in the original CL failed because the TPU converter did not filter out the  traced_save and traced_restore functions. The filter checks for the SaveV2/RestoreV2 ops, which the test-exported SavedModel did not have because the model had no variables.

PiperOrigin-RevId: 395862677
Change-Id: I04e09f257d5a09af6d18720d0861edd06459e0e2