Fix error messages in variables.py
I've focused on the messages that are relevant to resource variables or migration rather than the 1.x-only symbols / reference variables.

Removes an error message for tf.Variable.__iter__. It now just reads from the variable and iterates over the result, which is what you'd expect it to do. I think that's better than trying to fix an error message that shouldn't exist in the first place.

Fixes #23185.

PiperOrigin-RevId: 395965288
Change-Id: Id5ff6ed2545c4ef5e67acccf3fcbc4094d289e4f