Add a new config-proto option that enables CPU global JIT (aka autoclustering).
Before this change, CPU code would not be auto-jit'ed (i.e. autoclustered and
sent to XLA) unless the JIT level was >= ON_1 *and* the envvar TF_XLA_FLAGS
contained --tf_xla_cpu_global_jit.

The logic behind this is, if you have a GPU net, you probably *don't* want CPU
autojit.  CPU autojit can break up GPU clusters by sending metadata ops (which
are just ops on int32 tensors) to the CPU, instead of computing them cheaply on
the GPU.

This is right and fine, but a problem arises if you have a single binary that
mixes CPU and GPU nets and wants to autojit both of them.  You don't want to
autojit the CPU part of GPU nets, but you do want to autojit the CPU part of
CPU nets!  But the single envvar gets in your way.

None of this is perfect, but to make it slightly better, we introduce a new
flag, cpu_global_jit, that acts equivalently to the envvar, but is per-session
rather than global.

Now you can set the flag just on the CPU models that you're interested in
autojit'ing.

PiperOrigin-RevId: 396697548
Change-Id: Ib3eec6378181542c8fc3dfbafcafb8e8d4e8b6b9