[tfrt:jit] Add reduction codegen to TF CPU RT pipeline.
Unfortunately, it is not possible to move mlir cmd line options to CompilationOptions, because there is not copy constructor for it. It is also not possible to introduce a dependency of TFRT on TF. Instead `TfCpuRtOptions` struct is added to TFRT repo and then reused in tf_cpurt_kernels. The options from tf_cpurt_kernels.h are hidden then in tf_cpurt_kernels.cc.

PiperOrigin-RevId: 397286733
Change-Id: Iaef4ea770a5b0c58c87f787370accca04ce54ed4