Improve random port generation in PickUnusedPortOrDie.
This reduces the chance of hitting 'Address already in use' errors in tests
such as c_api_distributed_test or cluster_coordinator_test. Particularly when
tests are run in parallel on a machine. There is stil a chance that the same
port might be chosen in separate test processes because the port is not
checked for availability and then bound atomically.