Fix crash in tf.gather gradient when indices rank was unknown.
A crash will still occur if the rank is unknown and batch_dims < 0, since this case is hard to fix, but a sensible error message was added.

This fixes a crash in sparse_softmax_cross_entropy_with_logits when determinism is enabled, as that function uses gather with batch_dims >= 0 when determinism is enabled.

PiperOrigin-RevId: 398762499
Change-Id: I149567e53a5fcfbc69ab4bba35a73ca0ae81e967