Refactoring: Move all metric-related code to a common directory
This makes the file structure cleaner as it collects all metric-related file under "lite/python/" and it makes more sense for "converter_error_data.proto" to stay under a directory named "metrics" than "metrics_wrapper".

Besides, at the moment, if we want to add a new metric-related file, we have to choose between adding it to "metrics_wrapper" (which might not match the purpose of the new file) and "lite/python/" (which is too general and already crowded). Having a "metrics" directory will fill that gap.

PiperOrigin-RevId: 399091254
Change-Id: I789eca7d8699b2c786b6c4b6e8f2c6daf8f9a369