In make_deterministic, rewrite ops less often.
Some random image ops were added to kDeterministicStatefulOpsWhenAsync, so that asynchronous ops are allowed to have those image ops without being rewritten.

Also, the function of If and While ops are now recursively checked in FunctionMayIntroduceNondeterminism(). The effect, in practice, is that if a ParallelPrefetch op has a function containing If and While ops, the If and While ops can have deterministic stateful dataset ops like TextLineDataset without the ParallelPrefetch being rewritten. Before, If and While ops were recursively checked in the IsNodeStateful function, but IsNodeStateful did not take into account the existence of deterministic stateful ops.

PiperOrigin-RevId: 399299268
Change-Id: I829690c2ca304abeae62eaa20d56d094d8859518