Fix tfl-to-tosa quantization intermediate type and fully-connected shape
Intermediate quantized type should be an I32 under all circumstances however
the lowering assumed the intermediate type was the same as the output type.
This recently appeared due to some changes in the linalg compilation that found
illegal type conversion.

This also includes adding a reshape post fully-connected. Cases where the input
was reshaped sometimes resulted in the wrong replacement type in the model.

PiperOrigin-RevId: 399970010
Change-Id: I02d285d0d77562fede4e79977b5841e53cead71e