Don't accidentally upcast to double precision in CropAndResizeOp.
The C/C++ constant "0.5" is a double.  Multiplying by this implicitly upcasts
the remainder of the expression to double precision.

Don't do this.  This helps particularly on nvidia GPUs that don't have lots of
fp64 units.

In my benchmark on an RTX2080, this takes the kernel from 1.15ms to 660us, for
a ~2x speedup.

PiperOrigin-RevId: 401003860
Change-Id: I7752f3a5efe4abd8193d0d24d98e6ba74978a1cb