Support different input, bias, and output types for fused convolutions.
XLA's convolution integration expects to be able to specify the three types
independently, but the initial cuDNN frontend TensorFlow support expected to
decide them internally.

A note on naming: when only one scalar type is involved, we call it
`element_type`: it's the type of the elements of all the relevant tensors.
When we have separate types for different tensors, we'll avoid using
`element_type` at all, and name types according to the particular tensor
they're related to.

PiperOrigin-RevId: 401343329
Change-Id: Idb9b16b4bddbe020eda9a7abe563430336777327