Updates config=mkl_aarch64 build to oneDNN 2.4 and Compute Library 21.08
Updates oneDNN version used for `--config=mkl_aarch64` to 2.4
in line with x86 build.

Updates Compute Library (used as optimised oneDNN backend
on AArch64) to the latest release (21.08).

This enables support for SVE and BF16 kernels where
hardware supports.
This provides a performance uplift of between 1.4x and 3x
depending on the hardware support.

Removes some patches to oneDNN which are not required for
the 2.4 release.

NOTE: AArch64 build with --config=mkl_aarch64 requires GCC 10.

Change-Id: Iaf7562da729fc928dce922d33c02b8bb0ce74767