[XLA:Python] Add an explicit xla_client.pyi interface file.
This works around a bug in pytype where modules that are re-exported often have the types of their contents replaced with Any, and makes the errors from pytype for xla_client much more useful.

Fix a number of incorrect and missing type annotations, and make a few small cleanups in passing.

PiperOrigin-RevId: 403257225
Change-Id: Ic570f8277691fbdc79ab8d9e516187743b037728