Always return new group key from CollectiveKeys.get_group_key
We only get the group key once for the same strategy in practice. So we don't need to maintain a device set to group key mapping.

For different
strategies it's actually better to use different group keys for isolation. The
isolation is now achieved by clearing the group information in
update_server_def(), but this is subject to deadlock.

PiperOrigin-RevId: 404172644
Change-Id: I461da09c5c2ceb7a5faefe35e356213cbf46ba1f