Add _enable_new_dynamic_range_quantizer flag to control post-training dynamic range quantizer.
The default value of _experimental_new_dynamic_range_quantizer is false which
means MLIR quantizer will fallback to the old TOCO quantizer for dynamic range
quantization. If the value is set to true, MLIR quantizer will trigger the new
dynamic range quantizer in MLIR instead of the old one.

This flag is used only if experimental_new_quantizer flag is on.

example)
 converter = tf.lite.TFLiteConverter.from_saved_model(saved_model_dir)
 converter.optimizations = [lite.Optimize.DEFAULT]
 converter._experimental_new_dynamic_range_quantizer = True
 quantized_tflite_model = converter.convert()

PiperOrigin-RevId: 407515014
Change-Id: I8d61f55a016af6c0ac3c1596976fefa073102c9a