[tfrt:jit] Split the current output fusion into two parts.
The first part occurs when tiling the reduction. As a result we have two output arguments in the tiled_loop: one provides the initial values for the reduction result and the other is an init_tensor operation that will be used to initialize the temporary result (reduction of a single tile).

The second part happens after the padding/peeling. It sinks the init_tensor introduced in the first stage into the loop body and gets rid of the output operands for the intermediate results.

PiperOrigin-RevId: 408583476
Change-Id: Ifab0b25a36d646adabd472b91eeeacf8356a667a