Keras Metric.
Implement class `OneHotIoU` that inherits class `IoU`.

This class can be used to compute the IoU metric for a multi-class classification task where both the labels and predictions are in the one-hot encoding format.

This class will first convert both the labels and predictions into the all-class format, then applies the computation steps of class `IoU`.

PiperOrigin-RevId: 408709708
Change-Id: I43751791ab2efa04a4a25cb7701975dac1154fcc