Run canonicalizer before decomposing resource ops in cluster.
Running canonicalizer cleans up any dead ops in cluster so that resource ops decomposition does not have to spend time cleaning up dead ops. This also reduces the maximum number of iterations required to decompose resource ops in cluster which was set to 1000 earlier to accommodate for cleaning up chain of dead ops.

PiperOrigin-RevId: 410568001
Change-Id: I20c370b3234e06a01f5b90c949cfa7ed545abbb4