Fix XLA on cuDNN pre-8.1: missed a place that needed to fall back to legacy.
Since cuDNN didn't (sufficiently) support the frontend APIs as used by XLA
until 8.1, attempting to use it on older versions was an error.  A recent
change flipped the default of an XLA flag to enable cuDNN frontend, and this
was honoured regardless of the current cuDNN version.  Since the recent API
changes were carefully made to allow switching seamlessly between the two, we
can just second-guess that flag when cuDNN frontend is unavailable, and use
legacy APIs anyway.

PiperOrigin-RevId: 410942478
Change-Id: I95f34a06e55341c6538052c6ea39bd0de9c19bdb