When small function optimizations are enabled, allow some multiple-subgraph partitioned graphs to run synchronously by sorting subgraphs topologically.
Recent changes allowed sync execution for the 1-subgraph case; this extends it to many N-subgraph cases by reordering them so the non-blocking call ops are first.  Where the prior changes essentially enabled this* for CPU, this allows many e.g. GPU cases to be run with sync execution as well.

Specifically, in current usage, sends are non-blocking and receives are blocking.  This change allows synchronous execution of graph partitionings where a subgraph contain _Send/_HostSend ops, or _Recv/_HostRecv ops, but not both.

There is no restriction on the number of send or receive ops, and there is no restriction on the number of subgraphs in the partition. Though the number of subgraphs is generally at most 2, e.g. one for CPU and one for GPU, as they are indexed by device name.

(* again, this is all effective only when the small function optimizations flag is set and synchronous execution is requested within the runtime)

PiperOrigin-RevId: 412162117
Change-Id: Ie54494540ed2103274cf230b71dc958be7f87819