Add HLO -> MHLO importer for AllToAll HLO.
AllToAll HLO instruction can either have pre-split operands (and return a tuple) or a single operand that all-to-all splits across `split_dimension` into the number of replicas in a group. Only the latter case (array all-to-all) is supported in importer right now and the former (tuple all-to-all) is not supported yet because mhlo::AllToAllOp does not allow tuple output or variadic operands.

PiperOrigin-RevId: 413257417
Change-Id: Ibfeed33903017274c433781a7860edc6e4f84443