[tfrt:jit] Remove padding pass from TF CPURT.
There are several reasons to do that.
1. It is currently not used.
2. It does not improve the benchmarks if we disable peeling and in some cases
makes them worse.
3. It is not clear what padding value to select if we are doing padding of the
whole body of `linalg.tiled_loop`. For example, if we are padding a
`linalg.generic` that models a reduction, we might extract the padding value
from the `linalg.fill` that produces the output tensor for the reduction. For
other cases, we would need to write a giant switch-statement that would return
a padding value depending on what ops we have in `linalg.generic`. There is
another approach to that, which is being explored by gysit and nicolasvasilache: tile and pad a
single linalg.generic:

%input = input_op %input_of_input
loop
  %slice = extract_slice %input
  %pad = pad %slice %some_value_depending_on_the_generic_below
  %generic = linalg.generic %pad

and then fuse the input.

loop
  %slice = extract_slice %input_of_input
  %input = input_op %slice
  %pad = pad %slice with %some_value_depending_on_the_generic_below
  %generic = linalg.generic %pad

after that we have to move `pad` before `%input`:

loop
  %slice = extract_slice %input_of_input
  %pad = pad %slice with %previous_value_adjusted_depending_on_input_op
  %input = input_op %pad
  %generic = linalg.generic %pad

PiperOrigin-RevId: 414711137
Change-Id: Iaa77e22efe044f8f405f53b11ca24c8c74efdf90