Use global broadcast propagation pass.
This pass cannot replace yet the EarlyBroadcastInDimOp pattern because it does not allow to move elementwise ops into blocks (which is needed to be able to merge assuming ops). But in some cases, it already avoids the exponential time complexity
due to the EarlyBroadcastInDimOp pattern.
However we need to remove this pattern from the SymbolicShapeOptimization pass,
because that pass needs to run before BroadcastPropagation pass.
Re-enable a few broadcasting tests which are now fast enough.

PiperOrigin-RevId: 417372847
Change-Id: I93b510cf312847efadedaaa5114ee1c5a88aa6ec