Add missing dialect dependency on xla-legalize-tf-control-flow
Add missing dialect dependency on xla-legalize-tf-control-flow. This
fixes crashes when using tf-opt to lower tf.while to mhlo.while. The
reason the current test cases for the above pass don't crash is that
those test cases already have mhlo ops and so the dialect is loaded at
parse time hiding this issue.