[jax] Add default_device config to JitState
This can be used to override the system default device when running a
jit'd function. This is the prioritized list which determines which
settings take precedence when determining which device a jit'd
function runs on:

1. `device` or `backend` argument to jit
2. Input(s) with sticky device
3. default_device config (new!)
4. System default device

This also changes the `jit` C++ binding to take a `jit_device`
argument, which is the device determined by the `device` or `backend`
argument to jit (or None if not set), rather than determining it by
calling back into Python.