Squash REQUIRED_PACKAGES logic
This cleanup should be a no-op. I moved a bunch of package-selection logic into
the list of REQUIRED_PACKAGES so that all dependencies are specified in just one
place. For example, one big difference is that the TF ecosystem packages are
right next to each other.

I also know that gast 0.5.2 breaks one of our tests, so I've pinned the gast
version to 0.4.0 or below, which is what our CI currently installs for testing.

PiperOrigin-RevId: 419650242
Change-Id: I30a70e6a54f89e55e117d8484c84b2b39709b711