Add caution to math.segment_* documentation about bad segment_ids on GPU
In `tf.math`, `segment_max`, `segment_mean`, `segment_min`, `segment_prod` and `unsorted_segment_sum` all expect that `segments_ids` are sorted in increasing order. On CPU, these are validated, and an error is thrown for out-of-order indices. On Gpu, these are NOT validated, and out-of-order indices do not throw an error but instead result in safe but unspecified behavior.

PiperOrigin-RevId: 419867003
Change-Id: I95209a3367d0139f3384052b881066e864dfbf78