Update pybind11 to 2.9.0.
The proximate reason for the upgrade is to pick up a thread-safety fix (https://github.com/pybind/pybind11/pull/3237) for JAX, but we might as well just upgrade to the current version for everyone.

Remove a backport of a patch now present in the released version (https://github.com/pybind/pybind11/pull/2864).

Patch out the "VERSION" file from nsync to avoid conflicts with the <version> C++ header included by pybind11.

PiperOrigin-RevId: 420856622
Change-Id: Ied007ed62b93d96cf008d6ec6336017af1fa10fd