restore tf.ResizeBilinear from mhlo.convolution with non-all-1s lhs dilation
The existing pattern legalizes mhlo.convolution with non-all-ls lhs dilation to
tf.Conv2DBackpropInputOp with VALID and SAME paddings. If the mhlo.convolution
is converted from TF image resize ops (tf.ResizeBilinear, etc.), it usually has
different padding sizes which couldn't be matched (refer to the new unit test).
Besides, the tf.Conv2DBackpropInputOp usually doesn't have good support in some
of the inference backends, so tf.Resizebilinear is perferred when the padding
matches.

This patch added a new speical case to the pattern to match the
mhlo.convolution, which is legalized from tf.ResizeBilinear op. This change
improves the coverage of exsiting legalization.

PiperOrigin-RevId: 422614574
Change-Id: I4158767f2e9804e27d84e0e1e95ede60239d6a1e