Enable Gather Op Unit Test for Dynamic Range Quantization
Gather Op was supported by the existing MLIR dynamic range quantizer
but it was not able to pass the unit test from the old quantizer due
to the two reasons:
1) If the weights of the gather op has min/max quantization params,
the converter converts it not to Const but to Stats op.
2) The optimization op order pass is missing where it reorders the
dequantize op before ops with same scale requirements when memory
usage reduction is possible.

Therefore, this CL changed the test binary file not to contain
quantization parameters and also to increase weight size so that
it has memory usage reduction.

PiperOrigin-RevId: 422723291
Change-Id: I0a5fdd62dedaf631c588683d579987b75c7ee94a