Allow reopening tf_record_iterator() instances at prior read offset
This adds a `.reopen()` method to iterator instances returned from `tf.compat.v1.io.tf_record_iterator()`, which can be used on a closed iterator to re-open the originally provided filename, and resume reading from the file at the prior read offset.

This behavior is needed in order to handle the case where the tfrecord file in question has been replaced by a new version of the file with additional data appended, and the filesystem implementation is such that the file object opened internally by the iterator will not reflect any data from the new version of the file (e.g. it has an open handle to that specific version of the file).

By allowing the iterator to be closed and reopened, callers who detect this case (e.g. via a stat() call that shows the file has increased in size) can now continue reading, without having to directly expose and manage read offsets or deal with the potential quadratic overhead of re-reading the file from the beginning.

This behavior is intended primarily for use by TensorBoard.

PiperOrigin-RevId: 424452732
Change-Id: I4b0d1539b9e9ed5eecd7c210c411bf781a03021d