[tfrt:jit] Store sequential CompilationUnitAttr id in the Bef file
When compiled module emitted in the Bef attributes section, it is emitted for every call site, which means that if we use address as an `id`, we do not really benefit from any caching.

As a workaround rely on a sequential id generation during the mlir-to-bef translation. However it means that users must be careful not to share caches between Bef files.

PiperOrigin-RevId: 429623608
Change-Id: Iea5589903a7c00cfbcffdd4dbc734afae83ddf5d