Avoid cached definition and signature in __init__ of eagerly created function
When a concrete function is created but never called from Python, this change
saves us an extra copy of fdef. This tends to happen frequenctly in functions
loaded from a saved model, where only the top-level 'inference' concrete
functions are directly called from Python, whilst other supporting functions
are called as part of the graph execution.

PiperOrigin-RevId: 431814317