[mlir/tfg] Avoid name collision for identity node in the const dedupe hoist pass.
Adding an Identity node with the same name as a constant makes the MLIR TFG export to the GraphDef fail.
Let the name have `const_dedupe_hoist/` prefix followed by the const name. There is still a chance that this will lead to collision, but the surface is very limited.
It might worth to consider better unique name generation to avoid collisions, I left a TODO to address this.

PiperOrigin-RevId: 432489771