[XLA] Fix the bug in Literal storage: make it consistent with on-device buffer representation
Otherwise, copies using TransferManager may result in OOB read/writes. As a
bonus, the resulting code is also considerably simpler.

We have to over-allocate Literals (always allocate the "margin" for the dynamic
sizes), but given that this allocation is tiny and is not used on
performance-critical path, this should not be an issue.

PiperOrigin-RevId: 433393410