Fix tflite_flex_android_library for inclusion in android_binary
A previous change introduces a fix to disable layering checks when building for
Android or iOS to resolve an issue with iOS projects.

But this causes usages of tflite_flex_android_library to fail when included
with an android_binary, with the complaint about layering being disabled.

Changing the dependency selector to only disable layering for iOS build seems
to address the issue.

PiperOrigin-RevId: 433878104