[tf2xla] Don't cluster Fill ops that aren't used inside an XLA cluster.
There's no benefit to asking XLA to compile a Fill op that isn't used in the
cluster; this is just using XLA to call cudaMemset and return the result.
There is, however, a harm: Adding the Fill to the XLA cluster can significantly
increase its live range.  If the tensor is large, this can blow up a graph's
memory usage.

PiperOrigin-RevId: 435384722