Add xla_compile_device_type option to KernelAndDeviceFunc.
This option will annotate the nodes in the function with `_xla_compile_device_type` string attribute with the value of the `xla_compile_device_type` option.  This will used by the bridge to compile the function/ops for the specified device_type.

This is currently controlled by an EagerContext option which is disabled by default.

PiperOrigin-RevId: 436274145