[PJRT:CPU] Ensure TfrtCpuBuffer is defined and ready before letting caller acquire an ExternalReference for zero-copy raw pointer read.
TFRT CPU backend in theory allows CPU computations to be dispatched asynchronously, yet still allow PJRT callers to acquire an ExternalReference and read the buffer contents in a zero-copy way (see PyBuffer::AsNumpyArray implementation).

This CL fixes a bug where we did not synchronize and wait for the buffer to be ready before giving PJRT users a ExternalReference and allowing them to read via raw pointer.

This problem is however masked by a different bug in async dispatch. Due to a bug in retrieving flop count from HloCostAnalysis, all CPU computations are currently dispatched synchronously. This meant that the buffer content is always ready, and no users at HEAD is affected by the lack of synchronization that this CL fixes.

However, we do intend to allow async dispatch, in which case, the synchronization before handing out an ExternalReference is necessary.

PiperOrigin-RevId: 436327791