Copy TPU device propagation logic in distributed TPU rewrite pass to placer, without _Retval in the target op list.
MLIR bridge does not operate on V1 control flow ops hence we cannot implement the same logic in MLIR bridge. Moving the logic to placer can cover both the old bridge and MLIR bridge.

The logic may apply to GPU as well thus we rename the function to `PropagateDevices`. The function is still a NoOp for a graph without any TPU devices though. We can easily add GPU to `IsPropagatableDevice` later.

PiperOrigin-RevId: 439676624