call_with_layout shall pin to the mesh of the layout.
... instead of using the default mesh. This is only applicable in eager mode.

This means call_with_layout always follow the mesh of the output, and likely
implies the inputs must be explicitly copied to the mesh.

This model is easier to reason than the default mesh logic, and likely will not
cause un-avoidable communication overhead.

Fix reentrance of dtensor.run_on(). Before this CL with dtensor.run_on() always
restore to the global default mesh, which is undesirable if there are nested
run_on ops. (e.g. call_with_layout inside a default mesh scope).

PiperOrigin-RevId: 439700696