Increase tolerance for self_adjoint_eig_op_test.
The tolerance on self_adjoint_eig_op_test seems a bit tight. The test is currently failing
on aarch64 (#52544).

Playing around with small perturbations of the inputs and step size `delta` on x86_64,
the max error seems to be in the range 0.008-0.016. Increasing the test tolerance therefore
seems reasonable to account for this error range.

Fixes #52544.

PiperOrigin-RevId: 439758034