Copy TPU device propagation logic to **LowerFunctionalOpsPass**, but only operate on newly created nodes in the pass to avoid side-effects.
MLIR bridge does not operate on V1 control flow ops hence we cannot implement the same logic in MLIR bridge. Moving the logic to LowerFunctionalOpsPass can cover both the old bridge and MLIR bridge.

The logic may apply to GPU as well thus we rename the function to `PropagateDevices`. The function is still a NoOp for a graph without any TPU devices though. We can easily add GPU to `IsPropagatableDevice` later.

PiperOrigin-RevId: 440242573