Fix error calling tf.function with distinct-mesh small constant.
The error occurs due to constant folding merging the tf.Const Op
connecting to 2 distinct-mesh DTensorLayout ops to a same ConstOp, failing
at mesh propagation.

This really is a temporary solution. I think a full DTensor Dialect
(where DTensorLayout is not modeled as a separate Op, but directly on the
Value, if that's allowed by MLIR) to disable such constant folding/merging
sounds like a more correct solution.

PiperOrigin-RevId: 440956679