Optimize attr lookup by using `AttrSlice::FindByString()` when possible.
`AttrSlice::Find()` performs a linear scan of the attrs and can't be changed until `protobuf::Map` supports efficient lookups using a `StringPiece`.  But in many cases, `AttrSlice::Find()` is called with a `std::string`.  Switching to `AttrSlice::FindByString()`, which uses an O(1) map lookup, improves performance when there are many attrs (e.g. some `CsvDataset`s, which have an attr for every column and don't scale well to CSVs with O(100k) columns).

PiperOrigin-RevId: 441823425