Update pybind11 to 2.8.1.
The reason for the downgrade is that v2.9.x is causing a segmentation fault in the TF Lite interpreter in TF Text conversion tests.

The original reason for the 2.9.0 upgrade was to include a thread-safety fix (https://github.com/pybind/pybind11/pull/3237, https://github.com/pybind/pybind11/commit/0e599589fe821f86d18635c13636f3042d4c06b9) for JAX that was added in 2.8.0 and above, so performing this downgrade is not problematic for its intention.

PiperOrigin-RevId: 442586909