Updating tensorflow/python/eager:gradient_input_output_exclusions, splitting the main routine out from the code itself and adding a py_binary target.
It needs to be both a library and a binary. Previously it was one target; I changed it to a library to avoid the warning when linking into the test code, but wasn't aware of its usage as a binary to generate the pywrap exclusions file in this directory.

We could link in the python source twice -- into the library and binary -- without splitting. But by splitting the main out into a new .py source file , both the test and binary can now depend on the generator code within the py_library.  (Which seems the most correct end result, I believe.)

PiperOrigin-RevId: 443151843