Remove chain input and chain output from tfrt.case ODS.
The special chain input and chain output for tfrt.case operation aren't needed
as the branch functions have a chain input and a chain output. This change
removes the two fields from ODS and modifies tf-to-tfrt to use the chains from
the branch functions arguments and return values instead.

Modify relevant tests.

PiperOrigin-RevId: 443213704