Add initial support for int4 to MHLO
This CL adds very simple support for int4 to the MHLO dialect.

It doesn't worry too much about integrations (with frameworks or compilers)
or semantics (e.g. verifiers is a known improvement area for the dialect)
and simply aims to expand the representation power of MHLO to unblock experiments.

Figuring out integrations and semantics is also on the roadmap, but there is
consensus that we should unblock forward progress right away and figure out
the rest in parallel.

PiperOrigin-RevId: 445198171