[grappler][tfg] Swap to the new importer/exporter
This CL enabes the new TFG importer/exporter in Grappler. The importer is faster
(when converting to and from GraphDef) and supports multi-threading (when enabled
in the TFG optimizer).

This CL also makes two small changes to the TFG optimizer hook: the function
library is not copied into the output GraphDef and a function library is
default-instantiated in case there were no exported functions.

PiperOrigin-RevId: 446083421