Remove invalid pointer arithmetic from `SplitPath`.
`SplitPath` calls `ParseURI` to extract scheme, host, path triples. Then if `path` does not contain the filesystem separator, the existing implementation uses `host.end() - uri.begin()` to determine the length of the basename component of the return.

This pointer arithmetic works as long as `host` is valid within `uri` (that is, starts at least at the same address as `uri`). In the cases where `host` is invalid, the difference would be a negative number.

If this difference is a negative number, `StringPiece(addr, negative_number)` would result in a large memory allocation.

This bug has been silent in TF for years, since the code was written. This is because code uses `StringPiece(remaining.begin(), 0)` to build an empty `StringPiece` (which is a pointer within the same buffer as `uri` above but for a span of 0 characters; since it points within the buffer there is no potential for negative results).

However, once `StringPiece` (aka `absl::string_view`) is used in C++17 this can no longer work (and we have a subsequent change that replaces this with `StringPiece()` to build empty string views). This is because `absl::string_view` is aliased to `std::string_view` in C++17 and it is an error to do `std::string_view(pointer, 0)`.

PiperOrigin-RevId: 447870033