Migrate away from InferTensorType in HLO_CompatibleOperandsAndResultType
Existing infrastructure for Python bindings, combined with the existing type inference infrastructure provided by HLO_CompatibleOperandsAndResultType, makes it tricky for JAX/MHLO to create sparse programs:
  1) mhlo.FooOp.__init__ doesn't provide away to explicitly specify the result type, inferring it via _ods_ir.InferTypeOpInterface(FooOp).inferReturnTypes.
  2) mhlo.FooOp.build_generic allows to explicitly specify the result type, but also requires to specify the result type, which is a significant hit to ergonomics.

To unblock sparse work, we decided to make use of the current convention that inferReturnTypes infers the result type using the type of the first operand. Unfortunately, inferReturnTypes implemented by InferTensorType drops sparsity encodings, so we're migrating away from InferTensorType in favor of our own implementation.

In the long-term, we are planning to come up with a more principled solution that: a) allows to explicitly specify quantization and sparsity aspects of result types, b) doesn't impose a significant ergonomic hit.

PiperOrigin-RevId: 448545329