[XLA:SPMD] Improve propagation of sharding through reshape when sharded trivial dimensions are removed.
Currently we just return nullopt and propagate full replication, but instead we could move that
to partially replicated improving a bit usability when such a sharding is encountered.
Example of when this is encountered:

v = ..., sharding={replicated}
operand = f32[1,4,128] reshape(v), sharding={devices[2,1,16]...}
f32[2,4,128] dynamic-update-slice(base, operand,...indices...), sharding={devices[2,1,16]...}

When propagating bottom up we pass the dynamic-update-slice sharding to the operand, but because
the dimension on the operand is 1 when we remove the dimension from the reshape it doesn't propagate through.

PiperOrigin-RevId: 449643699