[tfg] Implement MemoryEffectOpInterface for TFG ops.
`MemoryEffectOpInterface` is a key core MLIR interface used in upstream passes,
such as CSE. This CL implements the interface generically for TFG ops as a
dialect interface.

The modeling is: if a TFG op is stateful, then its memory effects are that it
writes to the default resource (`MemoryWrite` in tablegen). Otherwise, it has
no side effects.

PiperOrigin-RevId: 449947852