Fix legalization of SplitV by handling negative axis values
The legalization for SplitV would not take negative values as axis at
the moment. Because no error is thrown, the rewrite will just discard
the operator resulting in a wrong MLIR.

This patch fixes the issue by adjusting the axis with a wrap around
approach. Furthermore, if the axis value exceeds the size of the input
tensor, an error is thrown so that the legalization does not take place
allowing other rewrites to be performed.

Signed-off-by: Michele Di Giorgio <michele.digiorgio@arm.com>
Change-Id: I2b873ddce9685b9105a142c20fc032afc9863bf3