Fix tf.image.non_max_suppression for half and mixed types.
The original implementation failed if the threshold type did not
match the input box/score type.  Internally, the op assumed they
were the same, which resulted in a bad memory access.

Unfortunately, we can't just check to ensure they are the same type
and fail otherwise.  The python op passes in default float values,
and we don't necessarily know the input type ahead-of-time to ensure
the types match.

The work-around is to internally parse threshold values based
on their original type, and then explicitly cast to the box/score
type.

Fixes #54264

PiperOrigin-RevId: 449991509