Use partial content of tensor as part of key to cache primitive
Since we cannot update tensors in-place when using Arm Compute Library once the Mkl primitive is created and to get benefit of reusing the primitive when it is called again as part of hash function for caching the primitive we are using memory address where the tensors whose content can be updated is stored. Unfortunately, when a new primitive needs to be created where the only difference is content of tensors that is stored at the same memory address (like in tests tensorflow/core/kernels:matmul_op_test_cpu and tensorflow/core/kernels/conv_ops_test_cpu or in running BERT) we will retrieve from the cache wrong primitive with stale tensors.

In order to avoid this problem this patch takes hash of partial content of tensor and use that as part of key for caching primitives. Once Arm Compute Library enables in-place updates (expected in 22.08 release) this patch will be removed.