[mhlo:linalg] Precisely match the documentation for mhlo.clamp
This doesn't make a difference if the upper bound is greater or equal to the
lower bound, but there is no verification of that. The documentation specifies
`clamp(a, x, b) = min(max(a, x), b)` so make sure we match that precisely.

This actually shows up when comparing numpy clip to mhlo.clamp with random
inputs. This diverges from TF behavior, which implements it the other way round
but explicitly leaves the behavior undefined when the bounds are mismatched.

PiperOrigin-RevId: 450509463