`mhlo.map`: allow heterogeneous operands
HLO Map to takes inputs of different element types per the documented op semantics: https://www.tensorflow.org/xla/operation_semantics#map. In practice XLA did not accept map ops with heterogeneous inputs in the past but as of PiperOrigin-RevId: 443525098, XLA allows this more flexible behavior to align more closely with the spec.

This brings mhlo in alignment with that change to HLO.

PiperOrigin-RevId: 451481051