#tf-data-service Fix flaky multi-process tests.
Error message:

```
Failed to perform worker heartbeat: Failed to assign an index for
worker localhost:38365. Configured workers list:
[localhost:39135, localhost:42759, localhost:45653, localhost:38831,
 localhost:36563].
```

The issue is the dispatcher may use the same port across the tests.
If a worker from the previous test is still running (due to multi-
processing), it may try to register itself to the dispatcher running in
the current test, and the registration will fail.

The fix is to use different ports for each dispatcher/worker. This way,
the workers won't register with the wrong dispatcher.

PiperOrigin-RevId: 451553309