[XLA:GPU] Decompose bitcast inside fusion into bitcast+transpose.
Within XLA:GPU fusions, bitcast ops can be either fast (ReshapeIsBitcast is
true) or slow (ReshapeIsBitcast is false).

Previously we tried to avoid creating slow ones by telling algsimp to avoid
translating transpose ops to bitcasts (which generally do not have
ReshapeIsBitcast true).

The problem with this is that if you have a transpose+bitcast that *doesn't*
end up as part of a larger fusion, this approach ends up creating a fusion just
for the transpose+bitcast (tantamount to a memcpy).  Whereas if we *had*
converted the transpose to a bitcast, then we'd have had bitcast+bitcast =>
bitcast, and, if that bitcast is outside of a fusion node, it's free.

This patch takes us in a different direction.  Now we always transform
transpose to bitcast where possible.  But then right before codegen, we find
all the bitcast ops inside fusion nodes and split them up into
transpose+reshape-is-bitcast.

We prove empirically that we can *always* split a bitcast into
transpose+reshape-is-bitcast, so this shouldn't make any fusions slower than
they were.

We leave removing the don't-rewrite-transposes-into-bitcasts feature in algsimp
for a later patch, after this one sticks.

PiperOrigin-RevId: 452191666