#tf-data-service `MultiProcessCluster` supports tsan.
This is explained in https://github.com/tensorflow/tensorflow/blob/5e6618333778a1cb7944cfe691b96caba9572ce3/tensorflow/python/distribute/multi_process_runner.py#L574:

`terminate` uses SIGTERM to terminate the subprocesses, in order to log
the stack traces of the subprocesses when they exit. However, this
results in timeout when the test runs with tsan (thread sanitizer); if
tsan is being run on the test targets that rely on timeout to assert
information, the process must be killed before the test exits, so the
subprocesses are terminated with SIGKILL, and data race is removed.

PiperOrigin-RevId: 453213487