Avoid segfault when init_value is not on default_mesh.
To actually fix the segfault in lower level (e.g. directly users of VarHandleOp),
I tried to add a validation in SPMD of AssignValueOp, but turns out it only
knows the resource_layout is an 'empty' layout without any mesh information.

So there is not enough information to compare if the mesh is correct. We
shall start tracking mesh of empty layout -- but changing the data model at
this point is not very easy to do or to justify.

PiperOrigin-RevId: 453453817