Allow TF quantizer's calibration step to work in graph mode.
TensorFlow quantizer's calibration step currently misbehaves when running in graph mode.
When running in graph mode, without `sess.run()` it will not execute any inference through the graph,
eventually leaving none of the `CustomAggregator` populated with the desired statistics (min and max values).
This results in the model not being quantized at all.

PiperOrigin-RevId: 454157742