Add monotonic timing functions to tensorflow/core.
TF core environment provides Now...() time functions (e.g. NowNanos(), NowMicros(), etc.) which returns the real time clock. This function is not guaranteed to be monotonically increasing, which is not ideal for measuring the time difference for profiling or benchmark. This CL introduces Monotonic...() time functions which don't return a meaningful time value but are guaranteed to be monotonically increasing and thus ideal for profiling.

PiperOrigin-RevId: 454277863