Improve `SlowOperationAlarm` to log the total duration on object destruction.
Right now, the object prints an error at soon as the timeout if finished, but the user never get to see the full-duration of the operation (it just knows it's above the timeout).

Thus, in addition to the alarm, we also log, if the operation took more than `timeout`, the total duration of the operation. This helps users identify the slowest parts of their compilation.

We also change the optional argument for `SlowCompilationAlarm` which used to be an optional name, into a mandatory argument used to define the context of the operation.
Instead of being appended to the warning message, it will be prepended.

PiperOrigin-RevId: 454857921