Fortify access to ScopedTFGraph with a context manager.
Such that the backing C object it is never destroyed in the middle of a
python function that uses it. It is unclear which functions that uses _c_graph can be called on the problematic 'use-during-deletion' path, thus I changed them all to use the fortified .get() API mechanically.

While at it, I added a tensor_util API to reduce number of protected access violations caused by TF_TryEvaluateConstant. It is unclear how the C++ evaluate constant relates to the _ConstantValue implemented in tensor_util.py. At least the names are different, and they are now in the same file, hopefully making a future consolidation easier.

PiperOrigin-RevId: 455227569