Allow an iterable for the representative dataset instead of a callable.
Representative dataset has been a callable that returns an iterator of representative samples.
Since there are no direct advantages of it being a caller and only complicates the format of the representative dataset,
directly accept an iterator instead for the quantizer.

The signature of the representative dataset changes from:

```
Callable[[], Iterable[RepresentativeSample]]
```

to:

```
Iterable[RepresentativeSample]
```

By doing so, any form of iterable is supported as a representative dataset, opening up many options for the users to use not only generators but also lists and `tf.data.Dataset` as the representative dataset.

PiperOrigin-RevId: 455498896