[GmlSt] Change syntax of SubsetYieldOp and ParallelOp yet again.
The changes are caused by the way the current in-place analysis works in
bufferization. If we don't have dedicated bbArgs that correspond to `outs`
arguments, then the `state.isValueRead` gets into infinite loop when analysing
the users of the output. Also, the terminator did not have `destination`
explicitly, because of that it was impossible to determine if we write to the
value.

PiperOrigin-RevId: 457098545