Use smaller number of representative samples in test cases.
Although the recommended number of samples in a representative dataset is more than 200,
there is no need for using such large number of samples in test cases.
It has been a source of slowdown and even timeouts (which led us to shard the test), especially for parameterized test cases
where multiple representative datasets are created and tested.

Unless we're explicitly seeking for behavior differences with larger number of samples, we can use a much smaller number.

PiperOrigin-RevId: 457890364