Improve side effect modeling for CollectiveReduceV2
Previously, the op did not have a specific effect attached which meant that it
was treated conservatively as "unknown side-effecting", leading to dependencies to/from all other side-effecting ops. Now, different instances of the op are kept in order but no dependencies to/from other side-effecting ops (except for unknown side-effecting) are created.
We don't rely on the existing `ordering_token` resource because that is currently not always set (DTensor).

PiperOrigin-RevId: 459358521