Bump JAX libs for CI
JAX libs are needed in CI as part of testing JAX <-> TFLite integration. This is only needed on Linux and Mac, as the corresponding tests are not run on Windows at this time.

There is no need for an equivalent change in `setup.py` as TF wheel does not depend on JAX. This is just a test-only dependency