[XLA:GPU Next] Use heuristics to pick tile size and unroll factors instead of hardcoding them.
Vectorize to 128 bytes per thread, try to find a block size that does not need a loop tail.

This is implemented by building a pass pipeline for each fusion op. We tried to avoid this in the past for performance reasons, but the advantage of this change is that we don't create the pipeline if no fusion op is rewritable. We could have both if we made the pass use a callback to retrieve tile/unroll sizes, but it's not clear whether it's worth the complication.

PiperOrigin-RevId: 460538701