[XLA:GPU Next] Make unrolled loop use static bounds if possible.
The `tile-loops` pass rewrites an `scf.parallel` into three nested ones. If the step size evenly divides a parent loop range, the child loop does not need to be clamped by `memref.min`.

This change allows for the inner loop to use this optimization even if the middle loop can't.

PiperOrigin-RevId: 460547881