Support MHLO Legalization for CustomCallWithComputation.
This CL refactors the existing MlirHloBuilder::CustomCall overload containing the XlaComputation& arg to now fully support called computations.

Since CustomCall (with computation) requires adding a new function to the encapsulating module, this op cannot be built during a Function pass, since that pass type is not allowed to modify anything outside the current op for thread safety reasons (https://mlir.llvm.org/docs/PassManagement/#operation-pass:~:text=Must%20not%20modify%20any%20state%20referenced%20or%20relied%20upon%20outside%20the%20current%20operation%20being%20operated%20on), so the builder now contains a new instance member indicating whether or not function creation is allowed. This will be utilized by a new LegalizeTFModulePass that will be added in a following CL.

Also includes a fix to HloFunctionImporter::ImportAsFunc which currently does not check whether or not the function_map_ is actually initialized and leads to a segfault trying to lookup whether the function is already imported. Fixes that by checking whether the function_map_ is initialized before checking it. Also makes it so that the caller should explicitly indicate whether or not the function being imported is the main function so that the caller is under control of whether the function is renamed to "main".

PiperOrigin-RevId: 462002829