DTensorCheckpointingV2 [3/n] tf.RestoreV2 Layout Propagation and SPMD expansion
tf.RestoreV2 previously had a dummy SPMD expander and a dummy layout propagator - This was because when using DCheckpointing, we made our custom DTensor op called DTensorRestoreV2. We made this custom op because we needed layouts and shapes of the tensors were restoring to, and DTensorRestoreV2 had shapes and layouts as attributes of the op.

Now that we have a new design with DTensorCheckpointingV2, where we are intergrating fully with tf.Checkpoint - that is, we do not want our own custom DCheckpointing class - we can properly do SPMD expansion of tf.RestoreV2 without the need for DTensorRestoreV2. The behavior of the SPMD expansion of tf.RestoreV2 is the same as the behavior of DTensorRestoreV2. The only difference is we get layouts and shapes of the tf.RestoreV2 properly - that is the layouts by the time we do SPMD expansion should be there as normally, and the shapes of the tf.RestoreV2 output should also be there.

PiperOrigin-RevId: 462248274