[xla:cpu] Try to pick the correct ABI for fp16 conversion runtime functions
The x86_64 psABI defines fp16 as passed in floating point registers instead of
general purpose registers. This means using uint16_t stopped working when LLVM
recently adopted that ABI rule.

The proper fix for this is using _Float16 instead but compiler support is still
spotty. GCC has supported it for a couple of versions, Clang as of now only in
trunk and other compilers don't have it at all.

To fix this we pick _Float16 when it's available on new GCC and Clang, and fall
back to float and careful casting if we don't have it.

PiperOrigin-RevId: 462640054