Extend the canonicalization rules for dynamic_iota.
Recognize a dynamic_iota whose argument is a constant,
and turn it into an iota.

I hoped we can get this done via shape inference, using the newly introduced
refineReturnTypes from MLIR upstream, but that doesn't work because we cannot
define inferReturnTypes for dynamic_iota, and it's still used in a number of
places.

PiperOrigin-RevId: 462987927