Fix crash from use-after-free in OpExpanderPass.
OpExpanderPass gets a list of HloInstruction*, then calls ExpandInstruction on each to rewrite it. ExpandInstruction may delete the instruction and its operands. It was possible this would delete a future HloInstruction* in the list, causing the pass to later access the already-freed HloInstruction*. This change causes the list to be created in post order, causing operands to always appear before their users.

PiperOrigin-RevId: 463165158