Prepare tf.module for making ResourceVariables as CompositeTensors.
We are going to let ResourceVariables subclass CompositeTensors in a future CL. With that change, we couldn't reply on the existing tf.nest to extract ResourceVariables embedded in CompositeTensors. Currently tf.nest can only treat CompositeTensors as leaves with expand_composites=False or fully expand CompositeTensors with expand_composites=True. If ResourceVariables are embedded in CompositeTensors, we will have to manually expand CompositeTensors layer by layer. One example of CompositeTensors contains ResourceVariables is ShardedVariable.

PiperOrigin-RevId: 463883129