Revert broken SplitV change.
PR #49215 introduced a SplitV optimization for a particular case of larger input tensors,
but as discovered in PR #56275, that implementation is incorrect.  Fixing the optimized
implementation leads to much slower compute times, making it no longer useful. Here
we revert the original PR and add test cases to catch the issue in future.

Fixes #56275.

PiperOrigin-RevId: 464004249