Enhance `tpu_sharding_identification_pass` when `use_sharding_op` attribute set as false in the frontend.
When sharding info is not set on the frontend, where attribute `use_sharding_op` is set as False, it implies that sharding attribute is added to the op itself.

If `TPUPartitionedOutput` does not contain the related `_XlaSharding` info, then `tpu_sharding_identification_pass` will not propagate `output_sharding_configuation` attribute to `tf_device.cluster_func` correctly.

PiperOrigin-RevId: 464690813