Fix TPUReplicatedVariable.handle access in eager mode.
When TPUReplicatedVariable.handle is accessed in a TPU replicated context, it is possible that the read is in an init_scope which causes tpu_partitioned_input not working correctly.

With init_scope the TPUReplicatedVariable.handle is accessed eagerly and thus we should return the handle of the first replica of the replicated variable.

PiperOrigin-RevId: 465148384