Add utility APIs to deal with input datasets efficiently in DTensor.
This change adds an input pipeline abstraction called `DTensorDataset` which extends `tf.data.Dataset`. It can be used as an iterator for an underlying dataset of DTensors which are correctly packed to the corresponding devices on each iteration, based on the provided layouts. This works in both single and multi-client DTensor use cases (with tf.data service).

PiperOrigin-RevId: 465256677