Add CheckPreemptionOp to detect if remote workers have received preemption signal.
When setting up ServerDef on remote workers, also register a signal listener to insert preemption information to the coordination service. Subsequent CheckPreemption ops running on any remote workers will be able to get the preemption notice and return an error with special payload, representing the first task that reports the preemption.

PiperOrigin-RevId: 465638259