Add GPU Delegate API to Lite Java docs.
This adds the [Java GpuDelegate](https://github.com/tensorflow/tensorflow/tree/master/tensorflow/lite/delegates/gpu/java/src/main/java/org/tensorflow/lite/gpu) API to the Lite reference docs, in a section marked "Delegates".

Note that this code is public, but has not yet been included in a public release, so this is safe to submit. Automation should take care of publishing it when 2.10 is released.

PiperOrigin-RevId: 465949074