lite: Make kStateUninvokable for Subgraph::ReleaseNonPersistentMemory()
It makes us to track memory status more correctly.

In the evaluation of control flow ops, kernel calls AllocateTensors()
for their subgraphs. If the subgraph state is kStateInvokable, the call just
return without going through actual allocation.
At the end of the evaluation, kernel calls ReleaseNonPersistentMemory()
to save memory.

This memory saving logic is already used for WHILE (b/197827490) and
we enable the same logic for IF kernel.

PiperOrigin-RevId: 466082405