Don't set MLIR_TABLEGEN_EXE
With llvm/llvm-project@112499f landed, `MLIR_TABLEGEN_EXE` is given as a
cache variable in the MLIR core project. Other external projects, such
as MLIR-HLO, should not set the variable as this breaks cross-compilation.

It might be necessary to wait for the integrate commit that lands commit
llvm/llvm-project@112499f, before this one can be merged.