[SavedModel Fingerprinting] Add version to the FingerprintDef.
For now, we only populate the `producer` field of the VersionDef, which describes the version of the code used to create the fingerprint. The `producer` field will be incremented every time the canonicalization code changes, the granularity of the change that warrants an increment is TBD.

In the future, we can add a `min_consumer` field if the need arises or have additional functions for version checking.

RFC: https://github.com/tensorflow/community/pull/415
PiperOrigin-RevId: 467075556