Add gradient support to IndexedSlices.
Background: The original goal was to add custom_gradient support to CompositeTensor. Then I realized IndexedSlices could be created by tf.gather and trying to extract gradient components from IndexedSlices could raise an exception.

It is probably not super exciting for users to directly take gradients of IndexedSlices. The main motivation is to unblock the custom gradient support for CompositeTensor.

PiperOrigin-RevId: 468514815