[XLA:GPU] Fix multi_output_fusion pass - do not fuse root instruction
The multi_output_fusion pass expects to match the following pattern:
↱ B ⬎
A    ROOT
↳ C ⬏
where the B/C are sibling fusions.

In a case where a fusion has no end users (if it has side effects, e.g. a custom-call), the root instruction of a computation may be passed down to "FuseSiblings" which then check fails.

PiperOrigin-RevId: 470711620