Disallow VariableSpecs in input_signature.
Because VariableSpec de-aliasing API is not yet determined. Since VariableSpec is not a public API, we should be safe to do so.

We can't dis-allow VariableSpecs passed to get_concrete_function at this moment because saved_model use them to create signatures during saving.

IteratorSpec is already a public API. So we can't disallow any dt_resource tensor spec inputs.

PiperOrigin-RevId: 470771027