Second attempt to fix a race condition in v1 control flow introduced by recent performance tuning change (https://github.com/tensorflow/tensorflow/commit/b4d798a4143590e7bf5c1f6eb2633cd46385cb15). After changing to use shared_lock instead of mutex lock, we need to ensure that inputs of the op should be set before adjusting its pending number. Since after we adjust the pending number, another thread may trigger the op execution.
This fix is based on the fact that Merge op doesn't support multiple valid edge.  However, it turns out that in production, we do have invalid model that have multiple active edge for merge op. In this second attempt, we only report a warning instead of error out when we encounter such invalid graphs to match previous behavior.

PiperOrigin-RevId: 470887318