[TF-TRT]: Revert default dynamic shape strategy to Range
Revert "Make default dynamic shape profile strategy to implicit batch compatible"

Currently, the ImplicitBatchModeCompatible strategy creates profiles with the
assumption that the first dim is always the batch dim. This does not necessarily
hold for TRTEngineops deeper in the graph, therefore we cannot simply override
dim 0 for all the min values. We have a case where this leads to inconsistent
profiles and engine creation failure in certain BERT networks.

The Range profile strategy does not have this issue, so revert to using
Range as the default.

This reverts commit 0737efbe374b2bc89134f153d54a45af6811592a.