Add Conv2D bfloat16 ops for CPU.
These are documented as available, but were never registered.  Intel, in an attempt
to allow oneDNN versions in graph mode, registered "fake" ops that lead to errors
in eager mode or when oneDNN is disabled (e.g. #56680).  Here we add correct ones.

Due do precision issues, the bfloat16 ops internally perform computation in float32 for
the forward pass, casting back-and-forth.

PiperOrigin-RevId: 471831125