Fix order-dependent test in debug_io_utils_test.
The subtests all used the same base folder for temporary files, and deleted the entire folder
at the end of the test.  If run concurrently, the result is that these tests would delete each
others' files.

Fix is to use separate unique subfolders, create the subfolder at the beginning of the test,
and only delete your own subfolder.

PiperOrigin-RevId: 472502667