Support TF Summary in DTensor [1/n]
https://www.tensorflow.org/api_docs/python/tf/summary

This is part of an effort to fix all things related to custom device pinning dtensor tests failures. The first one is TF.Summary.

Note that we are testing here with raw ops instead of the tf.summary api because we do not have full control of custom device pinning and do not yet support upcasting summary resource tensor to a dtensor.

PiperOrigin-RevId: 472557672