Fix result reporting description
The text printed along with results was changed in commit 48598dddf7a35471fb9dff431652df56306e1ad0.

The string that this document tells users to grep for will no longer work. This change fixes that problem, and replaces the grep string with one that will work with the new result message.