[mlir] Support prefixed getter/setters of attributes
This change tries to be as mechanical as possible leading to several awkward pieces of code:
1. op.getBody() -> op.SingleBlock::getBody()

Additionally, some methods were deleted as they are now created automatically.

All changes should be reversible to allow cleanup as desired afterwards.

This is an intermediate step as part of the pending deprecation of old accessor forms: https://discourse.llvm.org/t/psa-the-default-accessor-naming-has-been-flipped-to-prefixed/65111

PiperOrigin-RevId: 473961783