DTensor Remove ShardedPrefix Op. This is no longer needed as we can infer candidate prefixes.
This was an op used in the first version of dtensor checkpointing. In V2, we inferred the prefixes by generating all the possible candidates. Remove this op as we can use the V2 for all cases.

This is an effort to make some refactoring changes toward supporting saving in a tf function, which will help support multi-mesh saving and api consolidation.

PiperOrigin-RevId: 474123515