Migrate TensorFlow for --incompatible_use_platforms_repo_for_constraints
This is needed to make TensorFlow build with the upcoming Bazel 6.0 release.

We need newer version of rules_go and upb to adapt this incompatible changes, they are introduced in grpc_extra_deps(), but unfortunately, we cannot upgrade grpc in TF due to other problems, therefore we pin those two dependencies and add necessary patch to grpc.

PiperOrigin-RevId: 475567595