Avoid signed integer overflow when loading tensors with both 0 and large dims.
`TensorShapeBase` ensures `num_elements` doesn't overflow when adding a new
dimension. However, this check is insufficient to prevent other functions that
use a different multiplication order from hitting an overflow _if any of the
dimensions are 0_. For example, Eigen currently multiplies dimensions in
reverse order, so dimensions of (0, 4294967296,4294967296) will trigger an
overflow in Eigen code.

To prevent overflow for all multiplication orders, we can that `num_elements`
doesn't overflow if zero dimensions are skipped.

PiperOrigin-RevId: 477796726