Log sample numbers during calibration step.
This change adds logging sample numbers while running the calibration step for PTQ. The calibration step often runs slow, especially when the model is large or there are a lot of representative samples, and it may look stale when calibration step is running.
Logging the steps alleviates this problems and helps the user keep track of the calibration progress.

PiperOrigin-RevId: 477947702