Don't expose the "srcs" of :c_api.
Change the way that the tflite_custom_c_library target is implemented
so that we don't have to expose the source files for the :c_api target
and their corresponding dependencies.  In particular, rather than including
those targets in "srcs" of the generated rule, have the generated
rule only include the relevant "hdrs" and instead add "deps" on new
targets that contain the relevant "srcs".  We can't use the existing
":c_api" and ":c_api_experimental" targets for that, since those use
alwayslink=1, so instead add new "_without_alwayslink" targets that
don't include alwayslink=1.
PiperOrigin-RevId: 478078968