Add canellation APIs that supports cancelling in flight invocation.
During a long run tflite invoke, the application can call `Cancel` (probably from another thread) to cancel this invoke. This is similar to the existing `SetCancellationFunction` API but it differs in

1. Simple and straightforward API
1. `Cancel` only cancels current in flight invocation and will have no effect on subsequent `Invoke`.

By default this feature is turned off. Users need to enable it in interpreter option when building the interpreter. The API for that will be added in follow up changes.

PiperOrigin-RevId: 481842370