Attach "tf_saved_model.initializer_type" attribute to initializer functions during saved model import.
Also uses the `tf_saved_model.initializer_type` attribute to identify which init function should be added to the main function
from the `MergeInitializerFunctionOpsToMainPass`.

The attribute is used for distinguishing two different types of initializer functions: the init functions corresponding to the "init_op",
which is used for initializing resources such as hash tables, and "restore_op", which is used for restoring variables from checkpoint.

Accordingly, the attribute `tf_saved_model.initializer_type` is attached to various test files that introduce "init_op" or "restore_op".

The existence of `tf_saved_model.initializer_type` is not set as a hard requirement for initializer functions registered at
`tf_saved_model.session_initializer` op's `initializers` attribute in order to be compatible with old codes.

PiperOrigin-RevId: 482188631