Remove outdated docs re: Arduino Library Manager installation
For some years, Arduino maintained a TensorFlow Lite Micro mirror packaged as a standalone Arduino library named
"Arduino_TensorFlowLite". This allowed the library to be distributed via the Arduino Library Manager system.

The maintainers of the official tensorflow/tflite-micro-arduino-examples repository recently requested that library be
removed from the Arduino Library Manager. For this reason, it is no longer possible to install the library via Library
Manager. The library should now be installed following the official instructions here:

https://github.com/tensorflow/tflite-micro-arduino-examples#how-to-install