Make WhileRegion casts explicit. A WhileRegion's operands are implicitly cast to the iteration variable types. Iteration variables are declared as arguments of the body block and the condition block.
go/implicit-casts-and-variable-reformatting is the design doc.

With explicit casts, the WhileRegion operand type is the same as the iteration variable types. This is important for optimizations because the same types are needed for WhileRegionEliminatePassThrough to remove unnecessary iteration variables. Removed iteration variables are needed for tpu_variable_runtime_reformatting's passthrough detection.

This cl also converts explicit shape casts back to implicit in export from MLIR to TF Graph. Explicit shape casts are not represented in the TF graph. This is needed because casts in the TPU's partition must be implicit, since tf.cast is not registered for the TPU.

To handle more explicit casts, shape_inference for casts is more aggressive. It uses `GetCastCompatibleType` to handle subtypes, like variant's subtypes.

PiperOrigin-RevId: 482928204