Add unfreezing constants pass for TF quantization.
This is an initial implementation of the unfreezing constant pass where it transforms `ConstOp`s into `VarHandleOp -> ReadVariableOp` pairs. Also, the original `ConstOp`s are moved to the session initializer function where they are used to initialize the variables via `AssignVariableOp`s.

In this implementation there are no size threshold for `ConstOp`s for unfreezing and it also assumes that there are no `ConstOp`s existing in session initializer functions.
Also the `VarHandleOp`s shold be hoisted into the main graph in order to follow the TF2 semantics where shared variables aren't allowed.

PiperOrigin-RevId: 483287550