Support bfloat16 in GPU ops used in variable initializers.
In particular, a bfloat16 GPU kernel was added for ops Add, Sub, Mul, AddV2, Maximum, Minimum, which are used in various variable initializers.

The motivation is to fix an issue with bfloat16 when using XLA in TF1. Currently, XLA-GPU supports bfloat16, but TensorFlow without XLA does not support bfloat16.

The issue this change fixes was caused by the fact variables and their initializers are typically created outside XLA. When a variable initializer is a lambda function in TF1, the initializer is colocated with the variable. Before this change, since there were no bfloat16 GPU kernels for certain variable initializer ops, the initializer ops were run on the CPU, forcing the variable itself on the CPU due to the colocation. XLA does not support running a GPU computation with CPU variables, and even if it did, this would be slow. To fix this, this change adds bfloat16 GPU kernels for such initializer ops.

Ideally these ops would be generated by the MLIR kernel generator, but currently it dose not support bfloat16. Instead, normal non-MLIR kernels were added.

In the longer term, we plan on fully supporting bfloat16 in TensorFlow without XLA, by adding bfloat16 support for more ops.

PiperOrigin-RevId: 483420135