Make `quantization_options_proto`'s cc library header-only by default.
This fix makes the dependency relationship clearer and reduces the risk of ODR violation.

Currently, quantization_options_proto's implementation is transitively linked to `libtensorflow_cc.so`. This makes the dependency relationship difficult to reason about and introduces the risk of ODR violation for protobuf's static initialization of proto file registration.

PiperOrigin-RevId: 483843629