Emit both raw and prefix accessors for the TensorFlow dialect
This requires rename for operand named `t` in MLIR op definitions and move to manual definitions.

This in preparation for https://reviews.llvm.org/D136727 that removes support for raw accessors. With prefixed accessors, this will lead to conflict with 'T' as both accessors will become getT.

There are only five such ops so renaming inputs to something else is not that bad. Separately, we could figure out a better solution and auto-generate these ops.

PiperOrigin-RevId: 484488089