[xla:cpu-next] merge MLIR pipeline with hlo_xla_runtime_pipeline
This merges the two pipelines down to Linalg on buffers (the
MLIR pipeline then continues; the XLA Runtime pipeline stops
there since we have a separate pipeline to go down to LLVM).

I simply copied the MLIR pipeline in cpu_compiler.cc and pasted
it in hlo_xla_runtime_pipeline.cc.  Then just added options to
be able to do things slightly differently depending on whether
we are running XLA Runtime or not.

We could think about merging the remainder of the two pipelines
(i.e. from Linalg on buffers down to LLVM) later.  However those
parts of the pipelines have little in common so perhaps there
is little value in merging.

PiperOrigin-RevId: 484682154