Reset model object before using it to unpack a flatbuffer to avoid the memory leak
The existing code relies on UnPackTo to merge the existing state with the new incoming state. However, this behavior was changed in https://github.com/google/flatbuffers/pull/7527 and now leads to memory leak. Ignore the commit description there as it is outdated. During the review, they decided to not have the old behavior as an option.

PiperOrigin-RevId: 484798857