SpecializeType() handles attr default values
In SpecializeType() in full_type_util.cc, add default values (if defined) for any attributes not already specified. This supports (full) type constructors that use attributes.

Add a unit test in full_type_util.cc for the case of an op with a default for an attribute that is created without specifying that attribute.

PiperOrigin-RevId: 485174200