Use a single `tstring` to decompress non`memcpy`able tensors.
This change is primarily motivated by a forthcoming optimization that requires a preprocessing pass at the top of `UncompressElement` which incidentally allows us to count the amount of non`memcpy`able data and use a single `tstring`.  This approach has greater symmetry with `CompressElement`, is slightly more space efficient, and is negligibly more or less time efficient depending on the number of components.

PiperOrigin-RevId: 485175278