[XLA:TPU] Improve Pow in XLA, part 3/2: Improve the function which emits specialized code for known exponents. These calls to Pow can be implemented using only multiplications and calls to Cbrt and RsqrtNR, which are faster and more accurate than the general purpose Pow implementation. The code for special case handling is also much simpler than the general case.
In particular, this change adds or extends specializations for exponents that are known to be an integer plus multiples of 1/2, 1/3, and 1/4 at compile time.

Standard compliant special-case handling with proper test coverage is added for this branch of the code.

This fixes an old bug where pow(-inf, 0.5) was wrong. Exhaustive test updated accordingly.

Also, undo a regression introduced in cl/484042571: When base is known, we should compute log2(base) at compile time using std::log2. Thanks to Blake Hechtman for spotting this.

PiperOrigin-RevId: 485213398