Create standalone `pywrap_quantize_model`.
Instead of maintaining a separate `quantize_model_cc` target that is used just to be linked to `lib_pywrap_tensorflow_internal.so`, make `pywrap_quantize_model`
a standalone library for exporting symbols to python. To do so, `deps` of `pywrap_quantize_model` should all be header-only or only listed for this target.
Accordingly, `quantize_model_lib` and `calibrator_singleton` should be split into header-only and implementation targets.

* `quantize_model_lib` is renamed to `quantize_model_cc` and its implementation target is directly included only in `lib_pywrap_tensorflow_internal.so` to avoid ODR violation.
* `calibration_singleton`'s implementation target is linked with `custom_aggregator_op`.

PiperOrigin-RevId: 486001941