Refactor out lowering to HLO from XlaCompilationCache into it's own class
Currently the TfGraphToHloCompiler class just calls functions from the XlaCompiler class to do the lowering. Unit tests for this class would be added in the future in case more logic is added there or the lowering logic is moved there instead.

The TfToHloCompiler class is being used as an interface which the TfGraphToHloCompiler class implements.

PiperOrigin-RevId: 487332718