Fix subtle portability bug with exports in external_delegate_adaptor.cc.
Previously, the declarations of the API functions here were using
TFL_CAPI_EXPORT, which is a macro defined in c_api_types.h that on
most systems expands to '__attribute__((visibility("default")))', but
which on Windows expands to either '__declspec(dllimport)' or
'__declspec(dllexport)' depending on whether a particular flag macro
'TFL_COMPILE_LIBRARY' is defined.  The idea is that when building
the TF Lite C API DLL, you define that symbol, and then the declarations
annotated with TFL_CAPI_EXPORT will be treated as dll exports, causing
that DLL to export those symbols, but when merely using the DLL, you don't
define that symbol, and then those declarations are treated as dll imports,
allowing code outside the DLL to use symbols defined inside the DLL.

However, the symbols declared in this file are NOT intended to be part of the
TF Lite C API DLL, but are instead part of a DIFFERENT shared library,
containing an external delegate.  When building that library, we want
the TF Lite C API symbols to be imports, not exports, but we want the
two symbols defined by that library to be exports.  So we can't use
the same macro for both TF Lite C API symbols and the symbols defined
by that library.  Instead we define a new macro TFL_EXTERNAL_DELEGATE_EXPORT
whose expansion on Windows is controlled by a new flag macro
'TFL_EXTERNAL_DELEGATE_COMPILE_LIBRARY'.

This change is a necessary BUT NOT SUFFICIENT step towards making this
code work on Windows.  (There are definitely other issues, e.g. the build
rule unconditionally generates a .so file rather than a .dll, etc.)

This change should have no effect on non-Windows systems, since the two
macros both expand to the same thing on non-Windows systems.

PiperOrigin-RevId: 487400613