Automated g4 rollback of changelist 486223331.
*** Reason for rollback ***

TFLite team has decided to continue to use the macro

*** Original change description ***

TFLite Micro introduced the macro TF_LITE_REPORT_ERROR as an attempt to wrap the error logging functionality to check for TF_LITE_STRIP_ERROR_STRINGS (an ifdef var to ignore calls to message logging API).

This macro was meant to be used in the TFLite code that is shared with TFLite Micro. But its is being used in several places unintentionally.

This CL cleans up the usage of TF_LITE_REPORT_ERROR from TFLite code, to limit the usage of the macro to the code that TFLite shares with...

***

PiperOrigin-RevId: 488409656