[SavedModel Fingerprinting] Add instrumentation for the fingerprint in the following APIs: (1) tf.saved_model.save (2) tf.saved_model.load (3) tensorflow::LoadSavedModel (4) tensorflow::SavedModelV2Bundle::Load
This commit introduces two new string gauges:
(1) "/tensorflow/core/saved_model/write/fingerprint": This is the fingerprint of the SM that is written/saved/exported.
(2) "/tensorflow/core/saved_model/read/fingerprint": This is the fingerprint of the SM that is read/loaded/imported.

Each gauge (both with zero fields) contains the saved_model_checksum which corresponds to the field of the same name in the fingerprint.pb proto.

https://github.com/tensorflow/community/pull/415

PiperOrigin-RevId: 488782307