Implement saturate_cast for complex types.
The original implementation fails since complex types don't have a `min` or `max` property.
However, we should be able to still perform a saturate-cast by considering the real
and imaginary components separately.

We introduce some logic to correctly handle saturation for various input/output types,
and add a CPU kernel for `ClipByValue` (aka Clamp) for complex types that considers
the clamping region to be the 2D box in the complex plane bounded by the min/max values.

Fixes #58053.

PiperOrigin-RevId: 488998520