Add stablehlo-legalize-to-hlo to Phase 2 of the bridge
Currently, we still haven't migrated to StableHLO but this is required for jax2tf with jax migrating to StableHLO recently. Later on, we will anyway need this when we migrate to StableHLO.

This is a peculiar use-case and added a TODO to refactor to not invoke tf2xla bridge pipeline for modules that are already in MHLO/StableHLO. Adding this pass now to continue supporting the existing use-case.

PiperOrigin-RevId: 489092669