Constant fold MHLO broadcast_in_dim op before mul op.
Implements a pattern template so that support for other binary elementwise ops could be added easily. The pattern for constant folding:
(MHLO_MulOp
  (MHLO_BroadcastInDimOp
    (MHLO_Constant $lhs)),
  (MHLO_Constant $rhs))

Note that the binary op folder in hlo_ops.cc currently has an upper limit on how many elements can be folded by an op folder. So we manually constant fold MulOp in the stablehlo path.

PiperOrigin-RevId: 489596880