Avoid excessive creation of quantize/dequantize routines around same-scale ops
The current quantization strategy is greedy. That means it will quantize every quantizable ops/blocks when possible. This is not always a good strategy, especially around the same-scale ops.

Ex: A Transformer model X has 262 Reshape ops that is not connected to any convolution blocks. With the current strategy, we are adding:
`quantize -> Reshape -> dequantize`
for all those Reshape ops, which significantly degrades the performance since using the int8 Reshape bring very small performance gain compared to the cost of additional quantize/dequantize routines added.

This cl is a quick fix that avoids quantizing same-scale ops that are not connected to a Convolution/Matmul block. When our strategy changes, we might relax this policy in the future.

PiperOrigin-RevId: 490406751