Implement a pass that duplicates shape-determining constants.
This change implements `DuplicateShapeDeterminingConstantsPass`, which clones
constants that affect or determine the shape of a tensor after being used in a computation for some op.
Some specific operands of TF ops (like the `dim` argument for `TF::ExpandDimsOp`) determine the shape of the resulting tensor.

If these operands are constants, they are duplicated and replace the shape-determining operands.
Each duplicated constant will only be used as the shape-determining operand; it will not replace other usages of the original constant.
If the operands are not constants (i.e.  results of some other computation), then the pass recursively traverses the call tree upwards and duplicates all constants found in the subtree in a similar manner,
so that the whole subtree is "known at compile time"

PiperOrigin-RevId: 490664127