Host send/recv with maximal sharding shouldn't be treated as MPMD
Sharding propagation has a heuristic from MPMD days that propagates MPMD single-device instructions to their containing while loop. But it accidentally triggers on host send/recv instructions with maximal sharding, which are common in an SPMD context for outside compilation/debugging workflows. We add a "not a host transfer" condition to this heuristic.

PiperOrigin-RevId: 491722069