Remove functional to region control flow pass in TFXLA pass pipeline
This pass was introduced for transition and isn't necessary now that all tests are passing.

This required adding sink constants to control flow pass. The pass is in MHLO but is generic and sinks constants for any op with regions. Here, constants needs to sinked to regions so that tensorflow ops requiring constant operands can be legalized in later passes.

PiperOrigin-RevId: 491746367