Add DeviceExecutablePersistor class to persist/load executables on/from disk.
Currently it contains the persistence logic present in the XlaCompilationCache class (which exclusively deals with xla::LocalExecutable and xla::LocalClient for now).

The DeviceExecutablePersistor class is templated on the executable type and the client type. We may need to extend DeviceExecutablePersistor into XLA and PjRt specific subclasses in the future if the implementation diverges.

PiperOrigin-RevId: 492364175