Ensure that changes to configuration.proto don't break binary compatibility.
Add a test to verify that changes to the Protobuf schema in configuration.proto
don't break binary backwards compatibility for the autogenerated FlatBuffer
schema.

This required checking in the previous copy of the Protobuf schema into
the repository (since tests only have access to the current workspace,
not to previous workspaces).

To help ensure that the previous copy remains up-to-date, I have added
documentation warnings and LINT.IfChange ... LINT.ThenChange directives to
help ensure that the previous version in testdata/configuration.proto_prev
gets updated whenever configuration.proto is changed, and also a test
to verify that configuration.proto and testdata/configuration.proto_prev
are different.

PiperOrigin-RevId: 492989280