Save variables to checkpoint in TF quantizer.
Using BundleWriter, TF quantizer saves the tensors extracted from the unfrozen variables to checkpoint files.
The target variables are only searched within the initializer function with `initializer_type = "restore_op"`.
It identifies all `tf.AssignVariableOp` in the initializer function (which should be populated after `UnfreezingConstantsPass`) and finds the `tf.ConstOp`s that are used to initialize the `VarHandleOp`s.
These constant values are stored to the checkpoint file.

PiperOrigin-RevId: 493467590