Support ReduceWindowOp creation with a Region
Primarily, this is to make the type inference useful, as otherwise, it is
impossible to create an mhlo.reduce_window with type inference due to its
verification. The verification issue should be fixed, but this is also useful
to ensure that the op is valid right after construction similar to convenience
builders that scf dialect ops have.

PiperOrigin-RevId: 493548719