Move the model loader to tools. Change the benchmark tool to use the model loader.
The Delegate Performance Benchmark app is designed to qualify the stable delegate share libraries at model level. The app will embed a set of models in the app's assets directory for verification. Currently the app cannot use the benchmark tool to load the models in assets because the benchmark tool relies on absolute paths to the models and the app cannot provide absolute paths to the models in assets.

The model loader can load models with a model file descriptor. It allows the Delegate Performance Benchmark app to pass the models by file descriptors for triggering the benchmark tool.

PiperOrigin-RevId: 493671255