Lower log1p to libm if approximations are disabled.
log1p(x) has accuracy issues around 0 if implemented as log(1+x),
because 1+x is 1. This change optionally lowers it to a libm call if
approximations are disabled.

This fixes jax's lax_test_cpu and the remaining failure in lax_scipy_test_cpu
(but there are still tests that time out).

PiperOrigin-RevId: 493847049