Add another test for ABI backwards compatibility.
This one is designed to catch ABI breakage from possible future changes
to the FlatBuffer compiler's proto to FlatBuffer schema conversion
(e.g. see https://github.com/google/flatbuffers/issues/7645).
To do that, this CL checks in a snapshot of the .fbs file generated by
the FlatBuffer compiler as of this CL, and then at build/test time we
generate a new .fbs file, which will potentially use a different version
of the FlatBuffer compiler (if/when the FlatBuffer compiler is updated).
Then we use 'flatc --conform' to verify compatibility of those two.

PiperOrigin-RevId: 493867469