Add the compiler's resource_dir/share folder to the default inc path.
As discussed here (https://github.com/google/oss-fuzz/issues/3093), clang was updated to include the sanitizer ignorelist within the `.D`. This causes bazel to look for those files within the dependency list, so this directory has to be added to the default include path for bazel to not complain while running sanitizers with clang.

PiperOrigin-RevId: 493939307