Add experimental `disable_optimize_for_static_graph` option to `tf.ConfigProto`.
Currently, the Session implementation has functionality that allows it to
discard the original `GraphDef` object after creation, which results in lower
memory consumption.

This is part of an effort to enable such functionality by default.
Specifically, this new flag will allow users to opt-out from the optimization
if needed.

An opt-out will be required for users that intend to use the interface
`Session::Extend()` (or in case that `GraphOptions::place_pruned_graph()` is
true). Both of those cases are expected to be relatively rare.

PiperOrigin-RevId: 493979771