Add `TF::TensorArraySizeV3Op` to the list of ops whose outputs are considered compile time constants.
`TensorArraySizeV3` returns the size of a `TensorArrayV3`, which can then be passed onto ops that require compile time constant operands like `Reshape`.
Since it is just getting the size of `TensorArrayV3`, we can consider the output of `TensorArraySizeV3` a compile time constant (in terms of "the shape is known" from the XLA compiler's perspective).

PiperOrigin-RevId: 494056634