Rename the mlir::TFL::mhlo namespace to mlir::odml
This is part 1 of 5 of the CL chain that migrates odml_to_stablehlo from producing MHLO to producing StableHLO.

To prepare for the migration, I propose that we rename the old-style mlir::TFL::mhlo namespace to the new-style mlir::odml namespace. This new-style namespace is already used by several recent passes.

PiperOrigin-RevId: 495143635