Create a 64KiB lower threshold for constants during `UnfreezeConstantsPass`.
Current implementation of `UnfreezeConstantsPass` unfreezes all constants it encounters, which might not be desirable when even constants as small as few bytes are conveted to variables.
This change introduces a lower threshold of 64 KiB (arbitrarily determined) when unfreezing so that only "reasonably large" constants are converted to variables.

This option is not exposed to the users of `UnfreezeConstantsPass` unless we find a clear use case of adjusting this threshold.

PiperOrigin-RevId: 495198423