Rename components in lite/stablehlo from MHLO to StableHLO
This is part 3 of 5 of the CL chain that migrates odml_to_stablehlo from producing MHLO to producing StableHLO.

This CL renames almost everything that says "mhlo" to say "stablehlo", without changing any other behavior. I hope that having this change to be separate from the main CL that does the actual migration will simplify the review.

The only two exceptions were the fuse_mhlo_convolution.mlir and unfuse_mhlo_batch_norm.mlir which will be operating on MHLO even after the migration. As discussed, we aren't going to be doing optimization passes on StableHLO - that will happen in MHLO, even in the long term.

PiperOrigin-RevId: 495226095