[XLA] Make MaybeImproveInstructionSharding() better.
MaybeImproveInstructionSharding() tries to merge sharding only if
it doesn't cause additional communication.

Add a function in hlo_sharding_utils that checks if a sharding is a subsharding
of another. If sharding A is a subsharding of sharding B if going from sharding B
to sharding A doesn't cause any communication because each device already has the
data it needs based on how sharding A is compared to sharding B.

PiperOrigin-RevId: 495921146