[XLA] Add optional metadata for `GenericTransferManager::TransferLiteralFromDevice`, allowing for asynchronous transfers on CUDA.
CUDA callbacks are tricky as we cannot call any CUDA driver functions from within the callback. As a result of this, the version of `TransferLiteralFromDevice` with a callback is actually synchronous in the generic transfer manager. While this can be fixed by executing the driver functions in a separate thread, this still adds overhead. Here, we add some optional metadata allowing the user to declare their callback if safe to call from within a host callback.

PiperOrigin-RevId: 496395860