#tf-data Fix a test failure:
```
ValueError: Cannot capture a stateful node (name:DummySeedGenerator, type:DummySeedGenerator) by value.: A likely cause of this error is that the dataset for which you are calling `make_one_shot_iterator()` captures a stateful object, such as a `tf.Variable` or `tf.lookup.StaticHashTable`, which is not supported. Use `make_initializable_iterator()` instead.
```
PiperOrigin-RevId: 496489324