Add ClipByValue to TFLite flex delegate.
The op is now used by `tf.saturate_cast`, and will soon by used by `tf.clip_by_value`.
It is needed for handling complex inputs, and is more efficient than using a combination
of a `min` and `max`. We will also rely on `ClipByValue` more and more with new quantization
efforts, which need to pattern-match to created fused quantization ops.

Some model usages are already breaking due to lack of support in tflite. The forward
compatibility windows has already ended, so my understanding is that models may already
have this op in the wild.

PiperOrigin-RevId: 496696625