Update and pin all GitHub reusable workflows.
As part of Scorecards score, all dependencies (including GH Actions!) must be pinned. This change updates everything to latest version and adds the corresponding pin. This syntax is preferred by Dependabot, Renovate and other similar bots that can automatically update dependencies so in the future we can get PRs that will update these when there is a new release. This way, we don't need to keep track of them.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>
PiperOrigin-RevId: 497023994