Mark `XlaSetBoundOp`'s value as a bounded compile time value.
`tf.XlaSetBoundOp` is used to set the upper limit of an integer scalar value, which may hint the XLA compiler about the size of a tensor or a control flow (e.g. Range).
The output of this op is considered bounded, so we don't need to further traverse the tree upwards to duplicate constants.

The op itself also has a requirement that it's second input $bound should be compile-time constant.

PiperOrigin-RevId: 497068355