DTensorXlaSpmd [1/3] Fix Bug in Layout-to-Xla sharding.
Previously, the algorithm used a transposed mesh shape to calculate the new xla sharding assignment devices, but since the minor-to-major ordering was already changed, we should not be using the transposed mesh shape, but the normal mesh shape.

This has been working for symmetric meshes because the mesh dimensions are all equal, but it wouldnt work for asymmetric meshes.

PiperOrigin-RevId: 497250926