Add calling StablehloLegalizeToHloPass to ConvertMlirHloToHlo
To support the ongoing migration of XLA's compiler interface from MHLO to StableHLO, this CL inserts this fallback to provide support for backends which are converting incoming ModuleOps directly to HLO.

xla::MlirToXlaComputation is a better API for this purpose because it supports not just MHLO, but also CHLO and StableHLO, but we will temporarily support StableHLO to MHLO lowering here as well to ensure a smooth migration.

Once we have reasonable confidence that everyone has migrated from calling ConvertMlirHloToHlo directly, this fallback will be removed.

PiperOrigin-RevId: 498023214