Restore variables from TF Quntizer's python layer.
Once `MergeInitializerFunctionOpsToMainPass` exposes the control return node, the node can be used to identify the node in the graph for restoring variables.
Other information is required, such as the path to checkpoint directory, input file prefix tensor and shared names of the variables to be restored.
`Variable`s with empty value and unknown shape is created for each shared_name, just to register the variables to resource manager and collection defs.

`InlinerPass` should be disabled for post-calibration static-range PTQ to keep the lifted quantizable functions from the output of the pre-calibration step.

PiperOrigin-RevId: 498931590