Fix memory access error in SetSize.
If values are out of range, it can trigger an ASAN failure and memory corruption.
Added a quick guard.

Note that in the added test, the error is actually overwritten later on in the graph
with a different index-out-of-range error, so we don't actually observe the added
`InvalidArgument` message.  It still does fix the asan issue though.

PiperOrigin-RevId: 499228292