Remove MLIR legalization for Qr op in favor of fallback implementation
As Peter pointed out, this does not incorporate a number of improvements to the tf2xla kernel version in functionality and performance. We could simply make use of the fallback legalization and just have a single implementation for this complex op.

Longer term, We have a plan to move towards a world where there is exactly one implementation for each op either in MLIR or in tf2xla. Given that this is a complex op and is already outdated, better to remove MLIR implementation immediately.

PiperOrigin-RevId: 499236062