[tf-mlir] Canonicalize Einsums that do BatchMatmul to skip output transposes when lowered to a BatchMatmul.
Einsums are lowered to a sequence of tranposes -> BatchMatmul-> transpose. However, in many cases the output transpose is unnecessary. We can reduce the number of permuted dimensions by canonicalizing an ordering of the RHS and LHS of the einsum based on the earliest non-batch, non-contracted dimension. If the dimension variable in the einsum appears earlier in the OUTPUT, than swap the operands. Otherwise do not.

PiperOrigin-RevId: 499275887