Add env var to disable scatter deterministic implementation.
Setting the environment variable TF_DISABLE_SCATTER_OP_DETERMINISM to 1 causes GPU scatter ops such as ScatterNdAdd to use a potentially-nondeterministic implementation even when determinism is enabled. The reason is the deterministic implementation copies the inputs to the CPU and runs the op on the CPU, and therefore is very slow.

Scatter ops are nondeterministic only if there are duplicate indices, so if the user ensures there are no duplicate indices, setting the environment variable will make scatter ops runs significantly faster without being nondeterministic.

PiperOrigin-RevId: 499362625