Automated g4 rollback of changelist 493461477.
Rollforward of cl/493447633. The original CL broke some TAP tests. (Not sure
why `g4 presubmit` didn't catch those.)

Duplicate small constants placed on device.

This allows a small constant to be duplicated to different clusters.
Otherwise, the constant is copied back to host before it's passed to the second
cluster.

PiperOrigin-RevId: 499713008