Use `std::string` instead of `absl::string_view` for fields in `ExportOptions`.
`absl::string_view` member variables is not ideal for the case where the object lifetime of the references do not exceed `RunExportPasses`.
Example of this is happening for the `debug_name` field, which is passed a temporary object (the output of `absl::StrCat`),
essentially providing garbage values to the mlir dump file of the export step.

PiperOrigin-RevId: 500839575