[XLA] Delete the overload of CustomCall() that takes a single string_view.
Previously we we had two overloads:

  CustomCall(string_view target);
  CustomCall(absl::span<const string_view> targets);

This creates an issue wrt C++20.  In C++17, the code

  {"a", "b"}

is a span of two string_view's.  But in C++20, string_view adds a constructor
that takes two char*'s (as begin/end iterators), and this is now parsed as a
call to the string_view constructor.  Obviously this is not what we intend.

To fix this, I just got rid of the overload that takes a single string_view.
Now if you want to pass just one target to CustomCall, you have to write
CustomCall({"foo"}) whereas before you were able to write simply
CustomCall("foo").  Not great, but it works.

PiperOrigin-RevId: 501336895