Placement mechanisms for function arguments and return values using full type
Placement mechanisms for function arguments and return values that set `MemoryTypeVector` or `AllocatorAttributes` based dtype. "Weak" versions of these do not change existing behavior but logging of warnings if full type information is not correct can be enabled. Non-"Weak" versions of these generate an error if full type information is not consistent with existing behavior based on dtype DT_INT32 (for no behavior change if full type information is  correct).

These placement mechanisms are not used yet.

PiperOrigin-RevId: 501418067