Use output_path to export stablehlo always.
This CL use `output_path` always for exporting stablehlo mlir. It is required for some models with TF ops (e.g. skip-resize option enables tf.ResizeBilinear) and darwinn compiler can compile it. If not, it is generated a specific shape of `mhlo.gather` op not supported in DarwiNN compiler yet. So, even if this model compilation is failure in the standard point of view, we still want the darwinn compilation with the `output_path` name, not just switching to `debug_stablehlo.mlir`.

PiperOrigin-RevId: 501685513