Frontend: Add on-chip memory BW reports to the op profile xprof view.
This is the follow-on CL from the previous server-side changes.

Previously we only reported overall and hbm memory utilization.

However, each memory has its own bandwidth and utilization, so it's better to report them separately instead of in aggregate.

Eventually, we should probably remove the aggregate utilization, which is the only utilization we originally had.

PiperOrigin-RevId: 501913756