Reject bounded dynamic inputs for Split and SplitV lowering
If the Split op axis dimension size is odd, then it should return an error but we can't know that at the compile time if the input is bounded.

SplitV op takes split sizes so we could support that assuming that the input size matches the requested sizes. However, if one of the split size is `-1`, then we need to set dynamic dimension appropriate that we currently don't do.

PiperOrigin-RevId: 503234610