Add DTensor SPMD expander for IteratorGetNextOp.
This change adds an SPMD expander implementation for IteratorGetNextOp which will enable dataset iterators to be used inside tf.function. The element layouts are attached to the iterator resource DTensor as an attribute, and then used inside the SPMD Expansion algorithm. Iterator resources are created for each host mesh device and packed into a DTensor.

PiperOrigin-RevId: 503313097