Preserve function aliases through the TF quantizer.
Function aliases associate user-defined names to TF functions. This information is in the `MetaGraphDef` and
is used to identify specific functions of interest (the actual function names change every time the model is saved as SavedModel).

This change preserves the function aliases through the TF functions, so that the same functions (quantized, after the quantization passes) are still associated with the same aliases.
This is done by preventing the aliased functions from being inlined by `MarkFunctionsNoinlinePass`. The pass receives a list of names that should not be inlined, so the names should be passed along from the quantizer's python API.

PiperOrigin-RevId: 504396824