[XLA:GPU] Use a global thread pool in ShapeUtil::ForEachIndexInternalParallel (revised)
We got reports that always recreating the threads were slow.

We used the same global thread pool pattern as cuda_driver.cc/GetDriverExecutor.
https://github.com/tensorflow/tensorflow/blob/7ab78a5d7721ec10f3a742bb8f997ee34c2afe15/tensorflow/compiler/xla/stream_executor/cuda/cuda_driver.cc#L102

Added ShapeUtil::GetForEachIndexParallelThreadCount() to safely check thread pool size.

It is not safe to use tsl::port::MaxParallelism(), because it may change during the runtime of the program, so it may not be the same as the ForEachIndexParallel* thread pool size.

PiperOrigin-RevId: 504532192